+++ Makefile.js
@@ -1,3 +1,3 @@
 /* eslint-disable */
 // prettier-ignore
-"use strict";var e=require("os"),t=require("fs"),r=require("path"),n=require("events"),i=require("assert"),o=require("util"),s=require("child_process");function c(e){return e&&"object"==typeof e&&"default"in e?e:{default:e}}var a=c(e),u=c(t),l=c(r),f=c(n),p=c(i),h=c(o),d=c(s),v="undefined"!=typeof globalThis?globalThis:"undefined"!=typeof window?window:"undefined"!=typeof global?global:"undefined"!=typeof self?self:{};function g(e){return e&&e.__esModule&&Object.prototype.hasOwnProperty.call(e,"default")?e.default:e}var y=function(e){return e&&e.Math==Math&&e},m=y("object"==typeof globalThis&&globalThis)||y("object"==typeof window&&window)||y("object"==typeof self&&self)||y("object"==typeof v&&v)||function(){return this}()||Function("return this")(),E={},b=function(e){try{return!!e()}catch(e){return!0}},w=!b((function(){return 7!=Object.defineProperty({},1,{get:function(){return 7}})[1]})),O=!b((function(){var e=function(){}.bind();return"function"!=typeof e||e.hasOwnProperty("prototype")})),S=O,R=Function.prototype.call,I=S?R.bind(R):function(){return R.apply(R,arguments)},A={},j={}.propertyIsEnumerable,x=Object.getOwnPropertyDescriptor,k=x&&!j.call({1:2},1);A.f=k?function(e){var t=x(this,e);return!!t&&t.enumerable}:j;var L,_,T=function(e,t){return{enumerable:!(1&e),configurable:!(2&e),writable:!(4&e),value:t}},N=O,P=Function.prototype,F=P.call,D=N&&P.bind.bind(F,F),C=N?D:function(e){return function(){return F.apply(e,arguments)}},M=C,G=M({}.toString),$=M("".slice),U=function(e){return $(G(e),8,-1)},B=U,X=C,q=function(e){if("Function"===B(e))return X(e)},W=b,V=U,H=Object,z=q("".split),Y=W((function(){return!H("z").propertyIsEnumerable(0)}))?function(e){return"String"==V(e)?z(e,""):H(e)}:H,K=function(e){return null==e},J=K,Q=TypeError,Z=function(e){if(J(e))throw Q("Can't call method on "+e);return e},ee=Y,te=Z,re=function(e){return ee(te(e))},ne="object"==typeof document&&document.all,ie={all:ne,IS_HTMLDDA:void 0===ne&&void 0!==ne},oe=ie.all,se=ie.IS_HTMLDDA?function(e){return"function"==typeof e||e===oe}:function(e){return"function"==typeof e},ce=se,ae=ie.all,ue=ie.IS_HTMLDDA?function(e){return"object"==typeof e?null!==e:ce(e)||e===ae}:function(e){return"object"==typeof e?null!==e:ce(e)},le=m,fe=se,pe=function(e){return fe(e)?e:void 0},he=function(e,t){return arguments.length<2?pe(le[e]):le[e]&&le[e][t]},de=q({}.isPrototypeOf),ve=he("navigator","userAgent")||"",ge=m,ye=ve,me=ge.process,Ee=ge.Deno,be=me&&me.versions||Ee&&Ee.version,we=be&&be.v8;we&&(_=(L=we.split("."))[0]>0&&L[0]<4?1:+(L[0]+L[1])),!_&&ye&&(!(L=ye.match(/Edge\/(\d+)/))||L[1]>=74)&&(L=ye.match(/Chrome\/(\d+)/))&&(_=+L[1]);var Oe=_,Se=Oe,Re=b,Ie=!!Object.getOwnPropertySymbols&&!Re((function(){var e=Symbol();return!String(e)||!(Object(e)instanceof Symbol)||!Symbol.sham&&Se&&Se<41})),Ae=Ie&&!Symbol.sham&&"symbol"==typeof Symbol.iterator,je=he,xe=se,ke=de,Le=Object,_e=Ae?function(e){return"symbol"==typeof e}:function(e){var t=je("Symbol");return xe(t)&&ke(t.prototype,Le(e))},Te=String,Ne=function(e){try{return Te(e)}catch(e){return"Object"}},Pe=se,Fe=Ne,De=TypeError,Ce=function(e){if(Pe(e))return e;throw De(Fe(e)+" is not a function")},Me=Ce,Ge=K,$e=function(e,t){var r=e[t];return Ge(r)?void 0:Me(r)},Ue=I,Be=se,Xe=ue,qe=TypeError,We={exports:{}},Ve=m,He=Object.defineProperty,ze=function(e,t){try{He(Ve,e,{value:t,configurable:!0,writable:!0})}catch(r){Ve[e]=t}return t},Ye=ze,Ke=m["__core-js_shared__"]||Ye("__core-js_shared__",{}),Je=Ke;(We.exports=function(e,t){return Je[e]||(Je[e]=void 0!==t?t:{})})("versions",[]).push({version:"3.26.0",mode:"global",copyright:"© 2014-2022 Denis Pushkarev (zloirock.ru)",license:"https://github.com/zloirock/core-js/blob/v3.26.0/LICENSE",source:"https://github.com/zloirock/core-js"});var Qe=We.exports,Ze=Z,et=Object,tt=function(e){return et(Ze(e))},rt=tt,nt=q({}.hasOwnProperty),it=Object.hasOwn||function(e,t){return nt(rt(e),t)},ot=q,st=0,ct=Math.random(),at=ot(1..toString),ut=function(e){return"Symbol("+(void 0===e?"":e)+")_"+at(++st+ct,36)},lt=m,ft=it,pt=ut,ht=Ie,dt=Ae,vt=Qe("wks"),gt=lt.Symbol,yt=gt&&gt.for,mt=dt?gt:gt&&gt.withoutSetter||pt,Et=function(e){if(!ft(vt,e)||!ht&&"string"!=typeof vt[e]){var t="Symbol."+e;ht&&ft(gt,e)?vt[e]=gt[e]:vt[e]=dt&&yt?yt(t):mt(t)}return vt[e]},bt=I,wt=ue,Ot=_e,St=$e,Rt=function(e,t){var r,n;if("string"===t&&Be(r=e.toString)&&!Xe(n=Ue(r,e)))return n;if(Be(r=e.valueOf)&&!Xe(n=Ue(r,e)))return n;if("string"!==t&&Be(r=e.toString)&&!Xe(n=Ue(r,e)))return n;throw qe("Can't convert object to primitive value")},It=TypeError,At=Et("toPrimitive"),jt=function(e,t){if(!wt(e)||Ot(e))return e;var r,n=St(e,At);if(n){if(void 0===t&&(t="default"),r=bt(n,e,t),!wt(r)||Ot(r))return r;throw It("Can't convert object to primitive value")}return void 0===t&&(t="number"),Rt(e,t)},xt=_e,kt=function(e){var t=jt(e,"string");return xt(t)?t:t+""},Lt=ue,_t=m.document,Tt=Lt(_t)&&Lt(_t.createElement),Nt=function(e){return Tt?_t.createElement(e):{}},Pt=Nt,Ft=!w&&!b((function(){return 7!=Object.defineProperty(Pt("div"),"a",{get:function(){return 7}}).a})),Dt=w,Ct=I,Mt=A,Gt=T,$t=re,Ut=kt,Bt=it,Xt=Ft,qt=Object.getOwnPropertyDescriptor;E.f=Dt?qt:function(e,t){if(e=$t(e),t=Ut(t),Xt)try{return qt(e,t)}catch(e){}if(Bt(e,t))return Gt(!Ct(Mt.f,e,t),e[t])};var Wt={},Vt=w&&b((function(){return 42!=Object.defineProperty((function(){}),"prototype",{value:42,writable:!1}).prototype})),Ht=ue,zt=String,Yt=TypeError,Kt=function(e){if(Ht(e))return e;throw Yt(zt(e)+" is not an object")},Jt=w,Qt=Ft,Zt=Vt,er=Kt,tr=kt,rr=TypeError,nr=Object.defineProperty,ir=Object.getOwnPropertyDescriptor;Wt.f=Jt?Zt?function(e,t,r){if(er(e),t=tr(t),er(r),"function"==typeof e&&"prototype"===t&&"value"in r&&"writable"in r&&!r.writable){var n=ir(e,t);n&&n.writable&&(e[t]=r.value,r={configurable:"configurable"in r?r.configurable:n.configurable,enumerable:"enumerable"in r?r.enumerable:n.enumerable,writable:!1})}return nr(e,t,r)}:nr:function(e,t,r){if(er(e),t=tr(t),er(r),Qt)try{return nr(e,t,r)}catch(e){}if("get"in r||"set"in r)throw rr("Accessors not supported");return"value"in r&&(e[t]=r.value),e};var or=Wt,sr=T,cr=w?function(e,t,r){return or.f(e,t,sr(1,r))}:function(e,t,r){return e[t]=r,e},ar={exports:{}},ur=w,lr=it,fr=Function.prototype,pr=ur&&Object.getOwnPropertyDescriptor,hr=lr(fr,"name"),dr={EXISTS:hr,PROPER:hr&&"something"===function(){}.name,CONFIGURABLE:hr&&(!ur||ur&&pr(fr,"name").configurable)},vr=se,gr=Ke,yr=q(Function.toString);vr(gr.inspectSource)||(gr.inspectSource=function(e){return yr(e)});var mr,Er,br,wr=gr.inspectSource,Or=se,Sr=m.WeakMap,Rr=Or(Sr)&&/native code/.test(String(Sr)),Ir=ut,Ar=Qe("keys"),jr=function(e){return Ar[e]||(Ar[e]=Ir(e))},xr={},kr=Rr,Lr=m,_r=ue,Tr=cr,Nr=it,Pr=Ke,Fr=jr,Dr=xr,Cr=Lr.TypeError,Mr=Lr.WeakMap;if(kr||Pr.state){var Gr=Pr.state||(Pr.state=new Mr);Gr.get=Gr.get,Gr.has=Gr.has,Gr.set=Gr.set,mr=function(e,t){if(Gr.has(e))throw Cr("Object already initialized");return t.facade=e,Gr.set(e,t),t},Er=function(e){return Gr.get(e)||{}},br=function(e){return Gr.has(e)}}else{var $r=Fr("state");Dr[$r]=!0,mr=function(e,t){if(Nr(e,$r))throw Cr("Object already initialized");return t.facade=e,Tr(e,$r,t),t},Er=function(e){return Nr(e,$r)?e[$r]:{}},br=function(e){return Nr(e,$r)}}var Ur={set:mr,get:Er,has:br,enforce:function(e){return br(e)?Er(e):mr(e,{})},getterFor:function(e){return function(t){var r;if(!_r(t)||(r=Er(t)).type!==e)throw Cr("Incompatible receiver, "+e+" required");return r}}},Br=b,Xr=se,qr=it,Wr=w,Vr=dr.CONFIGURABLE,Hr=wr,zr=Ur.enforce,Yr=Ur.get,Kr=Object.defineProperty,Jr=Wr&&!Br((function(){return 8!==Kr((function(){}),"length",{value:8}).length})),Qr=String(String).split("String"),Zr=ar.exports=function(e,t,r){"Symbol("===String(t).slice(0,7)&&(t="["+String(t).replace(/^Symbol\(([^)]*)\)/,"$1")+"]"),r&&r.getter&&(t="get "+t),r&&r.setter&&(t="set "+t),(!qr(e,"name")||Vr&&e.name!==t)&&(Wr?Kr(e,"name",{value:t,configurable:!0}):e.name=t),Jr&&r&&qr(r,"arity")&&e.length!==r.arity&&Kr(e,"length",{value:r.arity});try{r&&qr(r,"constructor")&&r.constructor?Wr&&Kr(e,"prototype",{writable:!1}):e.prototype&&(e.prototype=void 0)}catch(e){}var n=zr(e);return qr(n,"source")||(n.source=Qr.join("string"==typeof t?t:"")),e};Function.prototype.toString=Zr((function(){return Xr(this)&&Yr(this).source||Hr(this)}),"toString");var en=ar.exports,tn=se,rn=Wt,nn=en,on=ze,sn=function(e,t,r,n){n||(n={});var i=n.enumerable,o=void 0!==n.name?n.name:t;if(tn(r)&&nn(r,o,n),n.global)i?e[t]=r:on(t,r);else{try{n.unsafe?e[t]&&(i=!0):delete e[t]}catch(e){}i?e[t]=r:rn.f(e,t,{value:r,enumerable:!1,configurable:!n.nonConfigurable,writable:!n.nonWritable})}return e},cn={},an=Math.ceil,un=Math.floor,ln=Math.trunc||function(e){var t=+e;return(t>0?un:an)(t)},fn=function(e){var t=+e;return t!=t||0===t?0:ln(t)},pn=fn,hn=Math.max,dn=Math.min,vn=function(e,t){var r=pn(e);return r<0?hn(r+t,0):dn(r,t)},gn=fn,yn=Math.min,mn=function(e){return e>0?yn(gn(e),9007199254740991):0},En=mn,bn=function(e){return En(e.length)},wn=re,On=vn,Sn=bn,Rn=function(e){return function(t,r,n){var i,o=wn(t),s=Sn(o),c=On(n,s);if(e&&r!=r){for(;s>c;)if((i=o[c++])!=i)return!0}else for(;s>c;c++)if((e||c in o)&&o[c]===r)return e||c||0;return!e&&-1}},In={includes:Rn(!0),indexOf:Rn(!1)},An=it,jn=re,xn=In.indexOf,kn=xr,Ln=q([].push),_n=function(e,t){var r,n=jn(e),i=0,o=[];for(r in n)!An(kn,r)&&An(n,r)&&Ln(o,r);for(;t.length>i;)An(n,r=t[i++])&&(~xn(o,r)||Ln(o,r));return o},Tn=["constructor","hasOwnProperty","isPrototypeOf","propertyIsEnumerable","toLocaleString","toString","valueOf"],Nn=_n,Pn=Tn.concat("length","prototype");cn.f=Object.getOwnPropertyNames||function(e){return Nn(e,Pn)};var Fn={};Fn.f=Object.getOwnPropertySymbols;var Dn=he,Cn=cn,Mn=Fn,Gn=Kt,$n=q([].concat),Un=Dn("Reflect","ownKeys")||function(e){var t=Cn.f(Gn(e)),r=Mn.f;return r?$n(t,r(e)):t},Bn=it,Xn=Un,qn=E,Wn=Wt,Vn=function(e,t,r){for(var n=Xn(t),i=Wn.f,o=qn.f,s=0;s<n.length;s++){var c=n[s];Bn(e,c)||r&&Bn(r,c)||i(e,c,o(t,c))}},Hn=b,zn=se,Yn=/#|\.prototype\./,Kn=function(e,t){var r=Qn[Jn(e)];return r==ei||r!=Zn&&(zn(t)?Hn(t):!!t)},Jn=Kn.normalize=function(e){return String(e).replace(Yn,".").toLowerCase()},Qn=Kn.data={},Zn=Kn.NATIVE="N",ei=Kn.POLYFILL="P",ti=Kn,ri=m,ni=E.f,ii=cr,oi=sn,si=ze,ci=Vn,ai=ti,ui=function(e,t){var r,n,i,o,s,c=e.target,a=e.global,u=e.stat;if(r=a?ri:u?ri[c]||si(c,{}):(ri[c]||{}).prototype)for(n in t){if(o=t[n],i=e.dontCallGetSet?(s=ni(r,n))&&s.value:r[n],!ai(a?n:c+(u?".":"#")+n,e.forced)&&void 0!==i){if(typeof o==typeof i)continue;ci(o,i)}(e.sham||i&&i.sham)&&ii(o,"sham",!0),oi(r,n,o,e)}},li=Wt.f,fi=it,pi=Et("toStringTag"),hi=function(e,t,r){e&&!r&&(e=e.prototype),e&&!fi(e,pi)&&li(e,pi,{configurable:!0,value:t})},di=m,vi=hi;ui({global:!0},{Reflect:{}}),vi(di.Reflect,"Reflect",!0);var gi={},yi=_n,mi=Tn,Ei=Object.keys||function(e){return yi(e,mi)},bi=w,wi=Vt,Oi=Wt,Si=Kt,Ri=re,Ii=Ei;gi.f=bi&&!wi?Object.defineProperties:function(e,t){Si(e);for(var r,n=Ri(t),i=Ii(t),o=i.length,s=0;o>s;)Oi.f(e,r=i[s++],n[r]);return e};var Ai,ji=he("document","documentElement"),xi=Kt,ki=gi,Li=Tn,_i=xr,Ti=ji,Ni=Nt,Pi=jr("IE_PROTO"),Fi=function(){},Di=function(e){return"<script>"+e+"<\/script>"},Ci=function(e){e.write(Di("")),e.close();var t=e.parentWindow.Object;return e=null,t},Mi=function(){try{Ai=new ActiveXObject("htmlfile")}catch(e){}var e,t;Mi="undefined"!=typeof document?document.domain&&Ai?Ci(Ai):((t=Ni("iframe")).style.display="none",Ti.appendChild(t),t.src=String("javascript:"),(e=t.contentWindow.document).open(),e.write(Di("document.F=Object")),e.close(),e.F):Ci(Ai);for(var r=Li.length;r--;)delete Mi.prototype[Li[r]];return Mi()};_i[Pi]=!0;var Gi=Object.create||function(e,t){var r;return null!==e?(Fi.prototype=xi(e),r=new Fi,Fi.prototype=null,r[Pi]=e):r=Mi(),void 0===t?r:ki.f(r,t)},$i=Et,Ui=Gi,Bi=Wt.f,Xi=$i("unscopables"),qi=Array.prototype;null==qi[Xi]&&Bi(qi,Xi,{configurable:!0,value:Ui(null)});var Wi,Vi,Hi,zi={},Yi=!b((function(){function e(){}return e.prototype.constructor=null,Object.getPrototypeOf(new e)!==e.prototype})),Ki=it,Ji=se,Qi=tt,Zi=Yi,eo=jr("IE_PROTO"),to=Object,ro=to.prototype,no=Zi?to.getPrototypeOf:function(e){var t=Qi(e);if(Ki(t,eo))return t[eo];var r=t.constructor;return Ji(r)&&t instanceof r?r.prototype:t instanceof to?ro:null},io=b,oo=se,so=ue,co=no,ao=sn,uo=Et("iterator"),lo=!1;[].keys&&("next"in(Hi=[].keys())?(Vi=co(co(Hi)))!==Object.prototype&&(Wi=Vi):lo=!0);var fo=!so(Wi)||io((function(){var e={};return Wi[uo].call(e)!==e}));fo&&(Wi={}),oo(Wi[uo])||ao(Wi,uo,(function(){return this}));var po={IteratorPrototype:Wi,BUGGY_SAFARI_ITERATORS:lo},ho=po.IteratorPrototype,vo=Gi,go=T,yo=hi,mo=zi,Eo=function(){return this},bo=se,wo=String,Oo=TypeError,So=q,Ro=Kt,Io=function(e){if("object"==typeof e||bo(e))return e;throw Oo("Can't set "+wo(e)+" as a prototype")},Ao=Object.setPrototypeOf||("__proto__"in{}?function(){var e,t=!1,r={};try{(e=So(Object.getOwnPropertyDescriptor(Object.prototype,"__proto__").set))(r,[]),t=r instanceof Array}catch(e){}return function(r,n){return Ro(r),Io(n),t?e(r,n):r.__proto__=n,r}}():void 0),jo=ui,xo=I,ko=se,Lo=function(e,t,r,n){var i=t+" Iterator";return e.prototype=vo(ho,{next:go(+!n,r)}),yo(e,i,!1),mo[i]=Eo,e},_o=no,To=Ao,No=hi,Po=cr,Fo=sn,Do=zi,Co=dr.PROPER,Mo=dr.CONFIGURABLE,Go=po.IteratorPrototype,$o=po.BUGGY_SAFARI_ITERATORS,Uo=Et("iterator"),Bo=function(){return this},Xo=re,qo=function(e){qi[Xi][e]=!0},Wo=zi,Vo=Ur,Ho=Wt.f,zo=function(e,t,r,n,i,o,s){Lo(r,t,n);var c,a,u,l=function(e){if(e===i&&v)return v;if(!$o&&e in h)return h[e];switch(e){case"keys":case"values":case"entries":return function(){return new r(this,e)}}return function(){return new r(this)}},f=t+" Iterator",p=!1,h=e.prototype,d=h[Uo]||h["@@iterator"]||i&&h[i],v=!$o&&d||l(i),g="Array"==t&&h.entries||d;if(g&&(c=_o(g.call(new e)))!==Object.prototype&&c.next&&(_o(c)!==Go&&(To?To(c,Go):ko(c[Uo])||Fo(c,Uo,Bo)),No(c,f,!0)),Co&&"values"==i&&d&&"values"!==d.name&&(Mo?Po(h,"name","values"):(p=!0,v=function(){return xo(d,this)})),i)if(a={values:l("values"),keys:o?v:l("keys"),entries:l("entries")},s)for(u in a)($o||p||!(u in h))&&Fo(h,u,a[u]);else jo({target:t,proto:!0,forced:$o||p},a);return h[Uo]!==v&&Fo(h,Uo,v,{name:i}),Do[t]=v,a},Yo=function(e,t){return{value:e,done:t}},Ko=w,Jo=Vo.set,Qo=Vo.getterFor("Array Iterator");zo(Array,"Array",(function(e,t){Jo(this,{type:"Array Iterator",target:Xo(e),index:0,kind:t})}),(function(){var e=Qo(this),t=e.target,r=e.kind,n=e.index++;return!t||n>=t.length?(e.target=void 0,Yo(void 0,!0)):Yo("keys"==r?n:"values"==r?t[n]:[n,t[n]],!1)}),"values");var Zo=Wo.Arguments=Wo.Array;if(qo("keys"),qo("values"),qo("entries"),Ko&&"values"!==Zo.name)try{Ho(Zo,"name",{value:"values"})}catch(e){}var es={};es[Et("toStringTag")]="z";var ts="[object z]"===String(es),rs=se,ns=U,is=Et("toStringTag"),os=Object,ss="Arguments"==ns(function(){return arguments}()),cs=ts?ns:function(e){var t,r,n;return void 0===e?"Undefined":null===e?"Null":"string"==typeof(r=function(e,t){try{return e[t]}catch(e){}}(t=os(e),is))?r:ss?ns(t):"Object"==(n=ns(t))&&rs(t.callee)?"Arguments":n},as=cs,us=String,ls=function(e){if("Symbol"===as(e))throw TypeError("Cannot convert a Symbol value to a string");return us(e)},fs=Kt,ps=function(){var e=fs(this),t="";return e.hasIndices&&(t+="d"),e.global&&(t+="g"),e.ignoreCase&&(t+="i"),e.multiline&&(t+="m"),e.dotAll&&(t+="s"),e.unicode&&(t+="u"),e.unicodeSets&&(t+="v"),e.sticky&&(t+="y"),t},hs=b,ds=m.RegExp,vs=hs((function(){var e=ds("a","y");return e.lastIndex=2,null!=e.exec("abcd")})),gs=vs||hs((function(){return!ds("a","y").sticky})),ys=vs||hs((function(){var e=ds("^r","gy");return e.lastIndex=2,null!=e.exec("str")})),ms={BROKEN_CARET:ys,MISSED_STICKY:gs,UNSUPPORTED_Y:vs},Es=b,bs=m.RegExp,ws=Es((function(){var e=bs(".","s");return!(e.dotAll&&e.exec("\n")&&"s"===e.flags)})),Os=b,Ss=m.RegExp,Rs=Os((function(){var e=Ss("(?<a>b)","g");return"b"!==e.exec("b").groups.a||"bc"!=="b".replace(e,"$<a>c")})),Is=I,As=q,js=ls,xs=ps,ks=ms,Ls=Gi,_s=Ur.get,Ts=ws,Ns=Rs,Ps=Qe("native-string-replace",String.prototype.replace),Fs=RegExp.prototype.exec,Ds=Fs,Cs=As("".charAt),Ms=As("".indexOf),Gs=As("".replace),$s=As("".slice),Us=function(){var e=/a/,t=/b*/g;return Is(Fs,e,"a"),Is(Fs,t,"a"),0!==e.lastIndex||0!==t.lastIndex}(),Bs=ks.BROKEN_CARET,Xs=void 0!==/()??/.exec("")[1];(Us||Xs||Bs||Ts||Ns)&&(Ds=function(e){var t,r,n,i,o,s,c,a=this,u=_s(a),l=js(e),f=u.raw;if(f)return f.lastIndex=a.lastIndex,t=Is(Ds,f,l),a.lastIndex=f.lastIndex,t;var p=u.groups,h=Bs&&a.sticky,d=Is(xs,a),v=a.source,g=0,y=l;if(h&&(d=Gs(d,"y",""),-1===Ms(d,"g")&&(d+="g"),y=$s(l,a.lastIndex),a.lastIndex>0&&(!a.multiline||a.multiline&&"\n"!==Cs(l,a.lastIndex-1))&&(v="(?: "+v+")",y=" "+y,g++),r=new RegExp("^(?:"+v+")",d)),Xs&&(r=new RegExp("^"+v+"$(?!\\s)",d)),Us&&(n=a.lastIndex),i=Is(Fs,h?r:a,y),h?i?(i.input=$s(i.input,g),i[0]=$s(i[0],g),i.index=a.lastIndex,a.lastIndex+=i[0].length):a.lastIndex=0:Us&&i&&(a.lastIndex=a.global?i.index+i[0].length:n),Xs&&i&&i.length>1&&Is(Ps,i[0],r,(function(){for(o=1;o<arguments.length-2;o++)void 0===arguments[o]&&(i[o]=void 0)})),i&&p)for(i.groups=s=Ls(null),o=0;o<p.length;o++)s[(c=p[o])[0]]=i[c[1]];return i});var qs=Ds;ui({target:"RegExp",proto:!0,forced:/./.exec!==qs},{exec:qs});var Ws=O,Vs=Function.prototype,Hs=Vs.apply,zs=Vs.call,Ys="object"==typeof Reflect&&Reflect.apply||(Ws?zs.bind(Hs):function(){return zs.apply(Hs,arguments)}),Ks=q,Js=sn,Qs=qs,Zs=b,ec=Et,tc=cr,rc=ec("species"),nc=RegExp.prototype,ic=function(e,t,r,n){var i=ec(e),o=!Zs((function(){var t={};return t[i]=function(){return 7},7!=""[e](t)})),s=o&&!Zs((function(){var t=!1,r=/a/;return"split"===e&&((r={}).constructor={},r.constructor[rc]=function(){return r},r.flags="",r[i]=/./[i]),r.exec=function(){return t=!0,null},r[i](""),!t}));if(!o||!s||r){var c=Ks(/./[i]),a=t(i,""[e],(function(e,t,r,n,i){var s=Ks(e),a=t.exec;return a===Qs||a===nc.exec?o&&!i?{done:!0,value:c(t,r,n)}:{done:!0,value:s(r,t,n)}:{done:!1}}));Js(String.prototype,e,a[0]),Js(nc,i,a[1])}n&&tc(nc[i],"sham",!0)},oc=q,sc=fn,cc=ls,ac=Z,uc=oc("".charAt),lc=oc("".charCodeAt),fc=oc("".slice),pc=function(e){return function(t,r){var n,i,o=cc(ac(t)),s=sc(r),c=o.length;return s<0||s>=c?e?"":void 0:(n=lc(o,s))<55296||n>56319||s+1===c||(i=lc(o,s+1))<56320||i>57343?e?uc(o,s):n:e?fc(o,s,s+2):i-56320+(n-55296<<10)+65536}},hc={codeAt:pc(!1),charAt:pc(!0)}.charAt,dc=function(e,t,r){return t+(r?hc(e,t).length:1)},vc=q,gc=tt,yc=Math.floor,mc=vc("".charAt),Ec=vc("".replace),bc=vc("".slice),wc=/\$([$&'`]|\d{1,2}|<[^>]*>)/g,Oc=/\$([$&'`]|\d{1,2})/g,Sc=function(e,t,r,n,i,o){var s=r+e.length,c=n.length,a=Oc;return void 0!==i&&(i=gc(i),a=wc),Ec(o,a,(function(o,a){var u;switch(mc(a,0)){case"$":return"$";case"&":return e;case"`":return bc(t,0,r);case"'":return bc(t,s);case"<":u=i[bc(a,1,-1)];break;default:var l=+a;if(0===l)return o;if(l>c){var f=yc(l/10);return 0===f?o:f<=c?void 0===n[f-1]?mc(a,1):n[f-1]+mc(a,1):o}u=n[l-1]}return void 0===u?"":u}))},Rc=I,Ic=Kt,Ac=se,jc=U,xc=qs,kc=TypeError,Lc=function(e,t){var r=e.exec;if(Ac(r)){var n=Rc(r,e,t);return null!==n&&Ic(n),n}if("RegExp"===jc(e))return Rc(xc,e,t);throw kc("RegExp#exec called on incompatible receiver")},_c=Ys,Tc=I,Nc=q,Pc=ic,Fc=b,Dc=Kt,Cc=se,Mc=K,Gc=fn,$c=mn,Uc=ls,Bc=Z,Xc=dc,qc=$e,Wc=Sc,Vc=Lc,Hc=Et("replace"),zc=Math.max,Yc=Math.min,Kc=Nc([].concat),Jc=Nc([].push),Qc=Nc("".indexOf),Zc=Nc("".slice),ea="$0"==="a".replace(/./,"$0"),ta=!!/./[Hc]&&""===/./[Hc]("a","$0"),ra=!Fc((function(){var e=/./;return e.exec=function(){var e=[];return e.groups={a:"7"},e},"7"!=="".replace(e,"$<a>")}));Pc("replace",(function(e,t,r){var n=ta?"$":"$0";return[function(e,r){var n=Bc(this),i=Mc(e)?void 0:qc(e,Hc);return i?Tc(i,e,n,r):Tc(t,Uc(n),e,r)},function(e,i){var o=Dc(this),s=Uc(e);if("string"==typeof i&&-1===Qc(i,n)&&-1===Qc(i,"$<")){var c=r(t,o,s,i);if(c.done)return c.value}var a=Cc(i);a||(i=Uc(i));var u=o.global;if(u){var l=o.unicode;o.lastIndex=0}for(var f=[];;){var p=Vc(o,s);if(null===p)break;if(Jc(f,p),!u)break;""===Uc(p[0])&&(o.lastIndex=Xc(s,$c(o.lastIndex),l))}for(var h,d="",v=0,g=0;g<f.length;g++){for(var y=Uc((p=f[g])[0]),m=zc(Yc(Gc(p.index),s.length),0),E=[],b=1;b<p.length;b++)Jc(E,void 0===(h=p[b])?h:String(h));var w=p.groups;if(a){var O=Kc([y],E,m,s);void 0!==w&&Jc(O,w);var S=Uc(_c(i,void 0,O))}else S=Wc(y,s,m,E,w,i);m>=v&&(d+=Zc(s,v,m)+S,v=m+y.length)}return d+Zc(s,v)}]}),!ra||!ea||ta);var na=ue,ia=U,oa=Et("match"),sa=function(e){var t;return na(e)&&(void 0!==(t=e[oa])?!!t:"RegExp"==ia(e))},ca=I,aa=it,ua=de,la=ps,fa=RegExp.prototype,pa=function(e){var t=e.flags;return void 0!==t||"flags"in fa||aa(e,"flags")||!ua(fa,e)?t:ca(la,e)},ha=ui,da=I,va=q,ga=Z,ya=se,ma=K,Ea=sa,ba=ls,wa=$e,Oa=pa,Sa=Sc,Ra=Et("replace"),Ia=TypeError,Aa=va("".indexOf);va("".replace);var ja=va("".slice),xa=Math.max,ka=function(e,t,r){return r>e.length?-1:""===t?r:Aa(e,t,r)};ha({target:"String",proto:!0},{replaceAll:function(e,t){var r,n,i,o,s,c,a,u,l=ga(this),f=0,p=0,h="";if(!ma(e)){if(Ea(e)&&(r=ba(ga(Oa(e))),!~Aa(r,"g")))throw Ia("`.replaceAll` does not allow non-global regexes");if(n=wa(e,Ra))return da(n,e,l,t)}for(i=ba(l),o=ba(e),(s=ya(t))||(t=ba(t)),c=o.length,a=xa(1,c),f=ka(i,o,0);-1!==f;)u=s?ba(t(o,f,i)):Sa(o,i,f,[],void 0,t),h+=ja(i,p,f)+u,p=f+c,f=ka(i,o,f+a);return p<i.length&&(h+=ja(i,p)),h}});var La=Wt.f,_a=function(e,t,r){r in e||La(e,r,{configurable:!0,get:function(){return t[r]},set:function(e){t[r]=e}})},Ta=se,Na=ue,Pa=Ao,Fa=function(e,t,r){var n,i;return Pa&&Ta(n=t.constructor)&&n!==r&&Na(i=n.prototype)&&i!==r.prototype&&Pa(e,i),e},Da=ls,Ca=ue,Ma=cr,Ga=Error,$a=q("".replace),Ua=String(Ga("zxcasd").stack),Ba=/\n\s*at [^:]*:[^\n]*/,Xa=Ba.test(Ua),qa=T,Wa=!b((function(){var e=Error("a");return!("stack"in e)||(Object.defineProperty(e,"stack",qa(1,7)),7!==e.stack)})),Va=he,Ha=it,za=cr,Ya=de,Ka=Ao,Ja=Vn,Qa=_a,Za=Fa,eu=function(e,t){return void 0===e?arguments.length<2?"":t:Da(e)},tu=function(e,t){Ca(t)&&"cause"in t&&Ma(e,"cause",t.cause)},ru=function(e,t){if(Xa&&"string"==typeof e&&!Ga.prepareStackTrace)for(;t--;)e=$a(e,Ba,"");return e},nu=Wa,iu=w,ou=ui,su=Ys,cu=function(e,t,r,n){var i="stackTraceLimit",o=n?2:1,s=e.split("."),c=s[s.length-1],a=Va.apply(null,s);if(a){var u=a.prototype;if(Ha(u,"cause")&&delete u.cause,!r)return a;var l=Va("Error"),f=t((function(e,t){var r=eu(n?t:e,void 0),i=n?new a(e):new a;return void 0!==r&&za(i,"message",r),nu&&za(i,"stack",ru(i.stack,2)),this&&Ya(u,this)&&Za(i,this,f),arguments.length>o&&tu(i,arguments[o]),i}));f.prototype=u,"Error"!==c?Ka?Ka(f,l):Ja(f,l,{name:!0}):iu&&i in a&&(Qa(f,a,i),Qa(f,a,"prepareStackTrace")),Ja(f,a);try{u.name!==c&&za(u,"name",c),u.constructor=f}catch(e){}return f}},au=m.WebAssembly,uu=7!==Error("e",{cause:7}).cause,lu=function(e,t){var r={};r[e]=cu(e,t,uu),ou({global:!0,constructor:!0,arity:1,forced:uu},r)},fu=function(e,t){if(au&&au[e]){var r={};r[e]=cu("WebAssembly."+e,t,uu),ou({target:"WebAssembly",stat:!0,constructor:!0,arity:1,forced:uu},r)}};lu("Error",(function(e){return function(t){return su(e,this,arguments)}})),lu("EvalError",(function(e){return function(t){return su(e,this,arguments)}})),lu("RangeError",(function(e){return function(t){return su(e,this,arguments)}})),lu("ReferenceError",(function(e){return function(t){return su(e,this,arguments)}})),lu("SyntaxError",(function(e){return function(t){return su(e,this,arguments)}})),lu("TypeError",(function(e){return function(t){return su(e,this,arguments)}})),lu("URIError",(function(e){return function(t){return su(e,this,arguments)}})),fu("CompileError",(function(e){return function(t){return su(e,this,arguments)}})),fu("LinkError",(function(e){return function(t){return su(e,this,arguments)}})),fu("RuntimeError",(function(e){return function(t){return su(e,this,arguments)}}));var pu=Ce,hu=O,du=q(q.bind),vu=U,gu=Array.isArray||function(e){return"Array"==vu(e)},yu=q,mu=b,Eu=se,bu=cs,wu=wr,Ou=function(){},Su=[],Ru=he("Reflect","construct"),Iu=/^\s*(?:class|function)\b/,Au=yu(Iu.exec),ju=!Iu.exec(Ou),xu=function(e){if(!Eu(e))return!1;try{return Ru(Ou,Su,e),!0}catch(e){return!1}},ku=function(e){if(!Eu(e))return!1;switch(bu(e)){case"AsyncFunction":case"GeneratorFunction":case"AsyncGeneratorFunction":return!1}try{return ju||!!Au(Iu,wu(e))}catch(e){return!0}};ku.sham=!0;var Lu=!Ru||mu((function(){var e;return xu(xu.call)||!xu(Object)||!xu((function(){e=!0}))||e}))?ku:xu,_u=gu,Tu=Lu,Nu=ue,Pu=Et("species"),Fu=Array,Du=function(e){var t;return _u(e)&&(t=e.constructor,(Tu(t)&&(t===Fu||_u(t.prototype))||Nu(t)&&null===(t=t[Pu]))&&(t=void 0)),void 0===t?Fu:t},Cu=function(e,t){return new(Du(e))(0===t?0:t)},Mu=function(e,t){return pu(e),void 0===t?e:hu?du(e,t):function(){return e.apply(t,arguments)}},Gu=Y,$u=tt,Uu=bn,Bu=Cu,Xu=q([].push),qu=function(e){var t=1==e,r=2==e,n=3==e,i=4==e,o=6==e,s=7==e,c=5==e||o;return function(a,u,l,f){for(var p,h,d=$u(a),v=Gu(d),g=Mu(u,l),y=Uu(v),m=0,E=f||Bu,b=t?E(a,y):r||s?E(a,0):void 0;y>m;m++)if((c||m in v)&&(h=g(p=v[m],m,d),e))if(t)b[m]=h;else if(h)switch(e){case 3:return!0;case 5:return p;case 6:return m;case 2:Xu(b,p)}else switch(e){case 4:return!1;case 7:Xu(b,p)}return o?-1:n||i?i:b}},Wu={forEach:qu(0),map:qu(1),filter:qu(2),some:qu(3),every:qu(4),find:qu(5),findIndex:qu(6),filterReject:qu(7)},Vu=b,Hu=Oe,zu=Et("species"),Yu=function(e){return Hu>=51||!Vu((function(){var t=[];return(t.constructor={})[zu]=function(){return{foo:1}},1!==t[e](Boolean).foo}))},Ku=Wu.map;ui({target:"Array",proto:!0,forced:!Yu("map")},{map:function(e){return Ku(this,e,arguments.length>1?arguments[1]:void 0)}});var Ju,Qu,Zu=ui,el=I,tl=se,rl=Kt,nl=ls,il=(Ju=!1,(Qu=/[ac]/).exec=function(){return Ju=!0,/./.exec.apply(this,arguments)},!0===Qu.test("abc")&&Ju),ol=/./.test;Zu({target:"RegExp",proto:!0,forced:!il},{test:function(e){var t=rl(this),r=nl(e),n=t.exec;if(!tl(n))return el(ol,t,r);var i=el(n,t,r);return null!==i&&(rl(i),!0)}});var sl=I,cl=Kt,al=K,ul=mn,ll=ls,fl=Z,pl=$e,hl=dc,dl=Lc;ic("match",(function(e,t,r){return[function(t){var r=fl(this),n=al(t)?void 0:pl(t,e);return n?sl(n,t,r):new RegExp(t)[e](ll(r))},function(e){var n=cl(this),i=ll(e),o=r(t,n,i);if(o.done)return o.value;if(!n.global)return dl(n,i);var s=n.unicode;n.lastIndex=0;for(var c,a=[],u=0;null!==(c=dl(n,i));){var l=ll(c[0]);a[u]=l,""===l&&(n.lastIndex=hl(i,ul(n.lastIndex),s)),u++}return 0===u?null:a}]}));var vl=he,gl=Wt,yl=w,ml=Et("species"),El=w,bl=m,wl=q,Ol=ti,Sl=Fa,Rl=cr,Il=cn.f,Al=de,jl=sa,xl=ls,kl=pa,Ll=ms,_l=_a,Tl=sn,Nl=b,Pl=it,Fl=Ur.enforce,Dl=function(e){var t=vl(e),r=gl.f;yl&&t&&!t[ml]&&r(t,ml,{configurable:!0,get:function(){return this}})},Cl=ws,Ml=Rs,Gl=Et("match"),$l=bl.RegExp,Ul=$l.prototype,Bl=bl.SyntaxError,Xl=wl(Ul.exec),ql=wl("".charAt),Wl=wl("".replace),Vl=wl("".indexOf),Hl=wl("".slice),zl=/^\?<[^\s\d!#%&*+<=>@^][^\s!#%&*+<=>@^]*>/,Yl=/a/g,Kl=/a/g,Jl=new $l(Yl)!==Yl,Ql=Ll.MISSED_STICKY,Zl=Ll.UNSUPPORTED_Y,ef=El&&(!Jl||Ql||Cl||Ml||Nl((function(){return Kl[Gl]=!1,$l(Yl)!=Yl||$l(Kl)==Kl||"/a/i"!=$l(Yl,"i")})));if(Ol("RegExp",ef)){for(var tf=function(e,t){var r,n,i,o,s,c,a=Al(Ul,this),u=jl(e),l=void 0===t,f=[],p=e;if(!a&&u&&l&&e.constructor===tf)return e;if((u||Al(Ul,e))&&(e=e.source,l&&(t=kl(p))),e=void 0===e?"":xl(e),t=void 0===t?"":xl(t),p=e,Cl&&"dotAll"in Yl&&(n=!!t&&Vl(t,"s")>-1)&&(t=Wl(t,/s/g,"")),r=t,Ql&&"sticky"in Yl&&(i=!!t&&Vl(t,"y")>-1)&&Zl&&(t=Wl(t,/y/g,"")),Ml&&(o=function(e){for(var t,r=e.length,n=0,i="",o=[],s={},c=!1,a=!1,u=0,l="";n<=r;n++){if("\\"===(t=ql(e,n)))t+=ql(e,++n);else if("]"===t)c=!1;else if(!c)switch(!0){case"["===t:c=!0;break;case"("===t:Xl(zl,Hl(e,n+1))&&(n+=2,a=!0),i+=t,u++;continue;case">"===t&&a:if(""===l||Pl(s,l))throw new Bl("Invalid capture group name");s[l]=!0,o[o.length]=[l,u],a=!1,l="";continue}a?l+=t:i+=t}return[i,o]}(e),e=o[0],f=o[1]),s=Sl($l(e,t),a?this:Ul,tf),(n||i||f.length)&&(c=Fl(s),n&&(c.dotAll=!0,c.raw=tf(function(e){for(var t,r=e.length,n=0,i="",o=!1;n<=r;n++)"\\"!==(t=ql(e,n))?o||"."!==t?("["===t?o=!0:"]"===t&&(o=!1),i+=t):i+="[\\s\\S]":i+=t+ql(e,++n);return i}(e),r)),i&&(c.sticky=!0),f.length&&(c.groups=f)),e!==p)try{Rl(s,"source",""===p?"(?:)":p)}catch(e){}return s},rf=Il($l),nf=0;rf.length>nf;)_l(tf,$l,rf[nf++]);Ul.constructor=tf,tf.prototype=Ul,Tl(bl,"RegExp",tf,{constructor:!0})}Dl("RegExp");var of=en,sf=Wt,cf=w,af=ws,uf=U,lf=function(e,t,r){return r.get&&of(r.get,t,{getter:!0}),r.set&&of(r.set,t,{setter:!0}),sf.f(e,t,r)},ff=Ur.get,pf=RegExp.prototype,hf=TypeError;cf&&af&&lf(pf,"dotAll",{configurable:!0,get:function(){if(this!==pf){if("RegExp"===uf(this))return!!ff(this).dotAll;throw hf("Incompatible receiver, RegExp required")}}});var df=Z,vf=ls,gf=q("".replace),yf="[\t\n\v\f\r                　\u2028\u2029\ufeff]",mf=RegExp("^"+yf+yf+"*"),Ef=RegExp(yf+yf+"*$"),bf=function(e){return function(t){var r=vf(df(t));return 1&e&&(r=gf(r,mf,"")),2&e&&(r=gf(r,Ef,"")),r}},wf={start:bf(1),end:bf(2),trim:bf(3)},Of=dr.PROPER,Sf=b,Rf="\t\n\v\f\r                　\u2028\u2029\ufeff",If=function(e){return Sf((function(){return!!Rf[e]()||"​᠎"!=="​᠎"[e]()||Of&&Rf[e].name!==e}))},Af=wf.trim;ui({target:"String",proto:!0,forced:If("trim")},{trim:function(){return Af(this)}});var jf=q([].slice),xf=ui,kf=he,Lf=Ys,_f=I,Tf=q,Nf=b,Pf=gu,Ff=se,Df=ue,Cf=_e,Mf=jf,Gf=Ie,$f=kf("JSON","stringify"),Uf=Tf(/./.exec),Bf=Tf("".charAt),Xf=Tf("".charCodeAt),qf=Tf("".replace),Wf=Tf(1..toString),Vf=/[\uD800-\uDFFF]/g,Hf=/^[\uD800-\uDBFF]$/,zf=/^[\uDC00-\uDFFF]$/,Yf=!Gf||Nf((function(){var e=kf("Symbol")();return"[null]"!=$f([e])||"{}"!=$f({a:e})||"{}"!=$f(Object(e))})),Kf=Nf((function(){return'"\\udf06\\ud834"'!==$f("\udf06\ud834")||'"\\udead"'!==$f("\udead")})),Jf=function(e,t){var r=Mf(arguments),n=t;if((Df(t)||void 0!==e)&&!Cf(e))return Pf(t)||(t=function(e,t){if(Ff(n)&&(t=_f(n,this,e,t)),!Cf(t))return t}),r[1]=t,Lf($f,null,r)},Qf=function(e,t,r){var n=Bf(r,t-1),i=Bf(r,t+1);return Uf(Hf,e)&&!Uf(zf,i)||Uf(zf,e)&&!Uf(Hf,n)?"\\u"+Wf(Xf(e,0),16):e};$f&&xf({target:"JSON",stat:!0,arity:3,forced:Yf||Kf},{stringify:function(e,t,r){var n=Mf(arguments),i=Lf(Yf?Jf:$f,null,n);return Kf&&"string"==typeof i?qf(i,Vf,Qf):i}});var Zf=kt,ep=Wt,tp=T,rp=function(e,t,r){var n=Zf(t);n in e?ep.f(e,n,tp(0,r)):e[n]=r},np=ui,ip=gu,op=Lu,sp=ue,cp=vn,ap=bn,up=re,lp=rp,fp=Et,pp=jf,hp=Yu("slice"),dp=fp("species"),vp=Array,gp=Math.max;np({target:"Array",proto:!0,forced:!hp},{slice:function(e,t){var r,n,i,o=up(this),s=ap(o),c=cp(e,s),a=cp(void 0===t?s:t,s);if(ip(o)&&(r=o.constructor,(op(r)&&(r===vp||ip(r.prototype))||sp(r)&&null===(r=r[dp]))&&(r=void 0),r===vp||void 0===r))return pp(o,c,a);for(n=new(void 0===r?vp:r)(gp(a-c,0)),i=0;c<a;c++,i++)c in o&&lp(n,i,o[c]);return n.length=i,n}});var yp=b,mp=function(e,t){var r=[][e];return!!r&&yp((function(){r.call(null,t||function(){return 1},1)}))},Ep=ui,bp=In.indexOf,wp=mp,Op=q([].indexOf),Sp=!!Op&&1/Op([1],1,-0)<0,Rp=wp("indexOf");Ep({target:"Array",proto:!0,forced:Sp||!Rp},{indexOf:function(e){var t=arguments.length>1?arguments[1]:void 0;return Sp?Op(this,e,t)||0:bp(this,e,t)}});var Ip=w,Ap=gu,jp=TypeError,xp=Object.getOwnPropertyDescriptor,kp=Ip&&!function(){if(void 0!==this)return!0;try{Object.defineProperty([],"length",{writable:!1}).length=1}catch(e){return e instanceof TypeError}}()?function(e,t){if(Ap(e)&&!xp(e,"length").writable)throw jp("Cannot set read only .length");return e.length=t}:function(e,t){return e.length=t},Lp=TypeError,_p=function(e){if(e>9007199254740991)throw Lp("Maximum allowed index exceeded");return e},Tp=ui,Np=tt,Pp=bn,Fp=kp,Dp=_p,Cp=b((function(){return 4294967297!==[].push.call({length:4294967296},1)})),Mp=!function(){try{Object.defineProperty([],"length",{writable:!1}).push()}catch(e){return e instanceof TypeError}}();Tp({target:"Array",proto:!0,arity:1,forced:Cp||Mp},{push:function(e){var t=Np(this),r=Pp(t),n=arguments.length;Dp(r+n);for(var i=0;i<n;i++)t[r]=arguments[i],r++;return Fp(t,r),r}});var Gp=ui,$p=b,Up=gu,Bp=ue,Xp=tt,qp=bn,Wp=_p,Vp=rp,Hp=Cu,zp=Yu,Yp=Oe,Kp=Et("isConcatSpreadable"),Jp=Yp>=51||!$p((function(){var e=[];return e[Kp]=!1,e.concat()[0]!==e})),Qp=zp("concat"),Zp=function(e){if(!Bp(e))return!1;var t=e[Kp];return void 0!==t?!!t:Up(e)};Gp({target:"Array",proto:!0,arity:1,forced:!Jp||!Qp},{concat:function(e){var t,r,n,i,o,s=Xp(this),c=Hp(s,0),a=0;for(t=-1,n=arguments.length;t<n;t++)if(Zp(o=-1===t?s:arguments[t]))for(i=qp(o),Wp(a+i),r=0;r<i;r++,a++)r in o&&Vp(c,a,o[r]);else Wp(a+1),Vp(c,a++,o);return c.length=a,c}});var eh=Ne,th=TypeError,rh=function(e,t){if(!delete e[t])throw th("Cannot delete property "+eh(t)+" of "+eh(e))},nh=ui,ih=tt,oh=bn,sh=kp,ch=rh,ah=_p,uh=1!==[].unshift(0),lh=!function(){try{Object.defineProperty([],"length",{writable:!1}).unshift()}catch(e){return e instanceof TypeError}}();nh({target:"Array",proto:!0,arity:1,forced:uh||lh},{unshift:function(e){var t=ih(this),r=oh(t),n=arguments.length;if(n){ah(r+n);for(var i=r;i--;){var o=i+n;i in t?t[o]=t[i]:ch(t,o)}for(var s=0;s<n;s++)t[s]=arguments[s]}return sh(t,r+n)}});var fh,ph,hh,dh={},vh={};function gh(){if(fh)return vh;fh=1;var e=l.default,t="win32"===process.platform,r=u.default,n=process.env.NODE_DEBUG&&/fs/.test(process.env.NODE_DEBUG);function i(e){return"function"==typeof e?e:function(){var e;if(n){var t=new Error;e=function(e){e&&(t.message=e.message,r(e=t))}}else e=r;return e;function r(e){if(e){if(process.throwDeprecation)throw e;if(!process.noDeprecation){var t="fs: missing callback "+(e.stack||e.message);process.traceDeprecation?console.trace(t):console.error(t)}}}}()}if(e.normalize,t)var o=/(.*?)(?:[\/\\]+|$)/g;else o=/(.*?)(?:[\/]+|$)/g;if(t)var s=/^(?:[a-zA-Z]:|[\\\/]{2}[^\\\/]+[\\\/][^\\\/]+)?[\\\/]*/;else s=/^[\/]*/;return vh.realpathSync=function(n,i){if(n=e.resolve(n),i&&Object.prototype.hasOwnProperty.call(i,n))return i[n];var c,a,u,l,f=n,p={},h={};function d(){var e=s.exec(n);c=e[0].length,a=e[0],u=e[0],l="",t&&!h[u]&&(r.lstatSync(u),h[u]=!0)}for(d();c<n.length;){o.lastIndex=c;var v=o.exec(n);if(l=a,a+=v[0],u=l+v[1],c=o.lastIndex,!(h[u]||i&&i[u]===u)){var g;if(i&&Object.prototype.hasOwnProperty.call(i,u))g=i[u];else{var y=r.lstatSync(u);if(!y.isSymbolicLink()){h[u]=!0,i&&(i[u]=u);continue}var m=null;if(!t){var E=y.dev.toString(32)+":"+y.ino.toString(32);p.hasOwnProperty(E)&&(m=p[E])}null===m&&(r.statSync(u),m=r.readlinkSync(u)),g=e.resolve(l,m),i&&(i[u]=g),t||(p[E]=m)}n=e.resolve(g,n.slice(c)),d()}}return i&&(i[f]=n),n},vh.realpath=function(n,c,a){if("function"!=typeof a&&(a=i(c),c=null),n=e.resolve(n),c&&Object.prototype.hasOwnProperty.call(c,n))return process.nextTick(a.bind(null,null,c[n]));var u,l,f,p,h=n,d={},v={};function g(){var e=s.exec(n);u=e[0].length,l=e[0],f=e[0],p="",t&&!v[f]?r.lstat(f,(function(e){if(e)return a(e);v[f]=!0,y()})):process.nextTick(y)}function y(){if(u>=n.length)return c&&(c[h]=n),a(null,n);o.lastIndex=u;var e=o.exec(n);return p=l,l+=e[0],f=p+e[1],u=o.lastIndex,v[f]||c&&c[f]===f?process.nextTick(y):c&&Object.prototype.hasOwnProperty.call(c,f)?b(c[f]):r.lstat(f,m)}function m(e,n){if(e)return a(e);if(!n.isSymbolicLink())return v[f]=!0,c&&(c[f]=f),process.nextTick(y);if(!t){var i=n.dev.toString(32)+":"+n.ino.toString(32);if(d.hasOwnProperty(i))return E(null,d[i],f)}r.stat(f,(function(e){if(e)return a(e);r.readlink(f,(function(e,r){t||(d[i]=r),E(e,r)}))}))}function E(t,r,n){if(t)return a(t);var i=e.resolve(p,r);c&&(c[n]=i),b(i)}function b(t){n=e.resolve(t,n.slice(u)),g()}g()},vh}function yh(){if(hh)return ph;hh=1,ph=c,c.realpath=c,c.sync=a,c.realpathSync=a,c.monkeypatch=function(){e.realpath=c,e.realpathSync=a},c.unmonkeypatch=function(){e.realpath=t,e.realpathSync=r};var e=u.default,t=e.realpath,r=e.realpathSync,n=process.version,i=/^v[0-5]\./.test(n),o=gh();function s(e){return e&&"realpath"===e.syscall&&("ELOOP"===e.code||"ENOMEM"===e.code||"ENAMETOOLONG"===e.code)}function c(e,r,n){if(i)return t(e,r,n);"function"==typeof r&&(n=r,r=null),t(e,r,(function(t,i){s(t)?o.realpath(e,r,n):n(t,i)}))}function a(e,t){if(i)return r(e,t);try{return r(e,t)}catch(r){if(s(r))return o.realpathSync(e,t);throw r}}return ph}var mh=Wu.filter;ui({target:"Array",proto:!0,forced:!Yu("filter")},{filter:function(e){return mh(this,e,arguments.length>1?arguments[1]:void 0)}});var Eh,bh,wh,Oh,Sh,Rh,Ih,Ah,jh=Lu,xh=Ne,kh=TypeError,Lh=Kt,_h=function(e){if(jh(e))return e;throw kh(xh(e)+" is not a constructor")},Th=K,Nh=Et("species"),Ph=vn,Fh=bn,Dh=rp,Ch=Array,Mh=Math.max,Gh=function(e,t,r){for(var n=Fh(e),i=Ph(t,n),o=Ph(void 0===r?n:r,n),s=Ch(Mh(o-i,0)),c=0;i<o;i++,c++)Dh(s,c,e[i]);return s.length=c,s},$h=Ys,Uh=I,Bh=q,Xh=ic,qh=Kt,Wh=K,Vh=sa,Hh=Z,zh=function(e,t){var r,n=Lh(e).constructor;return void 0===n||Th(r=Lh(n)[Nh])?t:_h(r)},Yh=dc,Kh=mn,Jh=ls,Qh=$e,Zh=Gh,ed=Lc,td=qs,rd=b,nd=ms.UNSUPPORTED_Y,id=Math.min,od=[].push,sd=Bh(/./.exec),cd=Bh(od),ad=Bh("".slice),ud=!rd((function(){var e=/(?:)/,t=e.exec;e.exec=function(){return t.apply(this,arguments)};var r="ab".split(e);return 2!==r.length||"a"!==r[0]||"b"!==r[1]}));function ld(){if(bh)return Eh;bh=1,Eh=function(t,r){for(var n=[],i=0;i<t.length;i++){var o=r(t[i],i);e(o)?n.push.apply(n,o):n.push(o)}return n};var e=Array.isArray||function(e){return"[object Array]"===Object.prototype.toString.call(e)};return Eh}function fd(){if(Rh)return Sh;Rh=1;var e=ld(),t=function(){if(Oh)return wh;function e(e,n,i){e instanceof RegExp&&(e=t(e,i)),n instanceof RegExp&&(n=t(n,i));var o=r(e,n,i);return o&&{start:o[0],end:o[1],pre:i.slice(0,o[0]),body:i.slice(o[0]+e.length,o[1]),post:i.slice(o[1]+n.length)}}function t(e,t){var r=t.match(e);return r?r[0]:null}function r(e,t,r){var n,i,o,s,c,a=r.indexOf(e),u=r.indexOf(t,a+1),l=a;if(a>=0&&u>0){for(n=[],o=r.length;l>=0&&!c;)l==a?(n.push(l),a=r.indexOf(e,l+1)):1==n.length?c=[n.pop(),u]:((i=n.pop())<o&&(o=i,s=u),u=r.indexOf(t,l+1)),l=a<u&&a>=0?a:u;n.length&&(c=[o,s])}return c}return Oh=1,wh=e,e.range=r,wh}();Sh=function(e){if(!e)return[];"{}"===e.substr(0,2)&&(e="\\{\\}"+e.substr(2));return d(function(e){return e.split("\\\\").join(r).split("\\{").join(n).split("\\}").join(i).split("\\,").join(o).split("\\.").join(s)}(e),!0).map(a)};var r="\0SLASH"+Math.random()+"\0",n="\0OPEN"+Math.random()+"\0",i="\0CLOSE"+Math.random()+"\0",o="\0COMMA"+Math.random()+"\0",s="\0PERIOD"+Math.random()+"\0";function c(e){return parseInt(e,10)==e?parseInt(e,10):e.charCodeAt(0)}function a(e){return e.split(r).join("\\").split(n).join("{").split(i).join("}").split(o).join(",").split(s).join(".")}function u(e){if(!e)return[""];var r=[],n=t("{","}",e);if(!n)return e.split(",");var i=n.pre,o=n.body,s=n.post,c=i.split(",");c[c.length-1]+="{"+o+"}";var a=u(s);return s.length&&(c[c.length-1]+=a.shift(),c.push.apply(c,a)),r.push.apply(r,c),r}function l(e){return"{"+e+"}"}function f(e){return/^-?0\d/.test(e)}function p(e,t){return e<=t}function h(e,t){return e>=t}function d(r,n){var o=[],s=t("{","}",r);if(!s||/\$$/.test(s.pre))return[r];var a,v=/^-?\d+\.\.-?\d+(?:\.\.-?\d+)?$/.test(s.body),g=/^[a-zA-Z]\.\.[a-zA-Z](?:\.\.-?\d+)?$/.test(s.body),y=v||g,m=s.body.indexOf(",")>=0;if(!y&&!m)return s.post.match(/,.*\}/)?d(r=s.pre+"{"+s.body+i+s.post):[r];if(y)a=s.body.split(/\.\./);else if(1===(a=u(s.body)).length&&1===(a=d(a[0],!1).map(l)).length)return(w=s.post.length?d(s.post,!1):[""]).map((function(e){return s.pre+a[0]+e}));var E,b=s.pre,w=s.post.length?d(s.post,!1):[""];if(y){var O=c(a[0]),S=c(a[1]),R=Math.max(a[0].length,a[1].length),I=3==a.length?Math.abs(c(a[2])):1,A=p;S<O&&(I*=-1,A=h);var j=a.some(f);E=[];for(var x=O;A(x,S);x+=I){var k;if(g)"\\"===(k=String.fromCharCode(x))&&(k="");else if(k=String(x),j){var L=R-k.length;if(L>0){var _=new Array(L+1).join("0");k=x<0?"-"+_+k.slice(1):_+k}}E.push(k)}}else E=e(a,(function(e){return d(e,!1)}));for(var T=0;T<E.length;T++)for(var N=0;N<w.length;N++){var P=b+E[T]+w[N];(!n||y||P)&&o.push(P)}return o}return Sh}function pd(){if(Ah)return Ih;Ah=1,Ih=u,u.Minimatch=l;var e=function(){try{return require("path")}catch(e){}}()||{sep:"/"};u.sep=e.sep;var t=u.GLOBSTAR=l.GLOBSTAR={},r=fd(),n={"!":{open:"(?:(?!(?:",close:"))[^/]*?)"},"?":{open:"(?:",close:")?"},"+":{open:"(?:",close:")+"},"*":{open:"(?:",close:")*"},"@":{open:"(?:",close:")"}},i="[^/]",o="[^/]*?",s="().*{}+?[]^$\\!".split("").reduce((function(e,t){return e[t]=!0,e}),{});var c=/\/+/;function a(e,t){t=t||{};var r={};return Object.keys(e).forEach((function(t){r[t]=e[t]})),Object.keys(t).forEach((function(e){r[e]=t[e]})),r}function u(e,t,r){return p(t),r||(r={}),!(!r.nocomment&&"#"===t.charAt(0))&&new l(t,r).match(e)}function l(t,r){if(!(this instanceof l))return new l(t,r);p(t),r||(r={}),t=t.trim(),r.allowWindowsEscape||"/"===e.sep||(t=t.split(e.sep).join("/")),this.options=r,this.set=[],this.pattern=t,this.regexp=null,this.negate=!1,this.comment=!1,this.empty=!1,this.partial=!!r.partial,this.make()}function f(e,t){return t||(t=this instanceof l?this.options:{}),e=void 0===e?this.pattern:e,p(e),t.nobrace||!/\{(?:(?!\{).)*\}/.test(e)?[e]:r(e)}u.filter=function(e,t){return t=t||{},function(r,n,i){return u(r,e,t)}},u.defaults=function(e){if(!e||"object"!=typeof e||!Object.keys(e).length)return u;var t=u,r=function(r,n,i){return t(r,n,a(e,i))};return(r.Minimatch=function(r,n){return new t.Minimatch(r,a(e,n))}).defaults=function(r){return t.defaults(a(e,r)).Minimatch},r.filter=function(r,n){return t.filter(r,a(e,n))},r.defaults=function(r){return t.defaults(a(e,r))},r.makeRe=function(r,n){return t.makeRe(r,a(e,n))},r.braceExpand=function(r,n){return t.braceExpand(r,a(e,n))},r.match=function(r,n,i){return t.match(r,n,a(e,i))},r},l.defaults=function(e){return u.defaults(e).Minimatch},l.prototype.debug=function(){},l.prototype.make=function(){var e=this.pattern,t=this.options;if(!t.nocomment&&"#"===e.charAt(0))return void(this.comment=!0);if(!e)return void(this.empty=!0);this.parseNegate();var r=this.globSet=this.braceExpand();t.debug&&(this.debug=function(){console.error.apply(console,arguments)});this.debug(this.pattern,r),r=this.globParts=r.map((function(e){return e.split(c)})),this.debug(this.pattern,r),r=r.map((function(e,t,r){return e.map(this.parse,this)}),this),this.debug(this.pattern,r),r=r.filter((function(e){return-1===e.indexOf(!1)})),this.debug(this.pattern,r),this.set=r},l.prototype.parseNegate=function(){var e=this.pattern,t=!1,r=this.options,n=0;if(r.nonegate)return;for(var i=0,o=e.length;i<o&&"!"===e.charAt(i);i++)t=!t,n++;n&&(this.pattern=e.substr(n));this.negate=t},u.braceExpand=function(e,t){return f(e,t)},l.prototype.braceExpand=f;var p=function(e){if("string"!=typeof e)throw new TypeError("invalid pattern");if(e.length>65536)throw new TypeError("pattern is too long")};l.prototype.parse=function(e,r){p(e);var c=this.options;if("**"===e){if(!c.noglobstar)return t;e="*"}if(""===e)return"";var a,u="",l=!!c.nocase,f=!1,d=[],v=[],g=!1,y=-1,m=-1,E="."===e.charAt(0)?"":c.dot?"(?!(?:^|\\/)\\.{1,2}(?:$|\\/))":"(?!\\.)",b=this;function w(){if(a){switch(a){case"*":u+=o,l=!0;break;case"?":u+=i,l=!0;break;default:u+="\\"+a}b.debug("clearStateChar %j %j",a,u),a=!1}}for(var O,S=0,R=e.length;S<R&&(O=e.charAt(S));S++)if(this.debug("%s\t%s %s %j",e,S,u,O),f&&s[O])u+="\\"+O,f=!1;else switch(O){case"/":return!1;case"\\":w(),f=!0;continue;case"?":case"*":case"+":case"@":case"!":if(this.debug("%s\t%s %s %j <-- stateChar",e,S,u,O),g){this.debug("  in class"),"!"===O&&S===m+1&&(O="^"),u+=O;continue}b.debug("call clearStateChar %j",a),w(),a=O,c.noext&&w();continue;case"(":if(g){u+="(";continue}if(!a){u+="\\(";continue}d.push({type:a,start:S-1,reStart:u.length,open:n[a].open,close:n[a].close}),u+="!"===a?"(?:(?!(?:":"(?:",this.debug("plType %j %j",a,u),a=!1;continue;case")":if(g||!d.length){u+="\\)";continue}w(),l=!0;var I=d.pop();u+=I.close,"!"===I.type&&v.push(I),I.reEnd=u.length;continue;case"|":if(g||!d.length||f){u+="\\|",f=!1;continue}w(),u+="|";continue;case"[":if(w(),g){u+="\\"+O;continue}g=!0,m=S,y=u.length,u+=O;continue;case"]":if(S===m+1||!g){u+="\\"+O,f=!1;continue}var A=e.substring(m+1,S);try{RegExp("["+A+"]")}catch(e){var j=this.parse(A,h);u=u.substr(0,y)+"\\["+j[0]+"\\]",l=l||j[1],g=!1;continue}l=!0,g=!1,u+=O;continue;default:w(),f?f=!1:!s[O]||"^"===O&&g||(u+="\\"),u+=O}g&&(A=e.substr(m+1),j=this.parse(A,h),u=u.substr(0,y)+"\\["+j[0],l=l||j[1]);for(I=d.pop();I;I=d.pop()){var x=u.slice(I.reStart+I.open.length);this.debug("setting tail",u,I),x=x.replace(/((?:\\{2}){0,64})(\\?)\|/g,(function(e,t,r){return r||(r="\\"),t+t+r+"|"})),this.debug("tail=%j\n   %s",x,x,I,u);var k="*"===I.type?o:"?"===I.type?i:"\\"+I.type;l=!0,u=u.slice(0,I.reStart)+k+"\\("+x}w(),f&&(u+="\\\\");var L=!1;switch(u.charAt(0)){case"[":case".":case"(":L=!0}for(var _=v.length-1;_>-1;_--){var T=v[_],N=u.slice(0,T.reStart),P=u.slice(T.reStart,T.reEnd-8),F=u.slice(T.reEnd-8,T.reEnd),D=u.slice(T.reEnd);F+=D;var C=N.split("(").length-1,M=D;for(S=0;S<C;S++)M=M.replace(/\)[+*?]?/,"");var G="";""===(D=M)&&r!==h&&(G="$"),u=N+P+D+G+F}""!==u&&l&&(u="(?=.)"+u);L&&(u=E+u);if(r===h)return[u,l];if(!l)return function(e){return e.replace(/\\(.)/g,"$1")}(e);var $=c.nocase?"i":"";try{var U=new RegExp("^"+u+"$",$)}catch(e){return new RegExp("$.")}return U._glob=e,U._src=u,U};var h={};return u.makeRe=function(e,t){return new l(e,t||{}).makeRe()},l.prototype.makeRe=function(){if(this.regexp||!1===this.regexp)return this.regexp;var e=this.set;if(!e.length)return this.regexp=!1,this.regexp;var r=this.options,n=r.noglobstar?o:r.dot?"(?:(?!(?:\\/|^)(?:\\.{1,2})($|\\/)).)*?":"(?:(?!(?:\\/|^)\\.).)*?",i=r.nocase?"i":"",s=e.map((function(e){return e.map((function(e){return e===t?n:"string"==typeof e?function(e){return e.replace(/[-[\]{}()*+?.,\\^$|#\s]/g,"\\$&")}(e):e._src})).join("\\/")})).join("|");s="^(?:"+s+")$",this.negate&&(s="^(?!"+s+").*$");try{this.regexp=new RegExp(s,i)}catch(e){this.regexp=!1}return this.regexp},u.match=function(e,t,r){var n=new l(t,r=r||{});return e=e.filter((function(e){return n.match(e)})),n.options.nonull&&!e.length&&e.push(t),e},l.prototype.match=function(t,r){if(void 0===r&&(r=this.partial),this.debug("match",t,this.pattern),this.comment)return!1;if(this.empty)return""===t;if("/"===t&&r)return!0;var n=this.options;"/"!==e.sep&&(t=t.split(e.sep).join("/")),t=t.split(c),this.debug(this.pattern,"split",t);var i,o,s=this.set;for(this.debug(this.pattern,"set",s),o=t.length-1;o>=0&&!(i=t[o]);o--);for(o=0;o<s.length;o++){var a=s[o],u=t;if(n.matchBase&&1===a.length&&(u=[i]),this.matchOne(u,a,r))return!!n.flipNegate||!this.negate}return!n.flipNegate&&this.negate},l.prototype.matchOne=function(e,r,n){var i=this.options;this.debug("matchOne",{this:this,file:e,pattern:r}),this.debug("matchOne",e.length,r.length);for(var o=0,s=0,c=e.length,a=r.length;o<c&&s<a;o++,s++){this.debug("matchOne loop");var u,l=r[s],f=e[o];if(this.debug(r,l,f),!1===l)return!1;if(l===t){this.debug("GLOBSTAR",[r,l,f]);var p=o,h=s+1;if(h===a){for(this.debug("** at the end");o<c;o++)if("."===e[o]||".."===e[o]||!i.dot&&"."===e[o].charAt(0))return!1;return!0}for(;p<c;){var d=e[p];if(this.debug("\nglobstar while",e,p,r,h,d),this.matchOne(e.slice(p),r.slice(h),n))return this.debug("globstar found match!",p,c,d),!0;if("."===d||".."===d||!i.dot&&"."===d.charAt(0)){this.debug("dot detected!",e,p,r,h);break}this.debug("globstar swallow a segment, and continue"),p++}return!(!n||(this.debug("\n>>> no match, partial?",e,p,r,h),p!==c))}if("string"==typeof l?(u=f===l,this.debug("string match",l,f,u)):(u=f.match(l),this.debug("pattern match",l,f,u)),!u)return!1}if(o===c&&s===a)return!0;if(o===c)return n;if(s===a)return o===c-1&&""===e[o];throw new Error("wtf?")},Ih}Xh("split",(function(e,t,r){var n;return n="c"=="abbc".split(/(b)*/)[1]||4!="test".split(/(?:)/,-1).length||2!="ab".split(/(?:ab)*/).length||4!=".".split(/(.?)(.?)/).length||".".split(/()()/).length>1||"".split(/.?/).length?function(e,r){var n=Jh(Hh(this)),i=void 0===r?4294967295:r>>>0;if(0===i)return[];if(void 0===e)return[n];if(!Vh(e))return Uh(t,n,e,i);for(var o,s,c,a=[],u=(e.ignoreCase?"i":"")+(e.multiline?"m":"")+(e.unicode?"u":"")+(e.sticky?"y":""),l=0,f=new RegExp(e.source,u+"g");(o=Uh(td,f,n))&&!((s=f.lastIndex)>l&&(cd(a,ad(n,l,o.index)),o.length>1&&o.index<n.length&&$h(od,a,Zh(o,1)),c=o[0].length,l=s,a.length>=i));)f.lastIndex===o.index&&f.lastIndex++;return l===n.length?!c&&sd(f,"")||cd(a,""):cd(a,ad(n,l)),a.length>i?Zh(a,0,i):a}:"0".split(void 0,0).length?function(e,r){return void 0===e&&0===r?[]:Uh(t,this,e,r)}:t,[function(t,r){var i=Hh(this),o=Wh(t)?void 0:Qh(t,e);return o?Uh(o,t,i,r):Uh(n,Jh(i),t,r)},function(e,i){var o=qh(this),s=Jh(e),c=r(n,o,s,i,n!==t);if(c.done)return c.value;var a=zh(o,RegExp),u=o.unicode,l=(o.ignoreCase?"i":"")+(o.multiline?"m":"")+(o.unicode?"u":"")+(nd?"g":"y"),f=new a(nd?"^(?:"+o.source+")":o,l),p=void 0===i?4294967295:i>>>0;if(0===p)return[];if(0===s.length)return null===ed(f,s)?[s]:[];for(var h=0,d=0,v=[];d<s.length;){f.lastIndex=nd?0:d;var g,y=ed(f,nd?ad(s,d):s);if(null===y||(g=id(Kh(f.lastIndex+(nd?d:0)),s.length))===h)d=Yh(s,d,u);else{if(cd(v,ad(s,h,d)),v.length===p)return v;for(var m=1;m<=y.length-1;m++)if(cd(v,y[m]),v.length===p)return v;d=h=g}}return cd(v,ad(s,h)),v}]}),!ud,nd);var hd,dd,vd={exports:{}},gd={exports:{}};function yd(){if(dd)return vd.exports;dd=1;try{var e=require("util");if("function"!=typeof e.inherits)throw"";vd.exports=e.inherits}catch(e){vd.exports=(hd||(hd=1,"function"==typeof Object.create?gd.exports=function(e,t){t&&(e.super_=t,e.prototype=Object.create(t.prototype,{constructor:{value:e,enumerable:!1,writable:!0,configurable:!0}}))}:gd.exports=function(e,t){if(t){e.super_=t;var r=function(){};r.prototype=t.prototype,e.prototype=new r,e.prototype.constructor=e}}),gd.exports)}return vd.exports}var md,Ed={exports:{}};function bd(){if(md)return Ed.exports;function e(e){return"/"===e.charAt(0)}function t(e){var t=/^([a-zA-Z]:|[\\\/]{2}[^\\\/]+[\\\/]+[^\\\/]+)?([\\\/])?([\s\S]*?)$/.exec(e),r=t[1]||"",n=Boolean(r&&":"!==r.charAt(1));return Boolean(t[2]||n)}return md=1,Ed.exports="win32"===process.platform?t:e,Ed.exports.posix=e,Ed.exports.win32=t,Ed.exports}var wd=Gh,Od=Math.floor,Sd=function(e,t){var r=e.length,n=Od(r/2);return r<8?Rd(e,t):Id(e,Sd(wd(e,0,n),t),Sd(wd(e,n),t),t)},Rd=function(e,t){for(var r,n,i=e.length,o=1;o<i;){for(n=o,r=e[o];n&&t(e[n-1],r)>0;)e[n]=e[--n];n!==o++&&(e[n]=r)}return e},Id=function(e,t,r,n){for(var i=t.length,o=r.length,s=0,c=0;s<i||c<o;)e[s+c]=s<i&&c<o?n(t[s],r[c])<=0?t[s++]:r[c++]:s<i?t[s++]:r[c++];return e},Ad=Sd,jd=ve.match(/firefox\/(\d+)/i),xd=!!jd&&+jd[1],kd=/MSIE|Trident/.test(ve),Ld=ve.match(/AppleWebKit\/(\d+)\./),_d=!!Ld&&+Ld[1],Td=ui,Nd=q,Pd=Ce,Fd=tt,Dd=bn,Cd=rh,Md=ls,Gd=b,$d=Ad,Ud=mp,Bd=xd,Xd=kd,qd=Oe,Wd=_d,Vd=[],Hd=Nd(Vd.sort),zd=Nd(Vd.push),Yd=Gd((function(){Vd.sort(void 0)})),Kd=Gd((function(){Vd.sort(null)})),Jd=Ud("sort"),Qd=!Gd((function(){if(qd)return qd<70;if(!(Bd&&Bd>3)){if(Xd)return!0;if(Wd)return Wd<603;var e,t,r,n,i="";for(e=65;e<76;e++){switch(t=String.fromCharCode(e),e){case 66:case 69:case 70:case 72:r=3;break;case 68:case 71:r=4;break;default:r=2}for(n=0;n<47;n++)Vd.push({k:t+n,v:r})}for(Vd.sort((function(e,t){return t.v-e.v})),n=0;n<Vd.length;n++)t=Vd[n].k.charAt(0),i.charAt(i.length-1)!==t&&(i+=t);return"DGBEFHACIJK"!==i}}));Td({target:"Array",proto:!0,forced:Yd||!Kd||!Jd||!Qd},{sort:function(e){void 0!==e&&Pd(e);var t=Fd(this);if(Qd)return void 0===e?Hd(t):Hd(t,e);var r,n,i=[],o=Dd(t);for(n=0;n<o;n++)n in t&&zd(i,t[n]);for($d(i,function(e){return function(t,r){return void 0===r?-1:void 0===t?1:void 0!==e?+e(t,r)||0:Md(t)>Md(r)?1:-1}}(e)),r=Dd(i),n=0;n<r;)t[n]=i[n++];for(;n<o;)Cd(t,n++);return t}});var Zd,ev,tv,rv={};function nv(){if(Zd)return rv;function e(e,t){return Object.prototype.hasOwnProperty.call(e,t)}Zd=1,rv.setopts=function(n,s,u){u||(u={});if(u.matchBase&&-1===s.indexOf("/")){if(u.noglobstar)throw new Error("base matching requires globstar");s="**/"+s}n.silent=!!u.silent,n.pattern=s,n.strict=!1!==u.strict,n.realpath=!!u.realpath,n.realpathCache=u.realpathCache||Object.create(null),n.follow=!!u.follow,n.dot=!!u.dot,n.mark=!!u.mark,n.nodir=!!u.nodir,n.nodir&&(n.mark=!0);n.sync=!!u.sync,n.nounique=!!u.nounique,n.nonull=!!u.nonull,n.nosort=!!u.nosort,n.nocase=!!u.nocase,n.stat=!!u.stat,n.noprocess=!!u.noprocess,n.absolute=!!u.absolute,n.fs=u.fs||t,n.maxLength=u.maxLength||1/0,n.cache=u.cache||Object.create(null),n.statCache=u.statCache||Object.create(null),n.symlinks=u.symlinks||Object.create(null),function(e,t){e.ignore=t.ignore||[],Array.isArray(e.ignore)||(e.ignore=[e.ignore]);e.ignore.length&&(e.ignore=e.ignore.map(c))}(n,u),n.changedCwd=!1;var l=process.cwd();e(u,"cwd")?(n.cwd=r.resolve(u.cwd),n.changedCwd=n.cwd!==l):n.cwd=l;n.root=u.root||r.resolve(n.cwd,"/"),n.root=r.resolve(n.root),"win32"===process.platform&&(n.root=n.root.replace(/\\/g,"/"));n.cwdAbs=i(n.cwd)?n.cwd:a(n,n.cwd),"win32"===process.platform&&(n.cwdAbs=n.cwdAbs.replace(/\\/g,"/"));n.nomount=!!u.nomount,u.nonegate=!0,u.nocomment=!0,u.allowWindowsEscape=!1,n.minimatch=new o(s,u),n.options=n.minimatch.options},rv.ownProp=e,rv.makeAbs=a,rv.finish=function(e){for(var t=e.nounique,r=t?[]:Object.create(null),n=0,i=e.matches.length;n<i;n++){var o=e.matches[n];if(o&&0!==Object.keys(o).length){var c=Object.keys(o);t?r.push.apply(r,c):c.forEach((function(e){r[e]=!0}))}else if(e.nonull){var u=e.minimatch.globSet[n];t?r.push(u):r[u]=!0}}t||(r=Object.keys(r));e.nosort||(r=r.sort(s));if(e.mark){for(n=0;n<r.length;n++)r[n]=e._mark(r[n]);e.nodir&&(r=r.filter((function(t){var r=!/\/$/.test(t),n=e.cache[t]||e.cache[a(e,t)];return r&&n&&(r="DIR"!==n&&!Array.isArray(n)),r})))}e.ignore.length&&(r=r.filter((function(t){return!f(e,t)})));e.found=r},rv.mark=function(e,t){var r=a(e,t),n=e.cache[r],i=t;if(n){var o="DIR"===n||Array.isArray(n),s="/"===t.slice(-1);if(o&&!s?i+="/":!o&&s&&(i=i.slice(0,-1)),i!==t){var c=a(e,i);e.statCache[c]=e.statCache[r],e.cache[c]=e.cache[r]}}return i},rv.isIgnored=f,rv.childrenIgnored=function(e,t){return!!e.ignore.length&&e.ignore.some((function(e){return!(!e.gmatcher||!e.gmatcher.match(t))}))};var t=u.default,r=l.default,n=pd(),i=bd(),o=n.Minimatch;function s(e,t){return e.localeCompare(t,"en")}function c(e){var t=null;if("/**"===e.slice(-3)){var r=e.replace(/(\/\*\*)+$/,"");t=new o(r,{dot:!0})}return{matcher:new o(e,{dot:!0}),gmatcher:t}}function a(e,t){var n=t;return n="/"===t.charAt(0)?r.join(e.root,t):i(t)||""===t?t:e.changedCwd?r.resolve(e.cwd,t):r.resolve(t),"win32"===process.platform&&(n=n.replace(/\\/g,"/")),n}function f(e,t){return!!e.ignore.length&&e.ignore.some((function(e){return e.matcher.match(t)||!(!e.gmatcher||!e.gmatcher.match(t))}))}return rv}var iv,ov,sv=ui,cv=tt,av=vn,uv=fn,lv=bn,fv=kp,pv=_p,hv=Cu,dv=rp,vv=rh,gv=Yu("splice"),yv=Math.max,mv=Math.min;function Ev(){if(ov)return iv;return ov=1,iv=function e(t,r){if(t&&r)return e(t)(r);if("function"!=typeof t)throw new TypeError("need wrapper function");return Object.keys(t).forEach((function(e){n[e]=t[e]})),n;function n(){for(var e=new Array(arguments.length),r=0;r<e.length;r++)e[r]=arguments[r];var n=t.apply(this,e),i=e[e.length-1];return"function"==typeof n&&n!==i&&Object.keys(i).forEach((function(e){n[e]=i[e]})),n}},iv}sv({target:"Array",proto:!0,forced:!gv},{splice:function(e,t){var r,n,i,o,s,c,a=cv(this),u=lv(a),l=av(e,u),f=arguments.length;for(0===f?r=n=0:1===f?(r=0,n=u-l):(r=f-2,n=mv(yv(uv(t),0),u-l)),pv(u+r-n),i=hv(a,n),o=0;o<n;o++)(s=l+o)in a&&dv(i,o,a[s]);if(i.length=n,r<n){for(o=l;o<u-n;o++)c=o+r,(s=o+n)in a?a[c]=a[s]:vv(a,c);for(o=u;o>u-n+r;o--)vv(a,o-1)}else if(r>n)for(o=u-n;o>l;o--)c=o+r-1,(s=o+n-1)in a?a[c]=a[s]:vv(a,c);for(o=0;o<r;o++)a[o+l]=arguments[o+2];return fv(a,u-n+r),i}});var bv,wv,Ov,Sv,Rv,Iv,Av,jv,xv,kv,Lv,_v,Tv,Nv,Pv={exports:{}};function Fv(){if(bv)return Pv.exports;bv=1;var e=Ev();function t(e){var t=function t(){return t.called?t.value:(t.called=!0,t.value=e.apply(this,arguments))};return t.called=!1,t}function r(e){var t=function t(){if(t.called)throw new Error(t.onceError);return t.called=!0,t.value=e.apply(this,arguments)},r=e.name||"Function wrapped with `once`";return t.onceError=r+" shouldn't be called more than once",t.called=!1,t}return Pv.exports=e(t),Pv.exports.strict=e(r),t.proto=t((function(){Object.defineProperty(Function.prototype,"once",{value:function(){return t(this)},configurable:!0}),Object.defineProperty(Function.prototype,"onceStrict",{value:function(){return r(this)},configurable:!0})})),Pv.exports}function Dv(){if(Ov)return wv;Ov=1;var e=Ev(),t=Object.create(null),r=Fv();function n(e){for(var t=e.length,r=[],n=0;n<t;n++)r[n]=e[n];return r}return wv=e((function(e,i){return t[e]?(t[e].push(i),null):(t[e]=[i],function(e){return r((function r(){var i=t[e],o=i.length,s=n(arguments);try{for(var c=0;c<o;c++)i[c].apply(null,s)}finally{i.length>o?(i.splice(0,o),process.nextTick((function(){r.apply(null,s)}))):delete t[e]}}))}(e))})),wv}function Cv(){if(Rv)return Sv;Rv=1,Sv=m;var e=yh(),t=pd();t.Minimatch;var r=yd(),n=f.default.EventEmitter,i=l.default,o=p.default,s=bd(),c=function(){if(tv)return ev;tv=1,ev=f,f.GlobSync=h;var e=yh(),t=pd();t.Minimatch,Cv().Glob;var r=l.default,n=p.default,i=bd(),o=nv(),s=o.setopts,c=o.ownProp,a=o.childrenIgnored,u=o.isIgnored;function f(e,t){if("function"==typeof t||3===arguments.length)throw new TypeError("callback provided to sync glob\nSee: https://github.com/isaacs/node-glob/issues/167");return new h(e,t).found}function h(e,t){if(!e)throw new Error("must provide pattern");if("function"==typeof t||3===arguments.length)throw new TypeError("callback provided to sync glob\nSee: https://github.com/isaacs/node-glob/issues/167");if(!(this instanceof h))return new h(e,t);if(s(this,e,t),this.noprocess)return this;var r=this.minimatch.set.length;this.matches=new Array(r);for(var n=0;n<r;n++)this._process(this.minimatch.set[n],n,!1);this._finish()}return h.prototype._finish=function(){if(n.ok(this instanceof h),this.realpath){var t=this;this.matches.forEach((function(r,n){var i=t.matches[n]=Object.create(null);for(var o in r)try{o=t._makeAbs(o),i[e.realpathSync(o,t.realpathCache)]=!0}catch(e){if("stat"!==e.syscall)throw e;i[t._makeAbs(o)]=!0}}))}o.finish(this)},h.prototype._process=function(e,r,o){n.ok(this instanceof h);for(var s,c=0;"string"==typeof e[c];)c++;switch(c){case e.length:return void this._processSimple(e.join("/"),r);case 0:s=null;break;default:s=e.slice(0,c).join("/")}var u,l=e.slice(c);null===s?u=".":i(s)||i(e.map((function(e){return"string"==typeof e?e:"[*]"})).join("/"))?(s&&i(s)||(s="/"+s),u=s):u=s;var f=this._makeAbs(u);a(this,u)||(l[0]===t.GLOBSTAR?this._processGlobStar(s,u,f,l,r,o):this._processReaddir(s,u,f,l,r,o))},h.prototype._processReaddir=function(e,t,n,i,o,s){var c=this._readdir(n,s);if(c){for(var a=i[0],u=!!this.minimatch.negate,l=a._glob,f=this.dot||"."===l.charAt(0),p=[],h=0;h<c.length;h++)("."!==(g=c[h]).charAt(0)||f)&&(u&&!e?!g.match(a):g.match(a))&&p.push(g);var d=p.length;if(0!==d)if(1!==i.length||this.mark||this.stat)for(i.shift(),h=0;h<d;h++){var v;g=p[h],v=e?[e,g]:[g],this._process(v.concat(i),o,s)}else{this.matches[o]||(this.matches[o]=Object.create(null));for(h=0;h<d;h++){var g=p[h];e&&(g="/"!==e.slice(-1)?e+"/"+g:e+g),"/"!==g.charAt(0)||this.nomount||(g=r.join(this.root,g)),this._emitMatch(o,g)}}}},h.prototype._emitMatch=function(e,t){if(!u(this,t)){var r=this._makeAbs(t);if(this.mark&&(t=this._mark(t)),this.absolute&&(t=r),!this.matches[e][t]){if(this.nodir){var n=this.cache[r];if("DIR"===n||Array.isArray(n))return}this.matches[e][t]=!0,this.stat&&this._stat(t)}}},h.prototype._readdirInGlobStar=function(e){if(this.follow)return this._readdir(e,!1);var t,r;try{r=this.fs.lstatSync(e)}catch(e){if("ENOENT"===e.code)return null}var n=r&&r.isSymbolicLink();return this.symlinks[e]=n,n||!r||r.isDirectory()?t=this._readdir(e,!1):this.cache[e]="FILE",t},h.prototype._readdir=function(e,t){if(t&&!c(this.symlinks,e))return this._readdirInGlobStar(e);if(c(this.cache,e)){var r=this.cache[e];if(!r||"FILE"===r)return null;if(Array.isArray(r))return r}try{return this._readdirEntries(e,this.fs.readdirSync(e))}catch(t){return this._readdirError(e,t),null}},h.prototype._readdirEntries=function(e,t){if(!this.mark&&!this.stat)for(var r=0;r<t.length;r++){var n=t[r];n="/"===e?e+n:e+"/"+n,this.cache[n]=!0}return this.cache[e]=t,t},h.prototype._readdirError=function(e,t){switch(t.code){case"ENOTSUP":case"ENOTDIR":var r=this._makeAbs(e);if(this.cache[r]="FILE",r===this.cwdAbs){var n=new Error(t.code+" invalid cwd "+this.cwd);throw n.path=this.cwd,n.code=t.code,n}break;case"ENOENT":case"ELOOP":case"ENAMETOOLONG":case"UNKNOWN":this.cache[this._makeAbs(e)]=!1;break;default:if(this.cache[this._makeAbs(e)]=!1,this.strict)throw t;this.silent||console.error("glob error",t)}},h.prototype._processGlobStar=function(e,t,r,n,i,o){var s=this._readdir(r,o);if(s){var c=n.slice(1),a=e?[e]:[],u=a.concat(c);this._process(u,i,!1);var l=s.length;if(!this.symlinks[r]||!o)for(var f=0;f<l;f++)if("."!==s[f].charAt(0)||this.dot){var p=a.concat(s[f],c);this._process(p,i,!0);var h=a.concat(s[f],n);this._process(h,i,!0)}}},h.prototype._processSimple=function(e,t){var n=this._stat(e);if(this.matches[t]||(this.matches[t]=Object.create(null)),n){if(e&&i(e)&&!this.nomount){var o=/[\/\\]$/.test(e);"/"===e.charAt(0)?e=r.join(this.root,e):(e=r.resolve(this.root,e),o&&(e+="/"))}"win32"===process.platform&&(e=e.replace(/\\/g,"/")),this._emitMatch(t,e)}},h.prototype._stat=function(e){var t=this._makeAbs(e),r="/"===e.slice(-1);if(e.length>this.maxLength)return!1;if(!this.stat&&c(this.cache,t)){var n=this.cache[t];if(Array.isArray(n)&&(n="DIR"),!r||"DIR"===n)return n;if(r&&"FILE"===n)return!1}var i=this.statCache[t];if(!i){var o;try{o=this.fs.lstatSync(t)}catch(e){if(e&&("ENOENT"===e.code||"ENOTDIR"===e.code))return this.statCache[t]=!1,!1}if(o&&o.isSymbolicLink())try{i=this.fs.statSync(t)}catch(e){i=o}else i=o}return this.statCache[t]=i,n=!0,i&&(n=i.isDirectory()?"DIR":"FILE"),this.cache[t]=this.cache[t]||n,(!r||"FILE"!==n)&&n},h.prototype._mark=function(e){return o.mark(this,e)},h.prototype._makeAbs=function(e){return o.makeAbs(this,e)},ev}(),a=nv(),u=a.setopts,h=a.ownProp,d=Dv(),v=a.childrenIgnored,g=a.isIgnored,y=Fv();function m(e,t,r){if("function"==typeof t&&(r=t,t={}),t||(t={}),t.sync){if(r)throw new TypeError("callback provided to sync glob");return c(e,t)}return new b(e,t,r)}m.sync=c;var E=m.GlobSync=c.GlobSync;function b(e,t,r){if("function"==typeof t&&(r=t,t=null),t&&t.sync){if(r)throw new TypeError("callback provided to sync glob");return new E(e,t)}if(!(this instanceof b))return new b(e,t,r);u(this,e,t),this._didRealPath=!1;var n=this.minimatch.set.length;this.matches=new Array(n),"function"==typeof r&&(r=y(r),this.on("error",r),this.on("end",(function(e){r(null,e)})));var i=this;if(this._processing=0,this._emitQueue=[],this._processQueue=[],this.paused=!1,this.noprocess)return this;if(0===n)return s();for(var o=0;o<n;o++)this._process(this.minimatch.set[o],o,!1,s);function s(){--i._processing,i._processing<=0&&i._finish()}}return m.glob=m,m.hasMagic=function(e,t){var r=function(e,t){if(null===t||"object"!=typeof t)return e;for(var r=Object.keys(t),n=r.length;n--;)e[r[n]]=t[r[n]];return e}({},t);r.noprocess=!0;var n=new b(e,r).minimatch.set;if(!e)return!1;if(n.length>1)return!0;for(var i=0;i<n[0].length;i++)if("string"!=typeof n[0][i])return!0;return!1},m.Glob=b,r(b,n),b.prototype._finish=function(){if(o(this instanceof b),!this.aborted){if(this.realpath&&!this._didRealpath)return this._realpath();a.finish(this),this.emit("end",this.found)}},b.prototype._realpath=function(){if(!this._didRealpath){this._didRealpath=!0;var e=this.matches.length;if(0===e)return this._finish();for(var t=this,r=0;r<this.matches.length;r++)this._realpathSet(r,n)}function n(){0==--e&&t._finish()}},b.prototype._realpathSet=function(t,r){var n=this.matches[t];if(!n)return r();var i=Object.keys(n),o=this,s=i.length;if(0===s)return r();var c=this.matches[t]=Object.create(null);i.forEach((function(n,i){n=o._makeAbs(n),e.realpath(n,o.realpathCache,(function(e,i){e?"stat"===e.syscall?c[n]=!0:o.emit("error",e):c[i]=!0,0==--s&&(o.matches[t]=c,r())}))}))},b.prototype._mark=function(e){return a.mark(this,e)},b.prototype._makeAbs=function(e){return a.makeAbs(this,e)},b.prototype.abort=function(){this.aborted=!0,this.emit("abort")},b.prototype.pause=function(){this.paused||(this.paused=!0,this.emit("pause"))},b.prototype.resume=function(){if(this.paused){if(this.emit("resume"),this.paused=!1,this._emitQueue.length){var e=this._emitQueue.slice(0);this._emitQueue.length=0;for(var t=0;t<e.length;t++){var r=e[t];this._emitMatch(r[0],r[1])}}if(this._processQueue.length){var n=this._processQueue.slice(0);this._processQueue.length=0;for(t=0;t<n.length;t++){var i=n[t];this._processing--,this._process(i[0],i[1],i[2],i[3])}}}},b.prototype._process=function(e,r,n,i){if(o(this instanceof b),o("function"==typeof i),!this.aborted)if(this._processing++,this.paused)this._processQueue.push([e,r,n,i]);else{for(var c,a=0;"string"==typeof e[a];)a++;switch(a){case e.length:return void this._processSimple(e.join("/"),r,i);case 0:c=null;break;default:c=e.slice(0,a).join("/")}var u,l=e.slice(a);null===c?u=".":s(c)||s(e.map((function(e){return"string"==typeof e?e:"[*]"})).join("/"))?(c&&s(c)||(c="/"+c),u=c):u=c;var f=this._makeAbs(u);if(v(this,u))return i();l[0]===t.GLOBSTAR?this._processGlobStar(c,u,f,l,r,n,i):this._processReaddir(c,u,f,l,r,n,i)}},b.prototype._processReaddir=function(e,t,r,n,i,o,s){var c=this;this._readdir(r,o,(function(a,u){return c._processReaddir2(e,t,r,n,i,o,u,s)}))},b.prototype._processReaddir2=function(e,t,r,n,o,s,c,a){if(!c)return a();for(var u=n[0],l=!!this.minimatch.negate,f=u._glob,p=this.dot||"."===f.charAt(0),h=[],d=0;d<c.length;d++){if("."!==(g=c[d]).charAt(0)||p)(l&&!e?!g.match(u):g.match(u))&&h.push(g)}var v=h.length;if(0===v)return a();if(1===n.length&&!this.mark&&!this.stat){this.matches[o]||(this.matches[o]=Object.create(null));for(d=0;d<v;d++){var g=h[d];e&&(g="/"!==e?e+"/"+g:e+g),"/"!==g.charAt(0)||this.nomount||(g=i.join(this.root,g)),this._emitMatch(o,g)}return a()}n.shift();for(d=0;d<v;d++){g=h[d];e&&(g="/"!==e?e+"/"+g:e+g),this._process([g].concat(n),o,s,a)}a()},b.prototype._emitMatch=function(e,t){if(!this.aborted&&!g(this,t))if(this.paused)this._emitQueue.push([e,t]);else{var r=s(t)?t:this._makeAbs(t);if(this.mark&&(t=this._mark(t)),this.absolute&&(t=r),!this.matches[e][t]){if(this.nodir){var n=this.cache[r];if("DIR"===n||Array.isArray(n))return}this.matches[e][t]=!0;var i=this.statCache[r];i&&this.emit("stat",t,i),this.emit("match",t)}}},b.prototype._readdirInGlobStar=function(e,t){if(!this.aborted){if(this.follow)return this._readdir(e,!1,t);var r=this,n=d("lstat\0"+e,(function(n,i){if(n&&"ENOENT"===n.code)return t();var o=i&&i.isSymbolicLink();r.symlinks[e]=o,o||!i||i.isDirectory()?r._readdir(e,!1,t):(r.cache[e]="FILE",t())}));n&&r.fs.lstat(e,n)}},b.prototype._readdir=function(e,t,r){if(!this.aborted&&(r=d("readdir\0"+e+"\0"+t,r))){if(t&&!h(this.symlinks,e))return this._readdirInGlobStar(e,r);if(h(this.cache,e)){var n=this.cache[e];if(!n||"FILE"===n)return r();if(Array.isArray(n))return r(null,n)}this.fs.readdir(e,function(e,t,r){return function(n,i){n?e._readdirError(t,n,r):e._readdirEntries(t,i,r)}}(this,e,r))}},b.prototype._readdirEntries=function(e,t,r){if(!this.aborted){if(!this.mark&&!this.stat)for(var n=0;n<t.length;n++){var i=t[n];i="/"===e?e+i:e+"/"+i,this.cache[i]=!0}return this.cache[e]=t,r(null,t)}},b.prototype._readdirError=function(e,t,r){if(!this.aborted){switch(t.code){case"ENOTSUP":case"ENOTDIR":var n=this._makeAbs(e);if(this.cache[n]="FILE",n===this.cwdAbs){var i=new Error(t.code+" invalid cwd "+this.cwd);i.path=this.cwd,i.code=t.code,this.emit("error",i),this.abort()}break;case"ENOENT":case"ELOOP":case"ENAMETOOLONG":case"UNKNOWN":this.cache[this._makeAbs(e)]=!1;break;default:this.cache[this._makeAbs(e)]=!1,this.strict&&(this.emit("error",t),this.abort()),this.silent||console.error("glob error",t)}return r()}},b.prototype._processGlobStar=function(e,t,r,n,i,o,s){var c=this;this._readdir(r,o,(function(a,u){c._processGlobStar2(e,t,r,n,i,o,u,s)}))},b.prototype._processGlobStar2=function(e,t,r,n,i,o,s,c){if(!s)return c();var a=n.slice(1),u=e?[e]:[],l=u.concat(a);this._process(l,i,!1,c);var f=this.symlinks[r],p=s.length;if(f&&o)return c();for(var h=0;h<p;h++){if("."!==s[h].charAt(0)||this.dot){var d=u.concat(s[h],a);this._process(d,i,!0,c);var v=u.concat(s[h],n);this._process(v,i,!0,c)}}c()},b.prototype._processSimple=function(e,t,r){var n=this;this._stat(e,(function(i,o){n._processSimple2(e,t,i,o,r)}))},b.prototype._processSimple2=function(e,t,r,n,o){if(this.matches[t]||(this.matches[t]=Object.create(null)),!n)return o();if(e&&s(e)&&!this.nomount){var c=/[\/\\]$/.test(e);"/"===e.charAt(0)?e=i.join(this.root,e):(e=i.resolve(this.root,e),c&&(e+="/"))}"win32"===process.platform&&(e=e.replace(/\\/g,"/")),this._emitMatch(t,e),o()},b.prototype._stat=function(e,t){var r=this._makeAbs(e),n="/"===e.slice(-1);if(e.length>this.maxLength)return t();if(!this.stat&&h(this.cache,r)){var i=this.cache[r];if(Array.isArray(i)&&(i="DIR"),!n||"DIR"===i)return t(null,i);if(n&&"FILE"===i)return t()}var o=this.statCache[r];if(void 0!==o){if(!1===o)return t(null,o);var s=o.isDirectory()?"DIR":"FILE";return n&&"FILE"===s?t():t(null,s,o)}var c=this,a=d("stat\0"+r,(function(n,i){if(i&&i.isSymbolicLink())return c.fs.stat(r,(function(n,o){n?c._stat2(e,r,null,i,t):c._stat2(e,r,n,o,t)}));c._stat2(e,r,n,i,t)}));a&&c.fs.lstat(r,a)},b.prototype._stat2=function(e,t,r,n,i){if(r&&("ENOENT"===r.code||"ENOTDIR"===r.code))return this.statCache[t]=!1,i();var o="/"===e.slice(-1);if(this.statCache[t]=n,"/"===t.slice(-1)&&n&&!n.isDirectory())return i(null,!1,n);var s=!0;return n&&(s=n.isDirectory()?"DIR":"FILE"),this.cache[t]=this.cache[t]||s,o&&"FILE"===s?i():i(null,s,n)},Sv}function Mv(){if(Iv)return dh;Iv=1;var e=a.default,t=u.default,r=Cv(),n=Iy(),i=Object.create(n);dh.extend=Object.assign;var o=Boolean(process.versions.electron),s={fatal:!1,globOptions:{},maxdepth:255,noglob:!1,silent:!1,verbose:!1,execPath:null,bufLength:65536},c={reset:function(){Object.assign(this,s),o||(this.execPath=process.execPath)},resetForTesting:function(){this.reset(),this.silent=!0}};c.reset(),dh.config=c;var l={error:null,errorCode:0,currentCmd:"shell.js"};function f(e){return"object"==typeof e&&null!==e}function p(){c.silent||console.error.apply(console,arguments)}function h(e){if("string"!=typeof e)throw new TypeError("input must be a string");return e.replace(/\\/g,"/")}function d(e,t,r){if("string"!=typeof e)throw new Error("msg must be a string");var n={continue:!1,code:1,prefix:l.currentCmd+": ",silent:!1};"number"==typeof t&&f(r)?r.code=t:f(t)?r=t:"number"==typeof t?r={code:t}:"number"!=typeof t&&(r={}),r=Object.assign({},n,r),l.errorCode||(l.errorCode=r.code);var i=h(r.prefix+e);if(l.error=l.error?l.error+"\n":"",l.error+=i,c.fatal)throw new Error(i);if(e.length>0&&!r.silent&&p(i),!r.continue)throw{msg:"earlyExit",retValue:new v("",l.error,l.errorCode)}}function v(e,t,r){var n;return e instanceof Array?((n=e).stdout=e.join("\n"),e.length>0&&(n.stdout+="\n")):(n=new String(e)).stdout=e,n.stderr=t,n.code=r,w.forEach((function(e){n[e]=i[e].bind(n)})),n}function g(e,t,r){if("string"!=typeof e&&!f(e))throw new Error("options must be strings or key-value pairs");if(!f(t))throw new Error("parseOptions() internal error: map must be an object");if(r&&!f(r))throw new Error("parseOptions() internal error: errorOptions must be object");if("--"===e)return{};var n={};if(Object.keys(t).forEach((function(e){var r=t[e];"!"!==r[0]&&(n[r]=!1)})),""===e)return n;if("string"==typeof e){if("-"!==e[0])throw new Error("Options string must start with a '-'");e.slice(1).split("").forEach((function(e){if(e in t){var i=t[e];"!"===i[0]?n[i.slice(1)]=!1:n[i]=!0}else d("option not recognized: "+e,r||{})}))}else Object.keys(e).forEach((function(i){var o=i[1];if(o in t){var s=t[o];n[s]=e[i]}else d("option not recognized: "+o,r||{})}));return n}function y(e){if(!Array.isArray(e))throw new TypeError("must be an array");var t=[];return e.forEach((function(e){if("string"!=typeof e)t.push(e);else{var n;try{n=(n=r.sync(e,c.globOptions)).length>0?n:[e]}catch(t){n=[e]}t=t.concat(n)}})),t}dh.state=l,delete process.env.OLDPWD,dh.isObject=f,dh.log=p,dh.convertErrorOutput=h,dh.error=d,dh.ShellString=v,dh.parseOptions=g,dh.expand=y;var m="function"==typeof Buffer.alloc?function(e){return Buffer.alloc(e||c.bufLength)}:function(e){return new Buffer(e||c.bufLength)};function E(t,r,n){return n=n||{},function(){var i=null;l.currentCmd=t,l.error=null,l.errorCode=0;try{var o=[].slice.call(arguments,0);if(c.verbose&&console.error.apply(console,[t].concat(o)),l.pipedValue=this&&"string"==typeof this.stdout?this.stdout:"",!1===n.unix)i=r.apply(this,o);else{f(o[0])&&"Object"===o[0].constructor.name||(0===o.length||"string"!=typeof o[0]||o[0].length<=1||"-"!==o[0][0])&&o.unshift(""),o=(o=o.reduce((function(e,t){return Array.isArray(t)?e.concat(t):(e.push(t),e)}),[])).map((function(e){return f(e)&&"String"===e.constructor.name?e.toString():e}));var s=e.homedir();o=o.map((function(e){return"string"==typeof e&&"~/"===e.slice(0,2)||"~"===e?e.replace(/^~/,s):e})),c.noglob||!0!==n.allowGlobbing||(o=o.slice(0,n.globStart).concat(y(o.slice(n.globStart))));try{f(n.cmdOptions)&&(o[0]=g(o[0],n.cmdOptions)),i=r.apply(this,o)}catch(e){if("earlyExit"!==e.msg)throw e;i=e.retValue}}}catch(e){if(!l.error)throw e.name="ShellJSInternalError",e;if(c.fatal)throw e}return n.wrapOutput&&("string"==typeof i||Array.isArray(i))&&(i=new v(i,l.error,l.errorCode)),l.currentCmd="shell.js",i}}dh.buffer=m,dh.unlinkSync=function(e){try{t.unlinkSync(e)}catch(r){if("EPERM"!==r.code)throw r;t.chmodSync(e,"0666"),t.unlinkSync(e)}},dh.statFollowLinks=function(){return t.statSync.apply(t,arguments)},dh.statNoFollowLinks=function(){return t.lstatSync.apply(t,arguments)},dh.randomFileName=function(){return"shelljs_"+function e(t){if(1===t)return parseInt(16*Math.random(),10).toString(16);for(var r="",n=0;n<t;n++)r+=e(1);return r}(20)},dh.wrap=E,dh.readFromPipe=function(){return l.pipedValue};var b={allowGlobbing:!0,canReceivePipe:!1,cmdOptions:null,globStart:1,pipeOnly:!1,wrapOutput:!0,unix:!0},w=[];return dh.register=function(e,t,r){if(r=r||{},Object.keys(r).forEach((function(e){if(!b.hasOwnProperty(e))throw new Error("Unknown option '"+e+"'");if(typeof r[e]!=typeof b[e])throw new TypeError("Unsupported type '"+typeof r[e]+"' for option '"+e+"'")})),r=Object.assign({},b,r),n.hasOwnProperty(e))throw new Error("Command `"+e+"` already exists");r.pipeOnly?(r.canReceivePipe=!0,i[e]=E(e,t,r)):n[e]=E(e,t,r),r.canReceivePipe&&w.push(e)},dh}function Gv(){if(jv)return Av;jv=1;var e=Mv(),t=u.default;function r(r,i){var o=e.readFromPipe();return i||o||e.error("no paths given"),(i=[].slice.call(arguments,1)).forEach((function(r){t.existsSync(r)?e.statFollowLinks(r).isDirectory()&&e.error(r+": Is a directory"):e.error("no such file or directory: "+r),o+=t.readFileSync(r,"utf8")})),r.number&&(o=n(o)),o}function n(e){var t=e.split("\n"),r=t.pop();return t=t.map((function(e,t){return i(t+1,e)})),r.length&&(r=i(t.length+1,r)),t.push(r),t.join("\n")}function i(e,t){return("     "+e).slice(-6)+"\t"+t}return e.register("cat",r,{canReceivePipe:!0,cmdOptions:{n:"number"}}),Av=r}function $v(){if(kv)return xv;kv=1;var e=a.default,t=Mv();function r(r,n){n||(n=e.homedir()),"-"===n&&(process.env.OLDPWD?n=process.env.OLDPWD:t.error("could not find previous directory"));try{var i=process.cwd();process.chdir(n),process.env.OLDPWD=i}catch(e){var o;try{t.statFollowLinks(n),o="not a directory: "+n}catch(e){o="no such file or directory: "+n}o&&t.error(o)}return""}return t.register("cd",r,{}),xv=r}function Uv(){if(_v)return Lv;_v=1;var e,t=Mv(),r=u.default,n=l.default,i={OTHER_EXEC:(e={EXEC:1,WRITE:2,READ:4}).EXEC,OTHER_WRITE:e.WRITE,OTHER_READ:e.READ,GROUP_EXEC:e.EXEC<<3,GROUP_WRITE:e.WRITE<<3,GROUP_READ:e.READ<<3,OWNER_EXEC:e.EXEC<<6,OWNER_WRITE:e.WRITE<<6,OWNER_READ:e.READ<<6,STICKY:parseInt("01000",8),SETGID:parseInt("02000",8),SETUID:parseInt("04000",8),TYPE_MASK:parseInt("0770000",8)};function o(e,o,s){var c;return s||(e.length>0&&"-"===e.charAt(0)?[].unshift.call(arguments,""):t.error("You must specify a file.")),e=t.parseOptions(e,{R:"recursive",c:"changes",v:"verbose"}),s=[].slice.call(arguments,2),e.recursive?(c=[],s.forEach((function e(n){var i=t.statNoFollowLinks(n);i.isSymbolicLink()||(c.push(n),i.isDirectory()&&r.readdirSync(n).forEach((function(t){e(n+"/"+t)})))}))):c=s,c.forEach((function(s){if(s=n.resolve(s),r.existsSync(s)||t.error("File not found: "+s),!e.recursive||!t.statNoFollowLinks(s).isSymbolicLink()){var c=t.statFollowLinks(s),a=c.isDirectory(),u=c.mode,l=u&i.TYPE_MASK,f=u;isNaN(parseInt(o,8))?o.split(",").forEach((function(n){var o=/([ugoa]*)([=\+-])([rwxXst]*)/i.exec(n);if(o){var c=o[1],p=o[2],h=o[3],d=-1!==c.indexOf("u")||"a"===c||""===c,v=-1!==c.indexOf("g")||"a"===c||""===c,g=-1!==c.indexOf("o")||"a"===c||""===c,y=-1!==h.indexOf("r"),m=-1!==h.indexOf("w"),E=-1!==h.indexOf("x"),b=-1!==h.indexOf("X"),w=-1!==h.indexOf("t"),O=-1!==h.indexOf("s");b&&a&&(E=!0);var S=0;switch(d&&(S|=(y?i.OWNER_READ:0)+(m?i.OWNER_WRITE:0)+(E?i.OWNER_EXEC:0)+(O?i.SETUID:0)),v&&(S|=(y?i.GROUP_READ:0)+(m?i.GROUP_WRITE:0)+(E?i.GROUP_EXEC:0)+(O?i.SETGID:0)),g&&(S|=(y?i.OTHER_READ:0)+(m?i.OTHER_WRITE:0)+(E?i.OTHER_EXEC:0)),w&&(S|=i.STICKY),p){case"+":f|=S;break;case"-":f&=~S;break;case"=":f=l+S,t.statFollowLinks(s).isDirectory()&&(f|=i.SETUID+i.SETGID&u);break;default:t.error("Could not recognize operator: `"+p+"`")}e.verbose&&console.log(s+" -> "+f.toString(8)),u!==f&&(!e.verbose&&e.changes&&console.log(s+" -> "+f.toString(8)),r.chmodSync(s,f),u=f)}else t.error("Invalid symbolic mode change: "+n)})):(f=l+parseInt(o,8),t.statFollowLinks(s).isDirectory()&&(f|=i.SETUID+i.SETGID&u),r.chmodSync(s,f))}})),""}return t.register("chmod",o,{}),Lv=o}function Bv(){if(Nv)return Tv;Nv=1;var e=u.default,t=l.default,r=Mv();function n(t,n,i){e.existsSync(t)||r.error("copyFileSync: no such file or directory: "+t);var o="win32"===process.platform;try{if(i.update&&r.statFollowLinks(t).mtime<e.statSync(n).mtime)return}catch(e){}if(r.statNoFollowLinks(t).isSymbolicLink()&&!i.followsymlink){try{r.statNoFollowLinks(n),r.unlinkSync(n)}catch(e){}var s=e.readlinkSync(t);e.symlinkSync(s,n,o?"junction":null)}else{var c=r.buffer(),a=c.length,u=a,l=0,f=null,p=null;try{f=e.openSync(t,"r")}catch(e){r.error("copyFileSync: could not read src file ("+t+")")}try{p=e.openSync(n,"w")}catch(e){r.error("copyFileSync: could not write to dest file (code="+e.code+"):"+n)}for(;u===a;)u=e.readSync(f,c,0,a,l),e.writeSync(p,c,0,u),l+=u;e.closeSync(f),e.closeSync(p),e.chmodSync(n,r.statFollowLinks(t).mode)}}function i(t,o,c,a){if(a||(a={}),!(c>=r.config.maxdepth)){c++;var u="win32"===process.platform;try{e.mkdirSync(o)}catch(e){if("EEXIST"!==e.code)throw e}for(var l=e.readdirSync(t),f=0;f<l.length;f++){var p,h=t+"/"+l[f],d=o+"/"+l[f],v=r.statNoFollowLinks(h);if(a.followsymlink&&s(t,h))console.error("Cycle link found."),p=e.readlinkSync(h),e.symlinkSync(p,d,u?"junction":null);else if(v.isDirectory())i(h,d,c,a);else if(v.isSymbolicLink()&&!a.followsymlink){p=e.readlinkSync(h);try{r.statNoFollowLinks(d),r.unlinkSync(d)}catch(e){}e.symlinkSync(p,d,u?"junction":null)}else v.isSymbolicLink()&&a.followsymlink?(v=r.statFollowLinks(h)).isDirectory()?i(h,d,c,a):n(h,d,a):e.existsSync(d)&&a.no_force?r.log("skipping existing file: "+l[f]):n(h,d,a)}var g=r.statFollowLinks(t);e.chmodSync(o,g.mode)}}function o(e,r){var n=e[r];return e.slice(0,r).some((function(e){return t.basename(e)===t.basename(n)}))}function s(t,n){if(r.statNoFollowLinks(n).isSymbolicLink()&&r.statFollowLinks(n).isDirectory()){var i=e.realpathSync(t),o=e.realpathSync(n);if(new RegExp(o).test(i))return!0}return!1}function c(s,c,a){s.followsymlink&&(s.noFollowsymlink=!1),s.recursive||s.noFollowsymlink||(s.followsymlink=!0),arguments.length<3?r.error("missing <source> and/or <dest>"):(c=[].slice.call(arguments,1,arguments.length-1),a=arguments[arguments.length-1]);var u=e.existsSync(a),l=u&&r.statFollowLinks(a);return u&&l.isDirectory()||!(c.length>1)||r.error("dest is not a directory (too many sources)"),u&&l.isFile()&&s.no_force?new r.ShellString("","",0):(c.forEach((function(u,f){if(!e.existsSync(u))return""===u&&(u="''"),void r.error("no such file or directory: "+u,{continue:!0});var p=r.statFollowLinks(u);if(!s.noFollowsymlink&&p.isDirectory())if(s.recursive){var h=l&&l.isDirectory()?t.join(a,t.basename(u)):a;try{r.statFollowLinks(t.dirname(a)),i(u,h,0,{no_force:s.no_force,followsymlink:s.followsymlink})}catch(e){r.error("cannot create directory '"+a+"': No such file or directory")}}else r.error("omitting directory '"+u+"'",{continue:!0});else{var d=a;l&&l.isDirectory()&&(d=t.normalize(a+"/"+t.basename(u)));var v=e.existsSync(d);if(v&&o(c,f))return void(s.no_force||r.error("will not overwrite just-created '"+d+"' with '"+u+"'",{continue:!0}));if(v&&s.no_force)return;if(""===t.relative(u,d))return void r.error("'"+d+"' and '"+u+"' are the same file",{continue:!0});n(u,d,s)}})),new r.ShellString("",r.state.error,r.state.errorCode))}return r.register("cp",c,{cmdOptions:{f:"!no_force",n:"no_force",u:"update",R:"recursive",r:"recursive",L:"followsymlink",P:"noFollowsymlink"},wrapOutput:!1}),Tv=c}var Xv,qv,Wv,Vv,Hv,zv={};function Yv(){if(Xv)return zv;Xv=1;var e=Mv(),t=$v(),r=l.default;e.register("dirs",u,{wrapOutput:!1}),e.register("pushd",c,{wrapOutput:!1}),e.register("popd",a,{wrapOutput:!1});var n=[];function i(e){return/^[\-+]\d+$/.test(e)}function o(t){if(i(t)){if(Math.abs(t)<n.length+1)return/^-/.test(t)?Number(t)-1:Number(t);e.error(t+": directory stack index out of range")}else e.error(t+": invalid number")}function s(){return[process.cwd()].concat(n)}function c(c,a){i(c)&&(a=c,c=""),c=e.parseOptions(c,{n:"no-cd",q:"quiet"});var l=s();if("+0"===a)return l;if(a)if(i(a)){var f=o(a);l=l.slice(f).concat(l.slice(0,f))}else c["no-cd"]?l.splice(1,0,a):l.unshift(a);else{if(!(l.length>1))return e.error("no other directory");l=l.splice(1,1).concat(l)}return c["no-cd"]?l=l.slice(1):(a=r.resolve(l.shift()),t("",a)),n=l,u(c.quiet?"-q":"")}function a(s,c){if(i(s)&&(c=s,s=""),s=e.parseOptions(s,{n:"no-cd",q:"quiet"}),!n.length)return e.error("directory stack empty");if(c=o(c||"+0"),s["no-cd"]||c>0||n.length+c===0)c=c>0?c-1:c,n.splice(c,1);else{var a=r.resolve(n.shift());t("",a)}return u(s.quiet?"-q":"")}function u(t,r){if(i(t)&&(r=t,t=""),(t=e.parseOptions(t,{c:"clear",q:"quiet"})).clear)return n=[];var c=s();return r?((r=o(r))<0&&(r=c.length+r),t.quiet||e.log(c[r]),c[r]):(t.quiet||e.log(c.join(" ")),c)}return zv.pushd=c,zv.popd=a,zv.dirs=u,zv}function Kv(){if(Wv)return qv;Wv=1;var e=h.default.format,t=Mv();function r(r){var n=[].slice.call(arguments,r?0:1),i={};try{i=t.parseOptions(n[0],{e:"escapes",n:"no_newline"},{silent:!0}),n[0]&&n.shift()}catch(e){t.state.error=null}var o=e.apply(null,n);return i.no_newline||(o+="\n"),process.stdout.write(o),o}return t.register("echo",r,{allowGlobbing:!1}),qv=r}function Jv(){if(Hv)return Vv;Hv=1;var e=Mv();return Vv=function(){return e.state.error}}var Qv,Zv={exports:{}};function eg(){return Qv||(Qv=1,function(e){if(require.main!==e)throw new Error("This file should not be required");var t=d.default,r=u.default,n=process.argv[2],i=r.readFileSync(n,"utf8"),o=JSON.parse(i),s=o.command,c=o.execOptions,a=o.pipe,l=o.stdoutFile,f=o.stderrFile,p=t.exec(s,c,(function(e){e?void 0===e.code?process.exitCode=1:process.exitCode=e.code:process.exitCode=0})),h=r.createWriteStream(l),v=r.createWriteStream(f);p.stdout.pipe(h),p.stderr.pipe(v),p.stdout.pipe(process.stdout),p.stderr.pipe(process.stderr),a&&p.stdin.end(a)}(Zv)),Zv.exports}var tg,rg,ng,ig,og,sg,cg,ag,ug,lg,fg,pg,hg,dg,vg,gg,yg,mg,Eg,bg,wg,Og={};function Sg(){if(tg)return Og;tg=1;var e,t=Mv(),r=a.default,n=u.default;function i(e){if(!e||!n.existsSync(e))return!1;if(!t.statFollowLinks(e).isDirectory())return!1;var r=e+"/"+t.randomFileName();try{return n.writeFileSync(r," "),t.unlinkSync(r),e}catch(e){return!1}}function o(){return e||(e=i(r.tmpdir())||i(process.env.TMPDIR)||i(process.env.TEMP)||i(process.env.TMP)||i(process.env.Wimp$ScrapDir)||i("C:\\TEMP")||i("C:\\TMP")||i("\\TEMP")||i("\\TMP")||i("/tmp")||i("/var/tmp")||i("/usr/tmp")||i("."))}return t.register("tempdir",o,{allowGlobbing:!1,wrapOutput:!1}),Og.tempDir=o,Og.isCached=function(){return e},Og.clearCache=function(){e=void 0},Og}function Rg(){if(ng)return rg;ng=1;var e=l.default;function t(){return e.resolve(process.cwd())}return Mv().register("pwd",t,{allowGlobbing:!1}),rg=t}function Ig(){if(og)return ig;og=1;var e=Mv(),t=Sg().tempDir,r=Rg(),n=l.default,i=u.default,o=d.default,s=20971520;function c(c,a,u){a=a||{},c||e.error("must specify command");var l=e.readFromPipe();return"function"==typeof a&&(u=a,a={async:!0}),"object"==typeof a&&"function"==typeof u&&(a.async=!0),(a=e.extend({silent:e.config.silent,async:!1},a)).async?function(t,n,i,c){n=e.extend({silent:e.config.silent,cwd:r().toString(),env:process.env,maxBuffer:s,encoding:"utf8"},n);var a=o.exec(t,n,(function(e,t,r){c&&(e?void 0===e.code?c(1,t,r):c(e.code,t,r):c(0,t,r))}));return i&&a.stdin.end(i),n.silent||(a.stdout.pipe(process.stdout),a.stderr.pipe(process.stderr)),a}(c,a,l,u):function(c,a,u){e.config.execPath||e.error("Unable to find a path to the node binary. Please manually set config.execPath");var l=t(),f=n.resolve(l+"/"+e.randomFileName()),p=n.resolve(l+"/"+e.randomFileName()),h=n.resolve(l+"/"+e.randomFileName());a=e.extend({silent:e.config.silent,cwd:r().toString(),env:process.env,maxBuffer:s,encoding:"utf8"},a),i.existsSync(f)&&e.unlinkSync(f),i.existsSync(p)&&e.unlinkSync(p),i.existsSync(h)&&e.unlinkSync(h),a.cwd=n.resolve(a.cwd);var d={command:c,execOptions:a,pipe:u,stdoutFile:h,stderrFile:p};function v(e,t){i.writeFileSync(e,t,{encoding:"utf8",mode:parseInt("600",8)})}v(h,""),v(p,""),v(f,JSON.stringify(d));var g=[n.join(__dirname,"exec-child.js"),f];a.silent?a.stdio="ignore":a.stdio=[0,1,2];var y=0;try{delete a.shell,o.execFileSync(e.config.execPath,g,a)}catch(e){y=e.status||1}var m="",E="";"buffer"===a.encoding?(m=i.readFileSync(h),E=i.readFileSync(p)):(m=i.readFileSync(h,a.encoding),E=i.readFileSync(p,a.encoding));try{e.unlinkSync(f)}catch(e){}try{e.unlinkSync(p)}catch(e){}try{e.unlinkSync(h)}catch(e){}return 0!==y&&e.error(E,y,{continue:!0,silent:!0}),e.ShellString(m,E,y)}(c,a,l)}return e.register("exec",c,{unix:!1,canReceivePipe:!0,wrapOutput:!1}),ig=c}function Ag(){if(cg)return sg;cg=1;var e=l.default,t=u.default,r=Mv(),n=Cv(),i=e.sep+"**";function o(o,c){o.all_deprecated&&(r.log("ls: Option -a is deprecated. Use -A instead"),o.all=!0),c=c?[].slice.call(arguments,1):["."];var a=[];function u(e,t,n){"win32"===process.platform&&(t=t.replace(/\\/g,"/")),o.long?(n=n||(o.link?r.statFollowLinks(e):r.statNoFollowLinks(e)),a.push(s(t,n))):a.push(t)}return c.forEach((function(s){var c;try{if((c=o.link?r.statFollowLinks(s):r.statNoFollowLinks(s)).isSymbolicLink())try{var a=r.statFollowLinks(s);a.isDirectory()&&(c=a)}catch(e){}}catch(e){return void r.error("no such file or directory: "+s,2,{continue:!0})}c.isDirectory()&&!o.directory?o.recursive?n.sync(s+i,{dot:o.all,follow:o.link}).forEach((function(t){e.relative(s,t)&&u(t,e.relative(s,t))})):o.all?t.readdirSync(s).forEach((function(t){u(e.join(s,t),t)})):t.readdirSync(s).forEach((function(t){"."!==t[0]&&u(e.join(s,t),t)})):u(s,s,c)})),a}function s(e,t){return t.name=e,t.toString=function(){return[this.mode,this.nlink,this.uid,this.gid,this.size,this.mtime,this.name].join(" ")},t}return r.register("ls",o,{cmdOptions:{R:"recursive",A:"all",L:"link",a:"all_deprecated",d:"directory",l:"long"}}),sg=o}function jg(){if(ug)return ag;ug=1;var e=l.default,t=Mv(),r=Ag();function n(n,i){i?"string"==typeof i&&(i=[].slice.call(arguments,1)):t.error("no path specified");var o=[];function s(e){"win32"===process.platform&&(e=e.replace(/\\/g,"/")),o.push(e)}return i.forEach((function(n){var i;try{i=t.statFollowLinks(n)}catch(e){t.error("no such file or directory: "+n)}s(n),i.isDirectory()&&r({recursive:!0,all:!0},n).forEach((function(t){s(e.join(n,t))}))})),o}return t.register("find",n,{}),ag=n}function xg(){if(fg)return lg;fg=1;var e=Mv(),t=u.default;function r(r,n,i){var o=e.readFromPipe();i||o||e.error("no paths given",2),i=[].slice.call(arguments,2),o&&i.unshift("-");var s=[];return r.ignoreCase&&(n=new RegExp(n,"i")),i.forEach((function(i){if(t.existsSync(i)||"-"===i){var c="-"===i?o:t.readFileSync(i,"utf8");if(r.nameOnly)c.match(n)&&s.push(i);else c.split("\n").forEach((function(e){var t=e.match(n);(r.inverse&&!t||!r.inverse&&t)&&s.push(e)}))}else e.error("no such file or directory: "+i,2,{continue:!0})})),s.join("\n")+"\n"}return e.register("grep",r,{globStart:2,canReceivePipe:!0,cmdOptions:{v:"inverse",l:"nameOnly",i:"ignoreCase"}}),lg=r}function kg(){if(hg)return pg;hg=1;var e=Mv(),t=u.default;function r(r,n){for(var i=e.buffer(),o=i.length,s=o,c=0,a=t.openSync(r,"r"),u=0,l="";s===o&&u<n;){s=t.readSync(a,i,0,o,c);var f=i.toString("utf8",0,s);u+=f.split("\n").length-1,l+=f,c+=s}return t.closeSync(a),l}function n(n,i){var o=[],s=e.readFromPipe();i||s||e.error("no paths given");var c=1;!0===n.numLines?(c=2,n.numLines=Number(arguments[1])):!1===n.numLines&&(n.numLines=10),i=[].slice.call(arguments,c),s&&i.unshift("-");var a=!1;return i.forEach((function(i){if("-"!==i){if(!t.existsSync(i))return void e.error("no such file or directory: "+i,{continue:!0});if(e.statFollowLinks(i).isDirectory())return void e.error("error reading '"+i+"': Is a directory",{continue:!0})}var c=("-"===i?s:n.numLines<0?t.readFileSync(i,"utf8"):r(i,n.numLines)).split("\n"),u=""===c[c.length-1];u&&c.pop(),a=u||n.numLines<c.length,o=o.concat(c.slice(0,n.numLines))})),a&&o.push(""),o.join("\n")}return e.register("head",n,{canReceivePipe:!0,cmdOptions:{n:"numLines"}}),pg=n}function Lg(){if(vg)return dg;vg=1;var e=u.default,t=l.default,r=Mv();function n(n,i,o){i&&o||r.error("Missing <source> and/or <dest>"),i=String(i);var s=t.normalize(i).replace(RegExp(t.sep+"$"),""),c=t.resolve(i)===s;if(o=t.resolve(process.cwd(),String(o)),e.existsSync(o)&&(n.force||r.error("Destination file exists",{continue:!0}),e.unlinkSync(o)),n.symlink){var a="win32"===process.platform,u=a?"file":null,l=c?s:t.resolve(process.cwd(),t.dirname(o),i);e.existsSync(l)?a&&r.statFollowLinks(l).isDirectory()&&(u="junction"):r.error("Source file does not exist",{continue:!0});try{e.symlinkSync("junction"===u?l:i,o,u)}catch(e){r.error(e.message)}}else{e.existsSync(i)||r.error("Source file does not exist",{continue:!0});try{e.linkSync(i,o)}catch(e){r.error(e.message)}}return""}return r.register("ln",n,{cmdOptions:{s:"symlink",f:"force"}}),dg=n}function _g(){if(yg)return gg;yg=1;var e=Mv(),t=u.default,r=l.default;function n(i){var o=r.dirname(i);o===i&&e.error("dirname() failed: ["+i+"]"),t.existsSync(o)||n(o),t.mkdirSync(i,parseInt("0777",8))}function i(i,o){return o||e.error("no paths given"),"string"==typeof o&&(o=[].slice.call(arguments,1)),o.forEach((function(o){try{var s=e.statNoFollowLinks(o);return void(i.fullpath?s.isFile()&&e.error("cannot create directory "+o+": File exists",{continue:!0}):e.error("path already exists: "+o,{continue:!0}))}catch(e){}var c=r.dirname(o);if(t.existsSync(c)||i.fullpath)try{i.fullpath?n(r.resolve(o)):t.mkdirSync(o,parseInt("0777",8))}catch(t){var a;if("EACCES"===t.code)a="Permission denied";else{if("ENOTDIR"!==t.code&&"ENOENT"!==t.code)throw t;a="Not a directory"}e.error("cannot create directory "+o+": "+a,{continue:!0})}else e.error("no such file or directory: "+c,{continue:!0})})),""}return e.register("mkdir",i,{cmdOptions:{p:"fullpath"}}),gg=i}function Tg(){if(Eg)return mg;Eg=1;var e=Mv(),t=u.default;function r(i,o,s){var c;c=t.readdirSync(i);for(var a=0;a<c.length;a++){var u=i+"/"+c[a];if(e.statNoFollowLinks(u).isDirectory())r(u,o);else if(o||n(u))try{e.unlinkSync(u)}catch(t){e.error("could not remove file (code "+t.code+"): "+u,{continue:!0})}}if(!s){var l;try{for(var f=Date.now();;)try{if(l=t.rmdirSync(i),t.existsSync(i))throw{code:"EAGAIN"};break}catch(e){if("win32"!==process.platform||"ENOTEMPTY"!==e.code&&"EBUSY"!==e.code&&"EPERM"!==e.code&&"EAGAIN"!==e.code){if("ENOENT"===e.code)break;throw e}if(Date.now()-f>1e3)throw e}}catch(t){e.error("could not remove directory (code "+t.code+"): "+i,{continue:!0})}return l}}function n(e){var r=!0;try{var n=t.openSync(e,"a");t.closeSync(n)}catch(e){r=!1}return r}function i(t,r){r.force||n(t)?e.unlinkSync(t):e.error("permission denied: "+t,{continue:!0})}function o(t,n){n.recursive?r(t,n.force):e.error("path is a directory",{continue:!0})}function s(t,n){var i;try{i=e.statFollowLinks(t)}catch(r){return void e.unlinkSync(t)}if(i.isFile())e.unlinkSync(t);else if(i.isDirectory())if("/"===t[t.length-1])if(n.recursive){r(t,n.force,!0)}else e.error("path is a directory",{continue:!0});else e.unlinkSync(t)}function c(t){e.unlinkSync(t)}function a(t,r){return r||e.error("no paths given"),(r=[].slice.call(arguments,1)).forEach((function(r){var n;try{var a="/"===r[r.length-1]?r.slice(0,-1):r;n=e.statNoFollowLinks(a)}catch(n){return void(t.force||e.error("no such file or directory: "+r,{continue:!0}))}n.isFile()?i(r,t):n.isDirectory()?o(r,t):n.isSymbolicLink()?s(r,t):n.isFIFO()&&c(r)})),""}return e.register("rm",a,{cmdOptions:{f:"force",r:"recursive",R:"recursive"}}),mg=a}function Ng(){if(wg)return bg;wg=1;var e=u.default,t=l.default,r=Mv(),n=Bv(),i=Tg();function o(e,r){var n=e[r];return e.slice(0,r).some((function(e){return t.basename(e)===t.basename(n)}))}function s(s,c,a){arguments.length<3?r.error("missing <source> and/or <dest>"):arguments.length>3?(c=[].slice.call(arguments,1,arguments.length-1),a=arguments[arguments.length-1]):"string"==typeof c?c=[c]:r.error("invalid arguments");var u=e.existsSync(a),l=u&&r.statFollowLinks(a);return u&&l.isDirectory()||!(c.length>1)||r.error("dest is not a directory (too many sources)"),u&&l.isFile()&&s.no_force&&r.error("dest file already exists: "+a),c.forEach((function(u,l){if(e.existsSync(u)){var f=a;if(e.existsSync(a)&&r.statFollowLinks(a).isDirectory()&&(f=t.normalize(a+"/"+t.basename(u))),e.existsSync(f)&&o(c,l))s.no_force||r.error("will not overwrite just-created '"+f+"' with '"+u+"'",{continue:!0});else if(e.existsSync(f)&&s.no_force)r.error("dest file already exists: "+f,{continue:!0});else if(t.resolve(u)!==t.dirname(t.resolve(f)))try{e.renameSync(u,f)}catch(e){"EXDEV"===e.code&&(n("-r",u,f),i("-rf",u))}else r.error("cannot move to self: "+u,{continue:!0})}else r.error("no such file or directory: "+u,{continue:!0})})),""}return r.register("mv",s,{cmdOptions:{f:"!no_force",n:"no_force"}}),bg=s}var Pg,Fg={};function Dg(){return Pg||(Pg=1),Fg}var Cg,Mg,Gg,$g,Ug,Bg={};function Xg(){return Cg||(Cg=1),Bg}function qg(){if(Gg)return Mg;Gg=1;var e=Mv(),t=u.default;function r(r,n,i,o){var s=e.readFromPipe();"string"!=typeof i&&"function"!=typeof i&&("number"==typeof i?i=i.toString():e.error("invalid replacement string")),"string"==typeof n&&(n=RegExp(n)),o||s||e.error("no files given"),o=[].slice.call(arguments,3),s&&o.unshift("-");var c=[];return o.forEach((function(o){if(t.existsSync(o)||"-"===o){var a=("-"===o?s:t.readFileSync(o,"utf8")).split("\n").map((function(e){return e.replace(n,i)})).join("\n");c.push(a),r.inplace&&t.writeFileSync(o,a,"utf8")}else e.error("no such file or directory: "+o,2,{continue:!0})})),c.join("\n")}return e.register("sed",r,{globStart:3,canReceivePipe:!0,cmdOptions:{i:"inplace"}}),Mg=r}function Wg(){if(Ug)return $g;Ug=1;var e=Mv();function t(t){if(!t){var r=[].slice.call(arguments,0);r.length<2&&e.error("must provide an argument"),t=r[1]}var n="+"===t[0];n&&(t="-"+t.slice(1)),t=e.parseOptions(t,{e:"fatal",v:"verbose",f:"noglob"}),n&&Object.keys(t).forEach((function(e){t[e]=!t[e]})),Object.keys(t).forEach((function(r){n!==t[r]&&(e.config[r]=t[r])}))}return e.register("set",t,{allowGlobbing:!1,wrapOutput:!1}),$g=t}var Vg=wf.end,Hg=If("trimEnd")?function(){return Vg(this)}:"".trimEnd;ui({target:"String",proto:!0,name:"trimEnd",forced:"".trimRight!==Hg},{trimRight:Hg});var zg,Yg,Kg,Jg,Qg,Zg,ey,ty,ry,ny,iy,oy,sy,cy,ay,uy,ly;function fy(){if(Yg)return zg;Yg=1;var e=Mv(),t=u.default;function r(e){var t=e.match(/^\s*(\d*)\s*(.*)$/);return{num:Number(t[1]),value:t[2]}}function n(e,t){var r=e.toLowerCase(),n=t.toLowerCase();return r===n?-1*e.localeCompare(t):r.localeCompare(n)}function i(e,t){var i=r(e),o=r(t);return i.hasOwnProperty("num")&&o.hasOwnProperty("num")&&i.num!==o.num?i.num-o.num:n(i.value,o.value)}function o(r,o){var s=e.readFromPipe();o||s||e.error("no files given"),o=[].slice.call(arguments,1),s&&o.unshift("-");var c=o.reduce((function(r,n){if("-"!==n){if(!t.existsSync(n))return e.error("no such file or directory: "+n,{continue:!0}),r;if(e.statFollowLinks(n).isDirectory())return e.error("read failed: "+n+": Is a directory",{continue:!0}),r}var i="-"===n?s:t.readFileSync(n,"utf8");return r.concat(i.trimRight().split("\n"))}),[]),a=c.sort(r.numerical?i:n);return r.reverse&&(a=a.reverse()),a.join("\n")+"\n"}return e.register("sort",o,{canReceivePipe:!0,cmdOptions:{r:"reverse",n:"numerical"}}),zg=o}function py(){if(Jg)return Kg;Jg=1;var e=Mv(),t=u.default;function r(r,n){var i=[],o=e.readFromPipe();n||o||e.error("no paths given");var s=1;!0===r.numLines?(s=2,r.numLines=Number(arguments[1])):!1===r.numLines&&(r.numLines=10),r.numLines=-1*Math.abs(r.numLines),n=[].slice.call(arguments,s),o&&n.unshift("-");var c=!1;return n.forEach((function(n){if("-"!==n){if(!t.existsSync(n))return void e.error("no such file or directory: "+n,{continue:!0});if(e.statFollowLinks(n).isDirectory())return void e.error("error reading '"+n+"': Is a directory",{continue:!0})}var s=("-"===n?o:t.readFileSync(n,"utf8")).split("\n");""===s[s.length-1]?(s.pop(),c=!0):c=!1,i=i.concat(s.slice(r.numLines))})),c&&i.push(""),i.join("\n")}return e.register("tail",r,{canReceivePipe:!0,cmdOptions:{n:"numLines"}}),Kg=r}function hy(){if(Zg)return Qg;Zg=1;var e=Mv(),t=u.default;function r(r,n){n||e.error("no path given");var i=!1;if(Object.keys(r).forEach((function(e){!0===r[e]&&(i=!0)})),i||e.error("could not interpret expression"),r.link)try{return e.statNoFollowLinks(n).isSymbolicLink()}catch(e){return!1}if(!t.existsSync(n))return!1;if(r.exists)return!0;var o=e.statFollowLinks(n);return r.block?o.isBlockDevice():r.character?o.isCharacterDevice():r.directory?o.isDirectory():r.file?o.isFile():r.pipe?o.isFIFO():!!r.socket&&o.isSocket()}return e.register("test",r,{cmdOptions:{b:"block",c:"character",d:"directory",e:"exists",f:"file",L:"link",p:"pipe",S:"socket"},wrapOutput:!1,allowGlobbing:!1}),Qg=r}function dy(){if(ty)return ey;ty=1;var e=Mv(),t=u.default,r=l.default;function n(n,i){i||e.error("wrong arguments"),t.existsSync(r.dirname(i))||e.error("no such file or directory: "+r.dirname(i));try{return t.writeFileSync(i,this.stdout||this.toString(),"utf8"),this}catch(t){e.error("could not write to file (code "+t.code+"): "+i,{continue:!0})}}return e.register("to",n,{pipeOnly:!0,wrapOutput:!1}),ey=n}function vy(){if(ny)return ry;ny=1;var e=Mv(),t=u.default,r=l.default;function n(n,i){i||e.error("wrong arguments"),t.existsSync(r.dirname(i))||e.error("no such file or directory: "+r.dirname(i));try{return t.appendFileSync(i,this.stdout||this.toString(),"utf8"),this}catch(t){e.error("could not append to file (code "+t.code+"): "+i,{continue:!0})}}return e.register("toEnd",n,{pipeOnly:!0,wrapOutput:!1}),ry=n}function gy(){if(oy)return iy;oy=1;var e=Mv(),t=u.default;function r(t,r){return r?"string"==typeof r?r=[].slice.call(arguments,1):e.error("file arg should be a string file path or an Array of string file paths"):e.error("no files given"),r.forEach((function(e){n(t,e)})),""}function n(r,n){var o=i(n);if((!o||!o.isDirectory())&&(o||!r.no_create)){t.closeSync(t.openSync(n,"a"));var s=new Date,c=r.date||s,a=r.date||s;if(r.reference){var u=i(r.reference);u||e.error("failed to get attributess of "+r.reference),c=u.mtime,a=u.atime}else r.date&&(c=r.date,a=r.date);r.atime_only&&r.mtime_only||(r.atime_only?c=o.mtime:r.mtime_only&&(a=o.atime)),t.utimesSync(n,a,c)}}function i(t){try{return e.statFollowLinks(t)}catch(e){return null}}return e.register("touch",r,{cmdOptions:{a:"atime_only",c:"no_create",d:"date",m:"mtime_only",r:"reference"}}),iy=r}function yy(){if(cy)return sy;cy=1;var e=Mv(),t=u.default;function r(r,n,i){var o=e.readFromPipe();o||(n||e.error("no input given"),t.existsSync(n)?e.statFollowLinks(n).isDirectory()&&e.error("error reading '"+n+"'"):e.error(n+": No such file or directory")),i&&t.existsSync(i)&&e.statFollowLinks(i).isDirectory()&&e.error(i+": Is a directory");var s=(n?t.readFileSync(n,"utf8"):o).trimRight().split("\n"),c=s.reduceRight((function(e,t){return 0===e.length?[{count:1,ln:t}]:0===(n=e[0].ln,i=t,r.ignoreCase?n.toLocaleLowerCase().localeCompare(i.toLocaleLowerCase()):n.localeCompare(i))?[{count:e[0].count+1,ln:t}].concat(e.slice(1)):[{count:1,ln:t}].concat(e);var n,i}),[]).filter((function(e){return!r.duplicates||e.count>1})).map((function(e){return(r.count?(t=7,n=e.count,(i=""+n).length<t&&(i=Array(t-i.length+1).join(" ")+i),i+" "):"")+e.ln;var t,n,i})).join("\n")+"\n";return i?(new e.ShellString(c).to(i),""):c}return e.register("uniq",r,{canReceivePipe:!0,cmdOptions:{i:"ignoreCase",c:"count",d:"duplicates"}}),sy=r}function my(){if(uy)return ay;uy=1;var e=Mv(),t=u.default,r=l.default;e.register("which",s,{allowGlobbing:!1,cmdOptions:{a:"all"}});function n(){return"win32"===process.platform}function i(e){return e?e.split(r.delimiter):[]}function o(r){return t.existsSync(r)&&!e.statFollowLinks(r).isDirectory()&&(n()||function(e){try{t.accessSync(e,1)}catch(e){return!1}return!0}(r))}function s(t,s){s||e.error("must specify command");var c=n(),a=i(process.env.PATH),u=[];if(-1===s.indexOf("/")){var l=[""];if(c)l=i((process.env.PATHEXT||".com;.exe;.bat;.cmd;.vbs;.vbe;.js;.jse;.wsf;.wsh").toUpperCase());for(var f=0;f<a.length&&(!(u.length>0)||t.all);f++){var p=r.resolve(a[f],s);c&&(p=p.toUpperCase());var h=p.match(/\.[^<>:"/\|?*.]+$/);if(h&&l.indexOf(h[0])>=0){if(o(p)){u.push(p);break}}else for(var d=0;d<l.length;d++){var v=p+l[d];if(o(v)){u.push(v);break}}}}else o(s)&&u.push(r.resolve(s));return u.length>0?t.all?u:u[0]:t.all?[]:null}return ay=s}function Ey(){return ly||(ly={"/node_modules/shelljs/src/cat.js":Gv,"/node_modules/shelljs/src/cd.js":$v,"/node_modules/shelljs/src/chmod.js":Uv,"/node_modules/shelljs/src/common.js":Mv,"/node_modules/shelljs/src/cp.js":Bv,"/node_modules/shelljs/src/dirs.js":Yv,"/node_modules/shelljs/src/echo.js":Kv,"/node_modules/shelljs/src/error.js":Jv,"/node_modules/shelljs/src/exec-child.js":eg,"/node_modules/shelljs/src/exec.js":Ig,"/node_modules/shelljs/src/find.js":jg,"/node_modules/shelljs/src/grep.js":xg,"/node_modules/shelljs/src/head.js":kg,"/node_modules/shelljs/src/ln.js":Lg,"/node_modules/shelljs/src/ls.js":Ag,"/node_modules/shelljs/src/mkdir.js":_g,"/node_modules/shelljs/src/mv.js":Ng,"/node_modules/shelljs/src/popd.js":Dg,"/node_modules/shelljs/src/pushd.js":Xg,"/node_modules/shelljs/src/pwd.js":Rg,"/node_modules/shelljs/src/rm.js":Tg,"/node_modules/shelljs/src/sed.js":qg,"/node_modules/shelljs/src/set.js":Wg,"/node_modules/shelljs/src/sort.js":fy,"/node_modules/shelljs/src/tail.js":py,"/node_modules/shelljs/src/tempdir.js":Sg,"/node_modules/shelljs/src/test.js":hy,"/node_modules/shelljs/src/to.js":dy,"/node_modules/shelljs/src/toEnd.js":vy,"/node_modules/shelljs/src/touch.js":gy,"/node_modules/shelljs/src/uniq.js":yy,"/node_modules/shelljs/src/which.js":my})}function by(e,t){var r,n=function(e){var t=e[0];if("/"===t||"\\"===t)return!1;var r=e[1],n=e[2];return!(!("."!==t||r&&"/"!==r&&"\\"!==r)||!("."!==t||"."!==r||n&&"/"!==n&&"\\"!==n))&&(":"!==r||"/"!==n&&"\\"!==n)}(e);"/"===(e=wy(e))[0]&&(t="");for(var i=Ey(),o=["",".js",".json"];!(r=wy(n?t+"/node_modules/"+e:t+"/"+e)).endsWith("/..");){for(var s=0;s<o.length;s++){var c=r+o[s];if(i[c])return c}if(!n)break;var a=wy(t+"/..");if(a===t)break;t=a}return null}function wy(e){for(var t=(e=e.replace(/\\/g,"/")).split("/"),r=""===t[0],n=1;n<t.length;n++)"."!==t[n]&&""!==t[n]||t.splice(n--,1);for(n=1;n<t.length;n++)".."===t[n]&&n>0&&".."!==t[n-1]&&"."!==t[n-1]&&(t.splice(--n,2),n--);return e=t.join("/"),r&&"/"!==e[0]?e="/"+e:0===e.length&&(e="."),e}ui({target:"String",proto:!0,name:"trimEnd",forced:"".trimEnd!==Hg},{trimEnd:Hg});var Oy,Sy={},Ry=["cat","cd","chmod","cp","dirs","echo","exec","find","grep","head","ln","ls","mkdir","mv","pwd","rm","sed","set","sort","tail","tempdir","test","to","toEnd","touch","uniq","which"];function Iy(){if(Oy)return Sy;Oy=1;var e=Mv();return Ry.forEach((function(e){!function(e){function t(t){var r=by(t,e);if(null!==r)return Ey()[r]();throw new Error('Could not dynamically require "'+t+'". Please configure the dynamicRequireTargets or/and ignoreDynamicRequires option of @rollup/plugin-commonjs appropriately for this require call to work.')}return t.resolve=function(t){var r=by(t,e);return null!==r?r:require.resolve(t)},t}("/node_modules/shelljs")("./src/"+e)})),Sy.exit=process.exit,Sy.error=Jv(),Sy.ShellString=e.ShellString,Sy.env=process.env,Sy.config=e.config,Sy}g(Ry);var Ay=Iy(),jy=Mv();Object.keys(Ay).forEach((function(e){v[e]=Ay[e]}));var xy=dy();String.prototype.to=jy.wrap("to",xy);var ky=vy();String.prototype.toEnd=jy.wrap("toEnd",ky),v.config.fatal=!0,v.target={};var Ly,_y=process.argv.slice(2),Ty=_y.indexOf("--");Ty>-1&&(Ly=_y.slice(Ty+1,_y.length),_y=_y.slice(0,Ty)),setTimeout((function(){var e;if(1!==_y.length||"--help"!==_y[0]){for(e in v.target)!function(e,t){v.target[e]=function(){return t.done||(t.done=!0,t.result=t.apply(t,arguments)),t.result}}(e,v.target[e]);_y.length>0?_y.forEach((function(e){e in v.target?v.target[e](Ly):console.log("no such target: "+e)})):"all"in v.target&&v.target.all(Ly)}else for(e in console.log("Available targets:"),v.target)console.log("  "+e)}),0);var Ny={exports:{}};!function(e,t){var r;t=Ny.exports=f,r="object"==typeof process&&process.env&&process.env.NODE_DEBUG&&/\bsemver\b/i.test(process.env.NODE_DEBUG)?function(){var e=Array.prototype.slice.call(arguments,0);e.unshift("SEMVER"),console.log.apply(console,e)}:function(){},t.SEMVER_SPEC_VERSION="2.0.0";var n=Number.MAX_SAFE_INTEGER||9007199254740991,i=t.re=[],o=t.src=[],s=t.tokens={},c=0;function a(e){s[e]=c++}a("NUMERICIDENTIFIER"),o[s.NUMERICIDENTIFIER]="0|[1-9]\\d*",a("NUMERICIDENTIFIERLOOSE"),o[s.NUMERICIDENTIFIERLOOSE]="[0-9]+",a("NONNUMERICIDENTIFIER"),o[s.NONNUMERICIDENTIFIER]="\\d*[a-zA-Z-][a-zA-Z0-9-]*",a("MAINVERSION"),o[s.MAINVERSION]="("+o[s.NUMERICIDENTIFIER]+")\\.("+o[s.NUMERICIDENTIFIER]+")\\.("+o[s.NUMERICIDENTIFIER]+")",a("MAINVERSIONLOOSE"),o[s.MAINVERSIONLOOSE]="("+o[s.NUMERICIDENTIFIERLOOSE]+")\\.("+o[s.NUMERICIDENTIFIERLOOSE]+")\\.("+o[s.NUMERICIDENTIFIERLOOSE]+")",a("PRERELEASEIDENTIFIER"),o[s.PRERELEASEIDENTIFIER]="(?:"+o[s.NUMERICIDENTIFIER]+"|"+o[s.NONNUMERICIDENTIFIER]+")",a("PRERELEASEIDENTIFIERLOOSE"),o[s.PRERELEASEIDENTIFIERLOOSE]="(?:"+o[s.NUMERICIDENTIFIERLOOSE]+"|"+o[s.NONNUMERICIDENTIFIER]+")",a("PRERELEASE"),o[s.PRERELEASE]="(?:-("+o[s.PRERELEASEIDENTIFIER]+"(?:\\."+o[s.PRERELEASEIDENTIFIER]+")*))",a("PRERELEASELOOSE"),o[s.PRERELEASELOOSE]="(?:-?("+o[s.PRERELEASEIDENTIFIERLOOSE]+"(?:\\."+o[s.PRERELEASEIDENTIFIERLOOSE]+")*))",a("BUILDIDENTIFIER"),o[s.BUILDIDENTIFIER]="[0-9A-Za-z-]+",a("BUILD"),o[s.BUILD]="(?:\\+("+o[s.BUILDIDENTIFIER]+"(?:\\."+o[s.BUILDIDENTIFIER]+")*))",a("FULL"),a("FULLPLAIN"),o[s.FULLPLAIN]="v?"+o[s.MAINVERSION]+o[s.PRERELEASE]+"?"+o[s.BUILD]+"?",o[s.FULL]="^"+o[s.FULLPLAIN]+"$",a("LOOSEPLAIN"),o[s.LOOSEPLAIN]="[v=\\s]*"+o[s.MAINVERSIONLOOSE]+o[s.PRERELEASELOOSE]+"?"+o[s.BUILD]+"?",a("LOOSE"),o[s.LOOSE]="^"+o[s.LOOSEPLAIN]+"$",a("GTLT"),o[s.GTLT]="((?:<|>)?=?)",a("XRANGEIDENTIFIERLOOSE"),o[s.XRANGEIDENTIFIERLOOSE]=o[s.NUMERICIDENTIFIERLOOSE]+"|x|X|\\*",a("XRANGEIDENTIFIER"),o[s.XRANGEIDENTIFIER]=o[s.NUMERICIDENTIFIER]+"|x|X|\\*",a("XRANGEPLAIN"),o[s.XRANGEPLAIN]="[v=\\s]*("+o[s.XRANGEIDENTIFIER]+")(?:\\.("+o[s.XRANGEIDENTIFIER]+")(?:\\.("+o[s.XRANGEIDENTIFIER]+")(?:"+o[s.PRERELEASE]+")?"+o[s.BUILD]+"?)?)?",a("XRANGEPLAINLOOSE"),o[s.XRANGEPLAINLOOSE]="[v=\\s]*("+o[s.XRANGEIDENTIFIERLOOSE]+")(?:\\.("+o[s.XRANGEIDENTIFIERLOOSE]+")(?:\\.("+o[s.XRANGEIDENTIFIERLOOSE]+")(?:"+o[s.PRERELEASELOOSE]+")?"+o[s.BUILD]+"?)?)?",a("XRANGE"),o[s.XRANGE]="^"+o[s.GTLT]+"\\s*"+o[s.XRANGEPLAIN]+"$",a("XRANGELOOSE"),o[s.XRANGELOOSE]="^"+o[s.GTLT]+"\\s*"+o[s.XRANGEPLAINLOOSE]+"$",a("COERCE"),o[s.COERCE]="(^|[^\\d])(\\d{1,16})(?:\\.(\\d{1,16}))?(?:\\.(\\d{1,16}))?(?:$|[^\\d])",a("COERCERTL"),i[s.COERCERTL]=new RegExp(o[s.COERCE],"g"),a("LONETILDE"),o[s.LONETILDE]="(?:~>?)",a("TILDETRIM"),o[s.TILDETRIM]="(\\s*)"+o[s.LONETILDE]+"\\s+",i[s.TILDETRIM]=new RegExp(o[s.TILDETRIM],"g");a("TILDE"),o[s.TILDE]="^"+o[s.LONETILDE]+o[s.XRANGEPLAIN]+"$",a("TILDELOOSE"),o[s.TILDELOOSE]="^"+o[s.LONETILDE]+o[s.XRANGEPLAINLOOSE]+"$",a("LONECARET"),o[s.LONECARET]="(?:\\^)",a("CARETTRIM"),o[s.CARETTRIM]="(\\s*)"+o[s.LONECARET]+"\\s+",i[s.CARETTRIM]=new RegExp(o[s.CARETTRIM],"g");a("CARET"),o[s.CARET]="^"+o[s.LONECARET]+o[s.XRANGEPLAIN]+"$",a("CARETLOOSE"),o[s.CARETLOOSE]="^"+o[s.LONECARET]+o[s.XRANGEPLAINLOOSE]+"$",a("COMPARATORLOOSE"),o[s.COMPARATORLOOSE]="^"+o[s.GTLT]+"\\s*("+o[s.LOOSEPLAIN]+")$|^$",a("COMPARATOR"),o[s.COMPARATOR]="^"+o[s.GTLT]+"\\s*("+o[s.FULLPLAIN]+")$|^$",a("COMPARATORTRIM"),o[s.COMPARATORTRIM]="(\\s*)"+o[s.GTLT]+"\\s*("+o[s.LOOSEPLAIN]+"|"+o[s.XRANGEPLAIN]+")",i[s.COMPARATORTRIM]=new RegExp(o[s.COMPARATORTRIM],"g");a("HYPHENRANGE"),o[s.HYPHENRANGE]="^\\s*("+o[s.XRANGEPLAIN]+")\\s+-\\s+("+o[s.XRANGEPLAIN]+")\\s*$",a("HYPHENRANGELOOSE"),o[s.HYPHENRANGELOOSE]="^\\s*("+o[s.XRANGEPLAINLOOSE]+")\\s+-\\s+("+o[s.XRANGEPLAINLOOSE]+")\\s*$",a("STAR"),o[s.STAR]="(<|>)?=?\\s*\\*";for(var u=0;u<c;u++)r(u,o[u]),i[u]||(i[u]=new RegExp(o[u]));function l(e,t){if(t&&"object"==typeof t||(t={loose:!!t,includePrerelease:!1}),e instanceof f)return e;if("string"!=typeof e)return null;if(e.length>256)return null;if(!(t.loose?i[s.LOOSE]:i[s.FULL]).test(e))return null;try{return new f(e,t)}catch(e){return null}}function f(e,t){if(t&&"object"==typeof t||(t={loose:!!t,includePrerelease:!1}),e instanceof f){if(e.loose===t.loose)return e;e=e.version}else if("string"!=typeof e)throw new TypeError("Invalid Version: "+e);if(e.length>256)throw new TypeError("version is longer than 256 characters");if(!(this instanceof f))return new f(e,t);r("SemVer",e,t),this.options=t,this.loose=!!t.loose;var o=e.trim().match(t.loose?i[s.LOOSE]:i[s.FULL]);if(!o)throw new TypeError("Invalid Version: "+e);if(this.raw=e,this.major=+o[1],this.minor=+o[2],this.patch=+o[3],this.major>n||this.major<0)throw new TypeError("Invalid major version");if(this.minor>n||this.minor<0)throw new TypeError("Invalid minor version");if(this.patch>n||this.patch<0)throw new TypeError("Invalid patch version");o[4]?this.prerelease=o[4].split(".").map((function(e){if(/^[0-9]+$/.test(e)){var t=+e;if(t>=0&&t<n)return t}return e})):this.prerelease=[],this.build=o[5]?o[5].split("."):[],this.format()}t.parse=l,t.valid=function(e,t){var r=l(e,t);return r?r.version:null},t.clean=function(e,t){var r=l(e.trim().replace(/^[=v]+/,""),t);return r?r.version:null},t.SemVer=f,f.prototype.format=function(){return this.version=this.major+"."+this.minor+"."+this.patch,this.prerelease.length&&(this.version+="-"+this.prerelease.join(".")),this.version},f.prototype.toString=function(){return this.version},f.prototype.compare=function(e){return r("SemVer.compare",this.version,this.options,e),e instanceof f||(e=new f(e,this.options)),this.compareMain(e)||this.comparePre(e)},f.prototype.compareMain=function(e){return e instanceof f||(e=new f(e,this.options)),h(this.major,e.major)||h(this.minor,e.minor)||h(this.patch,e.patch)},f.prototype.comparePre=function(e){if(e instanceof f||(e=new f(e,this.options)),this.prerelease.length&&!e.prerelease.length)return-1;if(!this.prerelease.length&&e.prerelease.length)return 1;if(!this.prerelease.length&&!e.prerelease.length)return 0;var t=0;do{var n=this.prerelease[t],i=e.prerelease[t];if(r("prerelease compare",t,n,i),void 0===n&&void 0===i)return 0;if(void 0===i)return 1;if(void 0===n)return-1;if(n!==i)return h(n,i)}while(++t)},f.prototype.compareBuild=function(e){e instanceof f||(e=new f(e,this.options));var t=0;do{var n=this.build[t],i=e.build[t];if(r("prerelease compare",t,n,i),void 0===n&&void 0===i)return 0;if(void 0===i)return 1;if(void 0===n)return-1;if(n!==i)return h(n,i)}while(++t)},f.prototype.inc=function(e,t){switch(e){case"premajor":this.prerelease.length=0,this.patch=0,this.minor=0,this.major++,this.inc("pre",t);break;case"preminor":this.prerelease.length=0,this.patch=0,this.minor++,this.inc("pre",t);break;case"prepatch":this.prerelease.length=0,this.inc("patch",t),this.inc("pre",t);break;case"prerelease":0===this.prerelease.length&&this.inc("patch",t),this.inc("pre",t);break;case"major":0===this.minor&&0===this.patch&&0!==this.prerelease.length||this.major++,this.minor=0,this.patch=0,this.prerelease=[];break;case"minor":0===this.patch&&0!==this.prerelease.length||this.minor++,this.patch=0,this.prerelease=[];break;case"patch":0===this.prerelease.length&&this.patch++,this.prerelease=[];break;case"pre":if(0===this.prerelease.length)this.prerelease=[0];else{for(var r=this.prerelease.length;--r>=0;)"number"==typeof this.prerelease[r]&&(this.prerelease[r]++,r=-2);-1===r&&this.prerelease.push(0)}t&&(this.prerelease[0]===t?isNaN(this.prerelease[1])&&(this.prerelease=[t,0]):this.prerelease=[t,0]);break;default:throw new Error("invalid increment argument: "+e)}return this.format(),this.raw=this.version,this},t.inc=function(e,t,r,n){"string"==typeof r&&(n=r,r=void 0);try{return new f(e,r).inc(t,n).version}catch(e){return null}},t.diff=function(e,t){if(y(e,t))return null;var r=l(e),n=l(t),i="";if(r.prerelease.length||n.prerelease.length){i="pre";var o="prerelease"}for(var s in r)if(("major"===s||"minor"===s||"patch"===s)&&r[s]!==n[s])return i+s;return o},t.compareIdentifiers=h;var p=/^[0-9]+$/;function h(e,t){var r=p.test(e),n=p.test(t);return r&&n&&(e=+e,t=+t),e===t?0:r&&!n?-1:n&&!r?1:e<t?-1:1}function d(e,t,r){return new f(e,r).compare(new f(t,r))}function v(e,t,r){return d(e,t,r)>0}function g(e,t,r){return d(e,t,r)<0}function y(e,t,r){return 0===d(e,t,r)}function m(e,t,r){return 0!==d(e,t,r)}function E(e,t,r){return d(e,t,r)>=0}function b(e,t,r){return d(e,t,r)<=0}function w(e,t,r,n){switch(t){case"===":return"object"==typeof e&&(e=e.version),"object"==typeof r&&(r=r.version),e===r;case"!==":return"object"==typeof e&&(e=e.version),"object"==typeof r&&(r=r.version),e!==r;case"":case"=":case"==":return y(e,r,n);case"!=":return m(e,r,n);case">":return v(e,r,n);case">=":return E(e,r,n);case"<":return g(e,r,n);case"<=":return b(e,r,n);default:throw new TypeError("Invalid operator: "+t)}}function O(e,t){if(t&&"object"==typeof t||(t={loose:!!t,includePrerelease:!1}),e instanceof O){if(e.loose===!!t.loose)return e;e=e.value}if(!(this instanceof O))return new O(e,t);r("comparator",e,t),this.options=t,this.loose=!!t.loose,this.parse(e),this.semver===S?this.value="":this.value=this.operator+this.semver.version,r("comp",this)}t.rcompareIdentifiers=function(e,t){return h(t,e)},t.major=function(e,t){return new f(e,t).major},t.minor=function(e,t){return new f(e,t).minor},t.patch=function(e,t){return new f(e,t).patch},t.compare=d,t.compareLoose=function(e,t){return d(e,t,!0)},t.compareBuild=function(e,t,r){var n=new f(e,r),i=new f(t,r);return n.compare(i)||n.compareBuild(i)},t.rcompare=function(e,t,r){return d(t,e,r)},t.sort=function(e,r){return e.sort((function(e,n){return t.compareBuild(e,n,r)}))},t.rsort=function(e,r){return e.sort((function(e,n){return t.compareBuild(n,e,r)}))},t.gt=v,t.lt=g,t.eq=y,t.neq=m,t.gte=E,t.lte=b,t.cmp=w,t.Comparator=O;var S={};function R(e,t){if(t&&"object"==typeof t||(t={loose:!!t,includePrerelease:!1}),e instanceof R)return e.loose===!!t.loose&&e.includePrerelease===!!t.includePrerelease?e:new R(e.raw,t);if(e instanceof O)return new R(e.value,t);if(!(this instanceof R))return new R(e,t);if(this.options=t,this.loose=!!t.loose,this.includePrerelease=!!t.includePrerelease,this.raw=e,this.set=e.split(/\s*\|\|\s*/).map((function(e){return this.parseRange(e.trim())}),this).filter((function(e){return e.length})),!this.set.length)throw new TypeError("Invalid SemVer Range: "+e);this.format()}function I(e,t){for(var r=!0,n=e.slice(),i=n.pop();r&&n.length;)r=n.every((function(e){return i.intersects(e,t)})),i=n.pop();return r}function A(e){return!e||"x"===e.toLowerCase()||"*"===e}function j(e,t,r,n,i,o,s,c,a,u,l,f,p){return((t=A(r)?"":A(n)?">="+r+".0.0":A(i)?">="+r+"."+n+".0":">="+t)+" "+(c=A(a)?"":A(u)?"<"+(+a+1)+".0.0":A(l)?"<"+a+"."+(+u+1)+".0":f?"<="+a+"."+u+"."+l+"-"+f:"<="+c)).trim()}function x(e,t,n){for(var i=0;i<e.length;i++)if(!e[i].test(t))return!1;if(t.prerelease.length&&!n.includePrerelease){for(i=0;i<e.length;i++)if(r(e[i].semver),e[i].semver!==S&&e[i].semver.prerelease.length>0){var o=e[i].semver;if(o.major===t.major&&o.minor===t.minor&&o.patch===t.patch)return!0}return!1}return!0}function k(e,t,r){try{t=new R(t,r)}catch(e){return!1}return t.test(e)}function L(e,t,r,n){var i,o,s,c,a;switch(e=new f(e,n),t=new R(t,n),r){case">":i=v,o=b,s=g,c=">",a=">=";break;case"<":i=g,o=E,s=v,c="<",a="<=";break;default:throw new TypeError('Must provide a hilo val of "<" or ">"')}if(k(e,t,n))return!1;for(var u=0;u<t.set.length;++u){var l=t.set[u],p=null,h=null;if(l.forEach((function(e){e.semver===S&&(e=new O(">=0.0.0")),p=p||e,h=h||e,i(e.semver,p.semver,n)?p=e:s(e.semver,h.semver,n)&&(h=e)})),p.operator===c||p.operator===a)return!1;if((!h.operator||h.operator===c)&&o(e,h.semver))return!1;if(h.operator===a&&s(e,h.semver))return!1}return!0}O.prototype.parse=function(e){var t=this.options.loose?i[s.COMPARATORLOOSE]:i[s.COMPARATOR],r=e.match(t);if(!r)throw new TypeError("Invalid comparator: "+e);this.operator=void 0!==r[1]?r[1]:"","="===this.operator&&(this.operator=""),r[2]?this.semver=new f(r[2],this.options.loose):this.semver=S},O.prototype.toString=function(){return this.value},O.prototype.test=function(e){if(r("Comparator.test",e,this.options.loose),this.semver===S||e===S)return!0;if("string"==typeof e)try{e=new f(e,this.options)}catch(e){return!1}return w(e,this.operator,this.semver,this.options)},O.prototype.intersects=function(e,t){if(!(e instanceof O))throw new TypeError("a Comparator is required");var r;if(t&&"object"==typeof t||(t={loose:!!t,includePrerelease:!1}),""===this.operator)return""===this.value||(r=new R(e.value,t),k(this.value,r,t));if(""===e.operator)return""===e.value||(r=new R(this.value,t),k(e.semver,r,t));var n=!(">="!==this.operator&&">"!==this.operator||">="!==e.operator&&">"!==e.operator),i=!("<="!==this.operator&&"<"!==this.operator||"<="!==e.operator&&"<"!==e.operator),o=this.semver.version===e.semver.version,s=!(">="!==this.operator&&"<="!==this.operator||">="!==e.operator&&"<="!==e.operator),c=w(this.semver,"<",e.semver,t)&&(">="===this.operator||">"===this.operator)&&("<="===e.operator||"<"===e.operator),a=w(this.semver,">",e.semver,t)&&("<="===this.operator||"<"===this.operator)&&(">="===e.operator||">"===e.operator);return n||i||o&&s||c||a},t.Range=R,R.prototype.format=function(){return this.range=this.set.map((function(e){return e.join(" ").trim()})).join("||").trim(),this.range},R.prototype.toString=function(){return this.range},R.prototype.parseRange=function(e){var t=this.options.loose;e=e.trim();var n=t?i[s.HYPHENRANGELOOSE]:i[s.HYPHENRANGE];e=e.replace(n,j),r("hyphen replace",e),e=e.replace(i[s.COMPARATORTRIM],"$1$2$3"),r("comparator trim",e,i[s.COMPARATORTRIM]),e=(e=(e=e.replace(i[s.TILDETRIM],"$1~")).replace(i[s.CARETTRIM],"$1^")).split(/\s+/).join(" ");var o=t?i[s.COMPARATORLOOSE]:i[s.COMPARATOR],c=e.split(" ").map((function(e){return function(e,t){return r("comp",e,t),e=function(e,t){return e.trim().split(/\s+/).map((function(e){return function(e,t){r("caret",e,t);var n=t.loose?i[s.CARETLOOSE]:i[s.CARET];return e.replace(n,(function(t,n,i,o,s){var c;return r("caret",e,t,n,i,o,s),A(n)?c="":A(i)?c=">="+n+".0.0 <"+(+n+1)+".0.0":A(o)?c="0"===n?">="+n+"."+i+".0 <"+n+"."+(+i+1)+".0":">="+n+"."+i+".0 <"+(+n+1)+".0.0":s?(r("replaceCaret pr",s),c="0"===n?"0"===i?">="+n+"."+i+"."+o+"-"+s+" <"+n+"."+i+"."+(+o+1):">="+n+"."+i+"."+o+"-"+s+" <"+n+"."+(+i+1)+".0":">="+n+"."+i+"."+o+"-"+s+" <"+(+n+1)+".0.0"):(r("no pr"),c="0"===n?"0"===i?">="+n+"."+i+"."+o+" <"+n+"."+i+"."+(+o+1):">="+n+"."+i+"."+o+" <"+n+"."+(+i+1)+".0":">="+n+"."+i+"."+o+" <"+(+n+1)+".0.0"),r("caret return",c),c}))}(e,t)})).join(" ")}(e,t),r("caret",e),e=function(e,t){return e.trim().split(/\s+/).map((function(e){return function(e,t){var n=t.loose?i[s.TILDELOOSE]:i[s.TILDE];return e.replace(n,(function(t,n,i,o,s){var c;return r("tilde",e,t,n,i,o,s),A(n)?c="":A(i)?c=">="+n+".0.0 <"+(+n+1)+".0.0":A(o)?c=">="+n+"."+i+".0 <"+n+"."+(+i+1)+".0":s?(r("replaceTilde pr",s),c=">="+n+"."+i+"."+o+"-"+s+" <"+n+"."+(+i+1)+".0"):c=">="+n+"."+i+"."+o+" <"+n+"."+(+i+1)+".0",r("tilde return",c),c}))}(e,t)})).join(" ")}(e,t),r("tildes",e),e=function(e,t){return r("replaceXRanges",e,t),e.split(/\s+/).map((function(e){return function(e,t){e=e.trim();var n=t.loose?i[s.XRANGELOOSE]:i[s.XRANGE];return e.replace(n,(function(n,i,o,s,c,a){r("xRange",e,n,i,o,s,c,a);var u=A(o),l=u||A(s),f=l||A(c),p=f;return"="===i&&p&&(i=""),a=t.includePrerelease?"-0":"",u?n=">"===i||"<"===i?"<0.0.0-0":"*":i&&p?(l&&(s=0),c=0,">"===i?(i=">=",l?(o=+o+1,s=0,c=0):(s=+s+1,c=0)):"<="===i&&(i="<",l?o=+o+1:s=+s+1),n=i+o+"."+s+"."+c+a):l?n=">="+o+".0.0"+a+" <"+(+o+1)+".0.0"+a:f&&(n=">="+o+"."+s+".0"+a+" <"+o+"."+(+s+1)+".0"+a),r("xRange return",n),n}))}(e,t)})).join(" ")}(e,t),r("xrange",e),e=function(e,t){return r("replaceStars",e,t),e.trim().replace(i[s.STAR],"")}(e,t),r("stars",e),e}(e,this.options)}),this).join(" ").split(/\s+/);return this.options.loose&&(c=c.filter((function(e){return!!e.match(o)}))),c=c.map((function(e){return new O(e,this.options)}),this)},R.prototype.intersects=function(e,t){if(!(e instanceof R))throw new TypeError("a Range is required");return this.set.some((function(r){return I(r,t)&&e.set.some((function(e){return I(e,t)&&r.every((function(r){return e.every((function(e){return r.intersects(e,t)}))}))}))}))},t.toComparators=function(e,t){return new R(e,t).set.map((function(e){return e.map((function(e){return e.value})).join(" ").trim().split(" ")}))},R.prototype.test=function(e){if(!e)return!1;if("string"==typeof e)try{e=new f(e,this.options)}catch(e){return!1}for(var t=0;t<this.set.length;t++)if(x(this.set[t],e,this.options))return!0;return!1},t.satisfies=k,t.maxSatisfying=function(e,t,r){var n=null,i=null;try{var o=new R(t,r)}catch(e){return null}return e.forEach((function(e){o.test(e)&&(n&&-1!==i.compare(e)||(i=new f(n=e,r)))})),n},t.minSatisfying=function(e,t,r){var n=null,i=null;try{var o=new R(t,r)}catch(e){return null}return e.forEach((function(e){o.test(e)&&(n&&1!==i.compare(e)||(i=new f(n=e,r)))})),n},t.minVersion=function(e,t){e=new R(e,t);var r=new f("0.0.0");if(e.test(r))return r;if(r=new f("0.0.0-0"),e.test(r))return r;r=null;for(var n=0;n<e.set.length;++n){e.set[n].forEach((function(e){var t=new f(e.semver.version);switch(e.operator){case">":0===t.prerelease.length?t.patch++:t.prerelease.push(0),t.raw=t.format();case"":case">=":r&&!v(r,t)||(r=t);break;case"<":case"<=":break;default:throw new Error("Unexpected operation: "+e.operator)}}))}if(r&&e.test(r))return r;return null},t.validRange=function(e,t){try{return new R(e,t).range||"*"}catch(e){return null}},t.ltr=function(e,t,r){return L(e,t,"<",r)},t.gtr=function(e,t,r){return L(e,t,">",r)},t.outside=L,t.prerelease=function(e,t){var r=l(e,t);return r&&r.prerelease.length?r.prerelease:null},t.intersects=function(e,t,r){return e=new R(e,r),t=new R(t,r),e.intersects(t)},t.coerce=function(e,t){if(e instanceof f)return e;"number"==typeof e&&(e=String(e));if("string"!=typeof e)return null;var r=null;if((t=t||{}).rtl){for(var n;(n=i[s.COERCERTL].exec(e))&&(!r||r.index+r[0].length!==e.length);)r&&n.index+n[0].length===r.index+r[0].length||(r=n),i[s.COERCERTL].lastIndex=n.index+n[1].length+n[2].length;i[s.COERCERTL].lastIndex=-1}else r=e.match(i[s.COERCE]);if(null===r)return null;return l(r[2]+"."+(r[3]||"0")+"."+(r[4]||"0"),t)}}(0,Ny.exports);var Py=g(Ny.exports);const Fy=global,Dy=new Proxy(global.target,{set:function(e,t,r){return Reflect.set(...arguments)},get:function(e,t,r){return Uy(`make ${t}`),Reflect.get(...arguments)}}),Cy=["packages","codemods","eslint"],My=["eslint","scripts","benchmark",...Cy,"*.{js,cjs,mjs,ts}","--format","codeframe","--ext",".js,.cjs,.mjs,.ts"],Gy=Fy.which("yarn").stdout,$y=process.execPath;function Uy(...e){console.log.apply(console,e)}function By(e,t,r,n=!0){Uy(`${e.replaceAll(Gy,"yarn").replaceAll($y,"node")} ${t.join(" ")}`);try{return s.execFileSync(e,t,{stdio:n?"inherit":void 0,cwd:r&&l.default.resolve(r),env:process.env})}catch(r){throw n&&0!=r.status&&(console.error(new Error(`\ncommand: ${e} ${t.join(" ")}\ncode: ${r.status}`)),process.exit(r.status)),r}}function Xy(e,t,r){return By(Gy,e,t,r)}function qy(e,t,r){return By($y,e,t,r)}function Wy(e,t){const r=process.env;process.env=Object.assign(Object.assign({},r),t),e(),process.env=r}function Vy(e,r,n){const i="./build/"+e.toLowerCase();Fy.rm("-rf",i),Fy.mkdir("-p","build"),By("git",["clone","--filter=blob:none","--sparse","--single-branch","--shallow-since='2 years ago'",r,i]),By("git",["sparse-checkout","set",...n],i),By("git",["checkout","-q",function(e){const r=t.readFileSync("./Makefile","utf8").match(new RegExp(`${e}_COMMIT = (\\w{40})`))[1];if(!r)throw new Error(`Could not find ${e}_COMMIT in Makefile`);return r}(e)],i)}Fy.config.verbose=!0,Dy["clean-all"]=function(){Fy.rm("-rf",["node_modules","package-lock.json",".changelog"]),Cy.forEach((e=>{Fy.rm("-rf",`${e}/*/test/tmp`),Fy.rm("-rf",`${e}/*/package-lock.json`)})),Dy.clean(),Dy["clean-lib"]()},Dy.clean=function(){Dy["test-clean"](),Fy.rm("-rf",[".npmrc","coverage","packages/*/npm-debug*","node_modules/.cache"])},Dy["test-clean"]=function(){Cy.forEach((e=>{Fy.rm("-rf",`${e}/*/test/tmp`),Fy.rm("-rf",`${e}/*/test-fixtures.json`)}))},Dy["clean-lib"]=function(){Fy.rm("-rf",Cy.map((e=>`${e}/*/lib`)))},Dy["clean-runtime-helpers"]=function(){Fy.rm("-rf",["packages/babel-runtime/helpers/**/*.js","packages/babel-runtime-corejs2/helpers/**/*.js","packages/babel-runtime-corejs3/helpers/**/*.js","packages/babel-runtime/helpers/**/*.mjs","packages/babel-runtime-corejs2/helpers/**/*.mjs","packages/babel-runtime-corejs3/helpers/**/*.mjs","packages/babel-runtime-corejs2/core-js"])},Dy["use-cjs"]=function(){qy(["scripts/set-module-type.js","script"]),Dy.bootstrap()},Dy["use-esm"]=function(){qy(["scripts/set-module-type.js","module"]),Dy.bootstrap()},Dy["bootstrap-only"]=function(){Dy["clean-all"](),Xy(["install"])},Dy.bootstrap=function(){Dy["bootstrap-only"](),Dy["generate-tsconfig"](),Dy.build()},Dy.build=function(){Dy["build-no-bundle"](),"true"!=process.env.BABEL_COVERAGE&&Dy["build-standalone"]()},Dy["build-standalone"]=function(){Xy(["gulp","build-babel-standalone"])},Dy["build-bundle"]=function(){Dy.clean(),Dy["clean-lib"](),qy(["scripts/set-module-type.js"]),Xy(["gulp","build"]),Dy["build-flow-typings"](),Dy["build-dist"]()},Dy["build-no-bundle"]=function(){Dy.clean(),Dy["clean-lib"](),qy(["scripts/set-module-type.js"]),Wy((()=>{Xy(["gulp","build-dev"])}),{BABEL_ENV:"development"}),Dy["build-flow-typings"](),Dy["build-dist"]()},Dy["build-flow-typings"]=function(){t.writeFileSync("packages/babel-types/lib/index.js.flow",qy(["packages/babel-types/scripts/generators/flow.js"],void 0,!1))},Dy["build-dist"]=function(){Dy["build-plugin-transform-runtime-dist"]()},Dy["build-plugin-transform-runtime-dist"]=function(){qy(["scripts/build-dist.js"],"packages/babel-plugin-transform-runtime")},Dy.prepublish=function(){process.env.BABEL_8_BREAKING?qy(["scripts/set-module-type.js","module"]):qy(["scripts/set-module-type.js","script"]),Dy["bootstrap-only"](),Wy((()=>{Dy["prepublish-build"](),Dy.test()}),{IS_PUBLISH:"true"}),qy(["scripts/set-module-type.js","clean"])},Dy["prepublish-build"]=function(){Dy["clean-lib"](),Dy["clean-runtime-helpers"](),Wy((()=>{Dy["build-bundle"]()}),{NODE_ENV:"production",BABEL_ENV:"production",STRIP_BABEL_8_FLAG:"true"}),Wy((()=>{Dy["prepublish-build-standalone"](),Dy["clone-license"](),Dy["prepublish-prepare-dts"]()}),{NODE_ENV:"production",STRIP_BABEL_8_FLAG:"true"})},Dy["prepublish-build-standalone"]=function(){Wy((()=>{Dy["build-standalone"]()}),{BABEL_ENV:"production",IS_PUBLISH:"true"})},Dy["prepublish-prepare-dts"]=function(){Dy.tscheck(),Xy(["gulp","bundle-dts"]),Dy["build-typescript-legacy-typings"]()},Dy.tscheck=function(){Dy["generate-tsconfig"](),Fy.rm("-rf","tsconfig.tsbuildinfo"),Fy.rm("-rf","dts"),Xy(["tsc","-b","."])},Dy["generate-tsconfig"]=function(){qy(["scripts/generators/tsconfig.js"]),qy(["scripts/generators/archived-libs-typings.js"])},Dy["generate-type-helpers"]=function(){Xy(["gulp","generate-type-helpers"])},Dy["build-typescript-legacy-typings"]=function(){t.writeFileSync("packages/babel-types/lib/index-legacy.d.ts",qy(["packages/babel-types/scripts/generators/typescript-legacy.js"],void 0,!1))},Dy["clone-license"]=function(){qy(["scripts/clone-license.js"])},Dy.lint=function(){Wy((()=>{Xy(My)}),{BABEL_ENV:"test"})},Dy.fix=function(){Dy["fix-json"](),Dy["fix-js"]()},Dy["fix-js"]=function(){Xy([...My,"--fix"])},Dy["fix-json"]=function(){Xy(["prettier",`{${Cy.join(",")}}/*/test/fixtures/**/options.json`,"--write","--loglevel","warn"])},Dy.watch=function(){Dy["build-no-bundle"](),Wy((()=>{Xy(["gulp","watch"])}),{BABEL_ENV:"development",WATCH_SKIP_BUILD:"true"})},Dy.test=function(){Dy.lint(),Dy["test-only"]()},Dy["test-only"]=function(e=[]){Xy(["jest",...e])},Dy["test-cov"]=function(){Dy.build(),Wy((()=>{Xy(["c8","jest"])}),{BABEL_ENV:"test",BABEL_COVERAGE:"true"})},Dy["bootstrap-test262"]=function(){Vy("TEST262","https://github.com/tc39/test262.git",["test","harness"])},Dy["bootstrap-typescript"]=function(){Vy("TYPESCRIPT","https://github.com/microsoft/TypeScript.git",["tests"])},Dy["bootstrap-flow"]=function(){Vy("FLOW","https://github.com/facebook/flow.git",["src/parser/test/flow"])},Dy["new-version-checklist"]=function(){console.log("\n!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!\n!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!\n!!!!!!                                                   !!!!!!\n!!!!!!              Update the version in                !!!!!!\n!!!!!! packages/babel-plugin-syntax-import-attributes/src/index.ts\n!!!!!!                    and                            !!!!!!\n!!!!!! /Users/nic/Documents/dev/babel/babel/packages/babel-preset-env/src/available-plugins.ts#L159\n!!!!!!                                                   !!!!!!\n!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!\n!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!\n    ".trim()),process.exit(1)},Dy["new-version"]=function(){Dy["new-version-checklist"](),By("git",["pull","--rebase"]),Xy(["release-tool","version","-f","@babel/standalone"])},Dy["new-babel-8-version-prepare"]=function(){By("git",["pull","--rebase"]);const e=JSON.parse(t.readFileSync("./package.json","utf8")),r=Py.inc(e.version_babel8,"prerelease");return e.version_babel8=r,t.writeFileSync("./package.json",JSON.stringify(e,null,2)+"\n"),By("git",["add","./package.json"]),By("git",["commit","-m","Bump Babel 8 version to "+r]),r},Dy["new-babel-8-version"]=function(){const e=Dy["new-babel-8-version-prepare"]();By("git",["checkout","-b","release/v"+e]),Cy.forEach((r=>{t.readdirSync(r).forEach((n=>{const i=`${r}/${n}/package.json`;try{const r=JSON.parse(t.readFileSync(i,"utf8"));r.peerDependencies["@babel/core"]=`^${e}`,t.writeFileSync(i,JSON.stringify(r,null,2)+"\n")}catch(e){}}))})),Xy(["release-tool","version",e,"--all"]),console.log(`Run \`git push upstream main release/v${e} --follow-tags\` to push the changes on GitHub and release them.`)};
+"use strict";var e=require("os"),r=require("fs"),t=require("path"),n=require("events"),i=require("assert"),o=require("util"),s=require("child_process");function c(e){return e&&"object"==typeof e&&"default"in e?e:{default:e}}var a=c(e),u=c(r),l=c(t),f=c(n),p=c(i),h=c(o),d=c(s),v="undefined"!=typeof globalThis?globalThis:"undefined"!=typeof window?window:"undefined"!=typeof global?global:"undefined"!=typeof self?self:{};function g(e){return e&&e.__esModule&&Object.prototype.hasOwnProperty.call(e,"default")?e.default:e}var y=function(e){return e&&e.Math==Math&&e},m=y("object"==typeof globalThis&&globalThis)||y("object"==typeof window&&window)||y("object"==typeof self&&self)||y("object"==typeof v&&v)||function(){return this}()||v||Function("return this")(),E={},b=function(e){try{return!!e()}catch(e){return!0}},w=!b((function(){return 7!=Object.defineProperty({},1,{get:function(){return 7}})[1]})),O=!b((function(){var e=function(){}.bind();return"function"!=typeof e||e.hasOwnProperty("prototype")})),S=O,R=Function.prototype.call,I=S?R.bind(R):function(){return R.apply(R,arguments)},A={},j={}.propertyIsEnumerable,x=Object.getOwnPropertyDescriptor,k=x&&!j.call({1:2},1);A.f=k?function(e){var r=x(this,e);return!!r&&r.enumerable}:j;var L,_,T=function(e,r){return{enumerable:!(1&e),configurable:!(2&e),writable:!(4&e),value:r}},N=O,P=Function.prototype,F=P.call,D=N&&P.bind.bind(F,F),C=N?D:function(e){return function(){return F.apply(e,arguments)}},M=C,G=M({}.toString),$=M("".slice),U=function(e){return $(G(e),8,-1)},B=b,X=U,q=Object,W=C("".split),V=B((function(){return!q("z").propertyIsEnumerable(0)}))?function(e){return"String"==X(e)?W(e,""):q(e)}:q,H=function(e){return null==e},z=H,Y=TypeError,K=function(e){if(z(e))throw Y("Can't call method on "+e);return e},J=V,Q=K,Z=function(e){return J(Q(e))},ee="object"==typeof document&&document.all,re={all:ee,IS_HTMLDDA:void 0===ee&&void 0!==ee},te=re.all,ne=re.IS_HTMLDDA?function(e){return"function"==typeof e||e===te}:function(e){return"function"==typeof e},ie=ne,oe=re.all,se=re.IS_HTMLDDA?function(e){return"object"==typeof e?null!==e:ie(e)||e===oe}:function(e){return"object"==typeof e?null!==e:ie(e)},ce=m,ae=ne,ue=function(e){return ae(e)?e:void 0},le=function(e,r){return arguments.length<2?ue(ce[e]):ce[e]&&ce[e][r]},fe=C({}.isPrototypeOf),pe="undefined"!=typeof navigator&&String(navigator.userAgent)||"",he=m,de=pe,ve=he.process,ge=he.Deno,ye=ve&&ve.versions||ge&&ge.version,me=ye&&ye.v8;me&&(_=(L=me.split("."))[0]>0&&L[0]<4?1:+(L[0]+L[1])),!_&&de&&(!(L=de.match(/Edge\/(\d+)/))||L[1]>=74)&&(L=de.match(/Chrome\/(\d+)/))&&(_=+L[1]);var Ee=_,be=Ee,we=b,Oe=m.String,Se=!!Object.getOwnPropertySymbols&&!we((function(){var e=Symbol();return!Oe(e)||!(Object(e)instanceof Symbol)||!Symbol.sham&&be&&be<41})),Re=Se&&!Symbol.sham&&"symbol"==typeof Symbol.iterator,Ie=le,Ae=ne,je=fe,xe=Object,ke=Re?function(e){return"symbol"==typeof e}:function(e){var r=Ie("Symbol");return Ae(r)&&je(r.prototype,xe(e))},Le=String,_e=function(e){try{return Le(e)}catch(e){return"Object"}},Te=ne,Ne=_e,Pe=TypeError,Fe=function(e){if(Te(e))return e;throw Pe(Ne(e)+" is not a function")},De=Fe,Ce=H,Me=function(e,r){var t=e[r];return Ce(t)?void 0:De(t)},Ge=I,$e=ne,Ue=se,Be=TypeError,Xe={exports:{}},qe=m,We=Object.defineProperty,Ve=function(e,r){try{We(qe,e,{value:r,configurable:!0,writable:!0})}catch(t){qe[e]=r}return r},He=Ve,ze=m["__core-js_shared__"]||He("__core-js_shared__",{}),Ye=ze;(Xe.exports=function(e,r){return Ye[e]||(Ye[e]=void 0!==r?r:{})})("versions",[]).push({version:"3.30.2",mode:"global",copyright:"© 2014-2023 Denis Pushkarev (zloirock.ru)",license:"https://github.com/zloirock/core-js/blob/v3.30.2/LICENSE",source:"https://github.com/zloirock/core-js"});var Ke=Xe.exports,Je=K,Qe=Object,Ze=function(e){return Qe(Je(e))},er=Ze,rr=C({}.hasOwnProperty),tr=Object.hasOwn||function(e,r){return rr(er(e),r)},nr=C,ir=0,or=Math.random(),sr=nr(1..toString),cr=function(e){return"Symbol("+(void 0===e?"":e)+")_"+sr(++ir+or,36)},ar=Ke,ur=tr,lr=cr,fr=Se,pr=Re,hr=m.Symbol,dr=ar("wks"),vr=pr?hr.for||hr:hr&&hr.withoutSetter||lr,gr=function(e){return ur(dr,e)||(dr[e]=fr&&ur(hr,e)?hr[e]:vr("Symbol."+e)),dr[e]},yr=I,mr=se,Er=ke,br=Me,wr=function(e,r){var t,n;if("string"===r&&$e(t=e.toString)&&!Ue(n=Ge(t,e)))return n;if($e(t=e.valueOf)&&!Ue(n=Ge(t,e)))return n;if("string"!==r&&$e(t=e.toString)&&!Ue(n=Ge(t,e)))return n;throw Be("Can't convert object to primitive value")},Or=TypeError,Sr=gr("toPrimitive"),Rr=function(e,r){if(!mr(e)||Er(e))return e;var t,n=br(e,Sr);if(n){if(void 0===r&&(r="default"),t=yr(n,e,r),!mr(t)||Er(t))return t;throw Or("Can't convert object to primitive value")}return void 0===r&&(r="number"),wr(e,r)},Ir=ke,Ar=function(e){var r=Rr(e,"string");return Ir(r)?r:r+""},jr=se,xr=m.document,kr=jr(xr)&&jr(xr.createElement),Lr=function(e){return kr?xr.createElement(e):{}},_r=Lr,Tr=!w&&!b((function(){return 7!=Object.defineProperty(_r("div"),"a",{get:function(){return 7}}).a})),Nr=w,Pr=I,Fr=A,Dr=T,Cr=Z,Mr=Ar,Gr=tr,$r=Tr,Ur=Object.getOwnPropertyDescriptor;E.f=Nr?Ur:function(e,r){if(e=Cr(e),r=Mr(r),$r)try{return Ur(e,r)}catch(e){}if(Gr(e,r))return Dr(!Pr(Fr.f,e,r),e[r])};var Br={},Xr=w&&b((function(){return 42!=Object.defineProperty((function(){}),"prototype",{value:42,writable:!1}).prototype})),qr=se,Wr=String,Vr=TypeError,Hr=function(e){if(qr(e))return e;throw Vr(Wr(e)+" is not an object")},zr=w,Yr=Tr,Kr=Xr,Jr=Hr,Qr=Ar,Zr=TypeError,et=Object.defineProperty,rt=Object.getOwnPropertyDescriptor;Br.f=zr?Kr?function(e,r,t){if(Jr(e),r=Qr(r),Jr(t),"function"==typeof e&&"prototype"===r&&"value"in t&&"writable"in t&&!t.writable){var n=rt(e,r);n&&n.writable&&(e[r]=t.value,t={configurable:"configurable"in t?t.configurable:n.configurable,enumerable:"enumerable"in t?t.enumerable:n.enumerable,writable:!1})}return et(e,r,t)}:et:function(e,r,t){if(Jr(e),r=Qr(r),Jr(t),Yr)try{return et(e,r,t)}catch(e){}if("get"in t||"set"in t)throw Zr("Accessors not supported");return"value"in t&&(e[r]=t.value),e};var tt=Br,nt=T,it=w?function(e,r,t){return tt.f(e,r,nt(1,t))}:function(e,r,t){return e[r]=t,e},ot={exports:{}},st=w,ct=tr,at=Function.prototype,ut=st&&Object.getOwnPropertyDescriptor,lt=ct(at,"name"),ft={EXISTS:lt,PROPER:lt&&"something"===function(){}.name,CONFIGURABLE:lt&&(!st||st&&ut(at,"name").configurable)},pt=ne,ht=ze,dt=C(Function.toString);pt(ht.inspectSource)||(ht.inspectSource=function(e){return dt(e)});var vt,gt,yt,mt=ht.inspectSource,Et=ne,bt=m.WeakMap,wt=Et(bt)&&/native code/.test(String(bt)),Ot=cr,St=Ke("keys"),Rt=function(e){return St[e]||(St[e]=Ot(e))},It={},At=wt,jt=m,xt=se,kt=it,Lt=tr,_t=ze,Tt=Rt,Nt=It,Pt=jt.TypeError,Ft=jt.WeakMap;if(At||_t.state){var Dt=_t.state||(_t.state=new Ft);Dt.get=Dt.get,Dt.has=Dt.has,Dt.set=Dt.set,vt=function(e,r){if(Dt.has(e))throw Pt("Object already initialized");return r.facade=e,Dt.set(e,r),r},gt=function(e){return Dt.get(e)||{}},yt=function(e){return Dt.has(e)}}else{var Ct=Tt("state");Nt[Ct]=!0,vt=function(e,r){if(Lt(e,Ct))throw Pt("Object already initialized");return r.facade=e,kt(e,Ct,r),r},gt=function(e){return Lt(e,Ct)?e[Ct]:{}},yt=function(e){return Lt(e,Ct)}}var Mt={set:vt,get:gt,has:yt,enforce:function(e){return yt(e)?gt(e):vt(e,{})},getterFor:function(e){return function(r){var t;if(!xt(r)||(t=gt(r)).type!==e)throw Pt("Incompatible receiver, "+e+" required");return t}}},Gt=C,$t=b,Ut=ne,Bt=tr,Xt=w,qt=ft.CONFIGURABLE,Wt=mt,Vt=Mt.enforce,Ht=Mt.get,zt=String,Yt=Object.defineProperty,Kt=Gt("".slice),Jt=Gt("".replace),Qt=Gt([].join),Zt=Xt&&!$t((function(){return 8!==Yt((function(){}),"length",{value:8}).length})),en=String(String).split("String"),rn=ot.exports=function(e,r,t){"Symbol("===Kt(zt(r),0,7)&&(r="["+Jt(zt(r),/^Symbol\(([^)]*)\)/,"$1")+"]"),t&&t.getter&&(r="get "+r),t&&t.setter&&(r="set "+r),(!Bt(e,"name")||qt&&e.name!==r)&&(Xt?Yt(e,"name",{value:r,configurable:!0}):e.name=r),Zt&&t&&Bt(t,"arity")&&e.length!==t.arity&&Yt(e,"length",{value:t.arity});try{t&&Bt(t,"constructor")&&t.constructor?Xt&&Yt(e,"prototype",{writable:!1}):e.prototype&&(e.prototype=void 0)}catch(e){}var n=Vt(e);return Bt(n,"source")||(n.source=Qt(en,"string"==typeof r?r:"")),e};Function.prototype.toString=rn((function(){return Ut(this)&&Ht(this).source||Wt(this)}),"toString");var tn=ot.exports,nn=ne,on=Br,sn=tn,cn=Ve,an=function(e,r,t,n){n||(n={});var i=n.enumerable,o=void 0!==n.name?n.name:r;if(nn(t)&&sn(t,o,n),n.global)i?e[r]=t:cn(r,t);else{try{n.unsafe?e[r]&&(i=!0):delete e[r]}catch(e){}i?e[r]=t:on.f(e,r,{value:t,enumerable:!1,configurable:!n.nonConfigurable,writable:!n.nonWritable})}return e},un={},ln=Math.ceil,fn=Math.floor,pn=Math.trunc||function(e){var r=+e;return(r>0?fn:ln)(r)},hn=function(e){var r=+e;return r!=r||0===r?0:pn(r)},dn=hn,vn=Math.max,gn=Math.min,yn=function(e,r){var t=dn(e);return t<0?vn(t+r,0):gn(t,r)},mn=hn,En=Math.min,bn=function(e){return e>0?En(mn(e),9007199254740991):0},wn=bn,On=function(e){return wn(e.length)},Sn=Z,Rn=yn,In=On,An=function(e){return function(r,t,n){var i,o=Sn(r),s=In(o),c=Rn(n,s);if(e&&t!=t){for(;s>c;)if((i=o[c++])!=i)return!0}else for(;s>c;c++)if((e||c in o)&&o[c]===t)return e||c||0;return!e&&-1}},jn={includes:An(!0),indexOf:An(!1)},xn=tr,kn=Z,Ln=jn.indexOf,_n=It,Tn=C([].push),Nn=function(e,r){var t,n=kn(e),i=0,o=[];for(t in n)!xn(_n,t)&&xn(n,t)&&Tn(o,t);for(;r.length>i;)xn(n,t=r[i++])&&(~Ln(o,t)||Tn(o,t));return o},Pn=["constructor","hasOwnProperty","isPrototypeOf","propertyIsEnumerable","toLocaleString","toString","valueOf"],Fn=Nn,Dn=Pn.concat("length","prototype");un.f=Object.getOwnPropertyNames||function(e){return Fn(e,Dn)};var Cn={};Cn.f=Object.getOwnPropertySymbols;var Mn=le,Gn=un,$n=Cn,Un=Hr,Bn=C([].concat),Xn=Mn("Reflect","ownKeys")||function(e){var r=Gn.f(Un(e)),t=$n.f;return t?Bn(r,t(e)):r},qn=tr,Wn=Xn,Vn=E,Hn=Br,zn=function(e,r,t){for(var n=Wn(r),i=Hn.f,o=Vn.f,s=0;s<n.length;s++){var c=n[s];qn(e,c)||t&&qn(t,c)||i(e,c,o(r,c))}},Yn=b,Kn=ne,Jn=/#|\.prototype\./,Qn=function(e,r){var t=ei[Zn(e)];return t==ti||t!=ri&&(Kn(r)?Yn(r):!!r)},Zn=Qn.normalize=function(e){return String(e).replace(Jn,".").toLowerCase()},ei=Qn.data={},ri=Qn.NATIVE="N",ti=Qn.POLYFILL="P",ni=Qn,ii=m,oi=E.f,si=it,ci=an,ai=Ve,ui=zn,li=ni,fi=function(e,r){var t,n,i,o,s,c=e.target,a=e.global,u=e.stat;if(t=a?ii:u?ii[c]||ai(c,{}):(ii[c]||{}).prototype)for(n in r){if(o=r[n],i=e.dontCallGetSet?(s=oi(t,n))&&s.value:t[n],!li(a?n:c+(u?".":"#")+n,e.forced)&&void 0!==i){if(typeof o==typeof i)continue;ui(o,i)}(e.sham||i&&i.sham)&&si(o,"sham",!0),ci(t,n,o,e)}},pi=Br.f,hi=tr,di=gr("toStringTag"),vi=function(e,r,t){e&&!t&&(e=e.prototype),e&&!hi(e,di)&&pi(e,di,{configurable:!0,value:r})},gi=m,yi=vi;fi({global:!0},{Reflect:{}}),yi(gi.Reflect,"Reflect",!0);var mi={},Ei=Nn,bi=Pn,wi=Object.keys||function(e){return Ei(e,bi)},Oi=w,Si=Xr,Ri=Br,Ii=Hr,Ai=Z,ji=wi;mi.f=Oi&&!Si?Object.defineProperties:function(e,r){Ii(e);for(var t,n=Ai(r),i=ji(r),o=i.length,s=0;o>s;)Ri.f(e,t=i[s++],n[t]);return e};var xi,ki=le("document","documentElement"),Li=Hr,_i=mi,Ti=Pn,Ni=It,Pi=ki,Fi=Lr,Di=Rt("IE_PROTO"),Ci=function(){},Mi=function(e){return"<script>"+e+"<\/script>"},Gi=function(e){e.write(Mi("")),e.close();var r=e.parentWindow.Object;return e=null,r},$i=function(){try{xi=new ActiveXObject("htmlfile")}catch(e){}var e,r;$i="undefined"!=typeof document?document.domain&&xi?Gi(xi):((r=Fi("iframe")).style.display="none",Pi.appendChild(r),r.src=String("javascript:"),(e=r.contentWindow.document).open(),e.write(Mi("document.F=Object")),e.close(),e.F):Gi(xi);for(var t=Ti.length;t--;)delete $i.prototype[Ti[t]];return $i()};Ni[Di]=!0;var Ui=Object.create||function(e,r){var t;return null!==e?(Ci.prototype=Li(e),t=new Ci,Ci.prototype=null,t[Di]=e):t=$i(),void 0===r?t:_i.f(t,r)},Bi=gr,Xi=Ui,qi=Br.f,Wi=Bi("unscopables"),Vi=Array.prototype;null==Vi[Wi]&&qi(Vi,Wi,{configurable:!0,value:Xi(null)});var Hi,zi,Yi,Ki={},Ji=!b((function(){function e(){}return e.prototype.constructor=null,Object.getPrototypeOf(new e)!==e.prototype})),Qi=tr,Zi=ne,eo=Ze,ro=Ji,to=Rt("IE_PROTO"),no=Object,io=no.prototype,oo=ro?no.getPrototypeOf:function(e){var r=eo(e);if(Qi(r,to))return r[to];var t=r.constructor;return Zi(t)&&r instanceof t?t.prototype:r instanceof no?io:null},so=b,co=ne,ao=se,uo=oo,lo=an,fo=gr("iterator"),po=!1;[].keys&&("next"in(Yi=[].keys())?(zi=uo(uo(Yi)))!==Object.prototype&&(Hi=zi):po=!0);var ho=!ao(Hi)||so((function(){var e={};return Hi[fo].call(e)!==e}));ho&&(Hi={}),co(Hi[fo])||lo(Hi,fo,(function(){return this}));var vo={IteratorPrototype:Hi,BUGGY_SAFARI_ITERATORS:po},go=vo.IteratorPrototype,yo=Ui,mo=T,Eo=vi,bo=Ki,wo=function(){return this},Oo=C,So=Fe,Ro=ne,Io=String,Ao=TypeError,jo=function(e,r,t){try{return Oo(So(Object.getOwnPropertyDescriptor(e,r)[t]))}catch(e){}},xo=Hr,ko=function(e){if("object"==typeof e||Ro(e))return e;throw Ao("Can't set "+Io(e)+" as a prototype")},Lo=Object.setPrototypeOf||("__proto__"in{}?function(){var e,r=!1,t={};try{(e=jo(Object.prototype,"__proto__","set"))(t,[]),r=t instanceof Array}catch(e){}return function(t,n){return xo(t),ko(n),r?e(t,n):t.__proto__=n,t}}():void 0),_o=fi,To=I,No=ne,Po=function(e,r,t,n){var i=r+" Iterator";return e.prototype=yo(go,{next:mo(+!n,t)}),Eo(e,i,!1),bo[i]=wo,e},Fo=oo,Do=Lo,Co=vi,Mo=it,Go=an,$o=Ki,Uo=ft.PROPER,Bo=ft.CONFIGURABLE,Xo=vo.IteratorPrototype,qo=vo.BUGGY_SAFARI_ITERATORS,Wo=gr("iterator"),Vo=function(){return this},Ho=Z,zo=function(e){Vi[Wi][e]=!0},Yo=Ki,Ko=Mt,Jo=Br.f,Qo=function(e,r,t,n,i,o,s){Po(t,r,n);var c,a,u,l=function(e){if(e===i&&v)return v;if(!qo&&e in h)return h[e];switch(e){case"keys":case"values":case"entries":return function(){return new t(this,e)}}return function(){return new t(this)}},f=r+" Iterator",p=!1,h=e.prototype,d=h[Wo]||h["@@iterator"]||i&&h[i],v=!qo&&d||l(i),g="Array"==r&&h.entries||d;if(g&&(c=Fo(g.call(new e)))!==Object.prototype&&c.next&&(Fo(c)!==Xo&&(Do?Do(c,Xo):No(c[Wo])||Go(c,Wo,Vo)),Co(c,f,!0)),Uo&&"values"==i&&d&&"values"!==d.name&&(Bo?Mo(h,"name","values"):(p=!0,v=function(){return To(d,this)})),i)if(a={values:l("values"),keys:o?v:l("keys"),entries:l("entries")},s)for(u in a)(qo||p||!(u in h))&&Go(h,u,a[u]);else _o({target:r,proto:!0,forced:qo||p},a);return h[Wo]!==v&&Go(h,Wo,v,{name:i}),$o[r]=v,a},Zo=function(e,r){return{value:e,done:r}},es=w,rs=Ko.set,ts=Ko.getterFor("Array Iterator");Qo(Array,"Array",(function(e,r){rs(this,{type:"Array Iterator",target:Ho(e),index:0,kind:r})}),(function(){var e=ts(this),r=e.target,t=e.kind,n=e.index++;return!r||n>=r.length?(e.target=void 0,Zo(void 0,!0)):Zo("keys"==t?n:"values"==t?r[n]:[n,r[n]],!1)}),"values");var ns=Yo.Arguments=Yo.Array;if(zo("keys"),zo("values"),zo("entries"),es&&"values"!==ns.name)try{Jo(ns,"name",{value:"values"})}catch(e){}var is={};is[gr("toStringTag")]="z";var os="[object z]"===String(is),ss=ne,cs=U,as=gr("toStringTag"),us=Object,ls="Arguments"==cs(function(){return arguments}()),fs=os?cs:function(e){var r,t,n;return void 0===e?"Undefined":null===e?"Null":"string"==typeof(t=function(e,r){try{return e[r]}catch(e){}}(r=us(e),as))?t:ls?cs(r):"Object"==(n=cs(r))&&ss(r.callee)?"Arguments":n},ps=fs,hs=String,ds=function(e){if("Symbol"===ps(e))throw TypeError("Cannot convert a Symbol value to a string");return hs(e)},vs=Hr,gs=function(){var e=vs(this),r="";return e.hasIndices&&(r+="d"),e.global&&(r+="g"),e.ignoreCase&&(r+="i"),e.multiline&&(r+="m"),e.dotAll&&(r+="s"),e.unicode&&(r+="u"),e.unicodeSets&&(r+="v"),e.sticky&&(r+="y"),r},ys=b,ms=m.RegExp,Es=ys((function(){var e=ms("a","y");return e.lastIndex=2,null!=e.exec("abcd")})),bs=Es||ys((function(){return!ms("a","y").sticky})),ws=Es||ys((function(){var e=ms("^r","gy");return e.lastIndex=2,null!=e.exec("str")})),Os={BROKEN_CARET:ws,MISSED_STICKY:bs,UNSUPPORTED_Y:Es},Ss=b,Rs=m.RegExp,Is=Ss((function(){var e=Rs(".","s");return!(e.dotAll&&e.exec("\n")&&"s"===e.flags)})),As=b,js=m.RegExp,xs=As((function(){var e=js("(?<a>b)","g");return"b"!==e.exec("b").groups.a||"bc"!=="b".replace(e,"$<a>c")})),ks=I,Ls=C,_s=ds,Ts=gs,Ns=Os,Ps=Ui,Fs=Mt.get,Ds=Is,Cs=xs,Ms=Ke("native-string-replace",String.prototype.replace),Gs=RegExp.prototype.exec,$s=Gs,Us=Ls("".charAt),Bs=Ls("".indexOf),Xs=Ls("".replace),qs=Ls("".slice),Ws=function(){var e=/a/,r=/b*/g;return ks(Gs,e,"a"),ks(Gs,r,"a"),0!==e.lastIndex||0!==r.lastIndex}(),Vs=Ns.BROKEN_CARET,Hs=void 0!==/()??/.exec("")[1];(Ws||Hs||Vs||Ds||Cs)&&($s=function(e){var r,t,n,i,o,s,c,a=this,u=Fs(a),l=_s(e),f=u.raw;if(f)return f.lastIndex=a.lastIndex,r=ks($s,f,l),a.lastIndex=f.lastIndex,r;var p=u.groups,h=Vs&&a.sticky,d=ks(Ts,a),v=a.source,g=0,y=l;if(h&&(d=Xs(d,"y",""),-1===Bs(d,"g")&&(d+="g"),y=qs(l,a.lastIndex),a.lastIndex>0&&(!a.multiline||a.multiline&&"\n"!==Us(l,a.lastIndex-1))&&(v="(?: "+v+")",y=" "+y,g++),t=new RegExp("^(?:"+v+")",d)),Hs&&(t=new RegExp("^"+v+"$(?!\\s)",d)),Ws&&(n=a.lastIndex),i=ks(Gs,h?t:a,y),h?i?(i.input=qs(i.input,g),i[0]=qs(i[0],g),i.index=a.lastIndex,a.lastIndex+=i[0].length):a.lastIndex=0:Ws&&i&&(a.lastIndex=a.global?i.index+i[0].length:n),Hs&&i&&i.length>1&&ks(Ms,i[0],t,(function(){for(o=1;o<arguments.length-2;o++)void 0===arguments[o]&&(i[o]=void 0)})),i&&p)for(i.groups=s=Ps(null),o=0;o<p.length;o++)s[(c=p[o])[0]]=i[c[1]];return i});var zs=$s;fi({target:"RegExp",proto:!0,forced:/./.exec!==zs},{exec:zs});var Ys=O,Ks=Function.prototype,Js=Ks.apply,Qs=Ks.call,Zs="object"==typeof Reflect&&Reflect.apply||(Ys?Qs.bind(Js):function(){return Qs.apply(Js,arguments)}),ec=U,rc=C,tc=function(e){if("Function"===ec(e))return rc(e)},nc=tc,ic=an,oc=zs,sc=b,cc=gr,ac=it,uc=cc("species"),lc=RegExp.prototype,fc=function(e,r,t,n){var i=cc(e),o=!sc((function(){var r={};return r[i]=function(){return 7},7!=""[e](r)})),s=o&&!sc((function(){var r=!1,t=/a/;return"split"===e&&((t={}).constructor={},t.constructor[uc]=function(){return t},t.flags="",t[i]=/./[i]),t.exec=function(){return r=!0,null},t[i](""),!r}));if(!o||!s||t){var c=nc(/./[i]),a=r(i,""[e],(function(e,r,t,n,i){var s=nc(e),a=r.exec;return a===oc||a===lc.exec?o&&!i?{done:!0,value:c(r,t,n)}:{done:!0,value:s(t,r,n)}:{done:!1}}));ic(String.prototype,e,a[0]),ic(lc,i,a[1])}n&&ac(lc[i],"sham",!0)},pc=C,hc=hn,dc=ds,vc=K,gc=pc("".charAt),yc=pc("".charCodeAt),mc=pc("".slice),Ec=function(e){return function(r,t){var n,i,o=dc(vc(r)),s=hc(t),c=o.length;return s<0||s>=c?e?"":void 0:(n=yc(o,s))<55296||n>56319||s+1===c||(i=yc(o,s+1))<56320||i>57343?e?gc(o,s):n:e?mc(o,s,s+2):i-56320+(n-55296<<10)+65536}},bc={codeAt:Ec(!1),charAt:Ec(!0)}.charAt,wc=function(e,r,t){return r+(t?bc(e,r).length:1)},Oc=C,Sc=Ze,Rc=Math.floor,Ic=Oc("".charAt),Ac=Oc("".replace),jc=Oc("".slice),xc=/\$([$&'`]|\d{1,2}|<[^>]*>)/g,kc=/\$([$&'`]|\d{1,2})/g,Lc=function(e,r,t,n,i,o){var s=t+e.length,c=n.length,a=kc;return void 0!==i&&(i=Sc(i),a=xc),Ac(o,a,(function(o,a){var u;switch(Ic(a,0)){case"$":return"$";case"&":return e;case"`":return jc(r,0,t);case"'":return jc(r,s);case"<":u=i[jc(a,1,-1)];break;default:var l=+a;if(0===l)return o;if(l>c){var f=Rc(l/10);return 0===f?o:f<=c?void 0===n[f-1]?Ic(a,1):n[f-1]+Ic(a,1):o}u=n[l-1]}return void 0===u?"":u}))},_c=I,Tc=Hr,Nc=ne,Pc=U,Fc=zs,Dc=TypeError,Cc=function(e,r){var t=e.exec;if(Nc(t)){var n=_c(t,e,r);return null!==n&&Tc(n),n}if("RegExp"===Pc(e))return _c(Fc,e,r);throw Dc("RegExp#exec called on incompatible receiver")},Mc=Zs,Gc=I,$c=C,Uc=fc,Bc=b,Xc=Hr,qc=ne,Wc=H,Vc=hn,Hc=bn,zc=ds,Yc=K,Kc=wc,Jc=Me,Qc=Lc,Zc=Cc,ea=gr("replace"),ra=Math.max,ta=Math.min,na=$c([].concat),ia=$c([].push),oa=$c("".indexOf),sa=$c("".slice),ca="$0"==="a".replace(/./,"$0"),aa=!!/./[ea]&&""===/./[ea]("a","$0"),ua=!Bc((function(){var e=/./;return e.exec=function(){var e=[];return e.groups={a:"7"},e},"7"!=="".replace(e,"$<a>")}));Uc("replace",(function(e,r,t){var n=aa?"$":"$0";return[function(e,t){var n=Yc(this),i=Wc(e)?void 0:Jc(e,ea);return i?Gc(i,e,n,t):Gc(r,zc(n),e,t)},function(e,i){var o=Xc(this),s=zc(e);if("string"==typeof i&&-1===oa(i,n)&&-1===oa(i,"$<")){var c=t(r,o,s,i);if(c.done)return c.value}var a=qc(i);a||(i=zc(i));var u=o.global;if(u){var l=o.unicode;o.lastIndex=0}for(var f=[];;){var p=Zc(o,s);if(null===p)break;if(ia(f,p),!u)break;""===zc(p[0])&&(o.lastIndex=Kc(s,Hc(o.lastIndex),l))}for(var h,d="",v=0,g=0;g<f.length;g++){for(var y=zc((p=f[g])[0]),m=ra(ta(Vc(p.index),s.length),0),E=[],b=1;b<p.length;b++)ia(E,void 0===(h=p[b])?h:String(h));var w=p.groups;if(a){var O=na([y],E,m,s);void 0!==w&&ia(O,w);var S=zc(Mc(i,void 0,O))}else S=Qc(y,s,m,E,w,i);m>=v&&(d+=sa(s,v,m)+S,v=m+y.length)}return d+sa(s,v)}]}),!ua||!ca||aa);var la=se,fa=U,pa=gr("match"),ha=function(e){var r;return la(e)&&(void 0!==(r=e[pa])?!!r:"RegExp"==fa(e))},da=I,va=tr,ga=fe,ya=gs,ma=RegExp.prototype,Ea=function(e){var r=e.flags;return void 0!==r||"flags"in ma||va(e,"flags")||!ga(ma,e)?r:da(ya,e)},ba=fi,wa=I,Oa=C,Sa=K,Ra=ne,Ia=H,Aa=ha,ja=ds,xa=Me,ka=Ea,La=Lc,_a=gr("replace"),Ta=TypeError,Na=Oa("".indexOf);Oa("".replace);var Pa=Oa("".slice),Fa=Math.max,Da=function(e,r,t){return t>e.length?-1:""===r?t:Na(e,r,t)};ba({target:"String",proto:!0},{replaceAll:function(e,r){var t,n,i,o,s,c,a,u,l=Sa(this),f=0,p=0,h="";if(!Ia(e)){if(Aa(e)&&(t=ja(Sa(ka(e))),!~Na(t,"g")))throw Ta("`.replaceAll` does not allow non-global regexes");if(n=xa(e,_a))return wa(n,e,l,r)}for(i=ja(l),o=ja(e),(s=Ra(r))||(r=ja(r)),c=o.length,a=Fa(1,c),f=Da(i,o,0);-1!==f;)u=s?ja(r(o,f,i)):La(o,i,f,[],void 0,r),h+=Pa(i,p,f)+u,p=f+c,f=Da(i,o,f+a);return p<i.length&&(h+=Pa(i,p)),h}});var Ca=Br.f,Ma=function(e,r,t){t in e||Ca(e,t,{configurable:!0,get:function(){return r[t]},set:function(e){r[t]=e}})},Ga=ne,$a=se,Ua=Lo,Ba=function(e,r,t){var n,i;return Ua&&Ga(n=r.constructor)&&n!==t&&$a(i=n.prototype)&&i!==t.prototype&&Ua(e,i),e},Xa=ds,qa=se,Wa=it,Va=Error,Ha=C("".replace),za=String(Va("zxcasd").stack),Ya=/\n\s*at [^:]*:[^\n]*/,Ka=Ya.test(za),Ja=T,Qa=!b((function(){var e=Error("a");return!("stack"in e)||(Object.defineProperty(e,"stack",Ja(1,7)),7!==e.stack)})),Za=it,eu=function(e,r){if(Ka&&"string"==typeof e&&!Va.prepareStackTrace)for(;r--;)e=Ha(e,Ya,"");return e},ru=Qa,tu=Error.captureStackTrace,nu=le,iu=tr,ou=it,su=fe,cu=Lo,au=zn,uu=Ma,lu=Ba,fu=function(e,r){return void 0===e?arguments.length<2?"":r:Xa(e)},pu=function(e,r){qa(r)&&"cause"in r&&Wa(e,"cause",r.cause)},hu=function(e,r,t,n){ru&&(tu?tu(e,r):Za(e,"stack",eu(t,n)))},du=w,vu=fi,gu=Zs,yu=function(e,r,t,n){var i="stackTraceLimit",o=n?2:1,s=e.split("."),c=s[s.length-1],a=nu.apply(null,s);if(a){var u=a.prototype;if(iu(u,"cause")&&delete u.cause,!t)return a;var l=nu("Error"),f=r((function(e,r){var t=fu(n?r:e,void 0),i=n?new a(e):new a;return void 0!==t&&ou(i,"message",t),hu(i,f,i.stack,2),this&&su(u,this)&&lu(i,this,f),arguments.length>o&&pu(i,arguments[o]),i}));f.prototype=u,"Error"!==c?cu?cu(f,l):au(f,l,{name:!0}):du&&i in a&&(uu(f,a,i),uu(f,a,"prepareStackTrace")),au(f,a);try{u.name!==c&&ou(u,"name",c),u.constructor=f}catch(e){}return f}},mu=m.WebAssembly,Eu=7!==Error("e",{cause:7}).cause,bu=function(e,r){var t={};t[e]=yu(e,r,Eu),vu({global:!0,constructor:!0,arity:1,forced:Eu},t)},wu=function(e,r){if(mu&&mu[e]){var t={};t[e]=yu("WebAssembly."+e,r,Eu),vu({target:"WebAssembly",stat:!0,constructor:!0,arity:1,forced:Eu},t)}};bu("Error",(function(e){return function(r){return gu(e,this,arguments)}})),bu("EvalError",(function(e){return function(r){return gu(e,this,arguments)}})),bu("RangeError",(function(e){return function(r){return gu(e,this,arguments)}})),bu("ReferenceError",(function(e){return function(r){return gu(e,this,arguments)}})),bu("SyntaxError",(function(e){return function(r){return gu(e,this,arguments)}})),bu("TypeError",(function(e){return function(r){return gu(e,this,arguments)}})),bu("URIError",(function(e){return function(r){return gu(e,this,arguments)}})),wu("CompileError",(function(e){return function(r){return gu(e,this,arguments)}})),wu("LinkError",(function(e){return function(r){return gu(e,this,arguments)}})),wu("RuntimeError",(function(e){return function(r){return gu(e,this,arguments)}}));var Ou=Fe,Su=O,Ru=tc(tc.bind),Iu=U,Au=Array.isArray||function(e){return"Array"==Iu(e)},ju=C,xu=b,ku=ne,Lu=fs,_u=mt,Tu=function(){},Nu=[],Pu=le("Reflect","construct"),Fu=/^\s*(?:class|function)\b/,Du=ju(Fu.exec),Cu=!Fu.exec(Tu),Mu=function(e){if(!ku(e))return!1;try{return Pu(Tu,Nu,e),!0}catch(e){return!1}},Gu=function(e){if(!ku(e))return!1;switch(Lu(e)){case"AsyncFunction":case"GeneratorFunction":case"AsyncGeneratorFunction":return!1}try{return Cu||!!Du(Fu,_u(e))}catch(e){return!0}};Gu.sham=!0;var $u=!Pu||xu((function(){var e;return Mu(Mu.call)||!Mu(Object)||!Mu((function(){e=!0}))||e}))?Gu:Mu,Uu=Au,Bu=$u,Xu=se,qu=gr("species"),Wu=Array,Vu=function(e){var r;return Uu(e)&&(r=e.constructor,(Bu(r)&&(r===Wu||Uu(r.prototype))||Xu(r)&&null===(r=r[qu]))&&(r=void 0)),void 0===r?Wu:r},Hu=function(e,r){return new(Vu(e))(0===r?0:r)},zu=function(e,r){return Ou(e),void 0===r?e:Su?Ru(e,r):function(){return e.apply(r,arguments)}},Yu=V,Ku=Ze,Ju=On,Qu=Hu,Zu=C([].push),el=function(e){var r=1==e,t=2==e,n=3==e,i=4==e,o=6==e,s=7==e,c=5==e||o;return function(a,u,l,f){for(var p,h,d=Ku(a),v=Yu(d),g=zu(u,l),y=Ju(v),m=0,E=f||Qu,b=r?E(a,y):t||s?E(a,0):void 0;y>m;m++)if((c||m in v)&&(h=g(p=v[m],m,d),e))if(r)b[m]=h;else if(h)switch(e){case 3:return!0;case 5:return p;case 6:return m;case 2:Zu(b,p)}else switch(e){case 4:return!1;case 7:Zu(b,p)}return o?-1:n||i?i:b}},rl={forEach:el(0),map:el(1),filter:el(2),some:el(3),every:el(4),find:el(5),findIndex:el(6),filterReject:el(7)},tl=b,nl=Ee,il=gr("species"),ol=function(e){return nl>=51||!tl((function(){var r=[];return(r.constructor={})[il]=function(){return{foo:1}},1!==r[e](Boolean).foo}))},sl=rl.map;fi({target:"Array",proto:!0,forced:!ol("map")},{map:function(e){return sl(this,e,arguments.length>1?arguments[1]:void 0)}});var cl,al,ul=fi,ll=I,fl=ne,pl=Hr,hl=ds,dl=(cl=!1,(al=/[ac]/).exec=function(){return cl=!0,/./.exec.apply(this,arguments)},!0===al.test("abc")&&cl),vl=/./.test;ul({target:"RegExp",proto:!0,forced:!dl},{test:function(e){var r=pl(this),t=hl(e),n=r.exec;if(!fl(n))return ll(vl,r,t);var i=ll(n,r,t);return null!==i&&(pl(i),!0)}});var gl=I,yl=Hr,ml=H,El=bn,bl=ds,wl=K,Ol=Me,Sl=wc,Rl=Cc;fc("match",(function(e,r,t){return[function(r){var t=wl(this),n=ml(r)?void 0:Ol(r,e);return n?gl(n,r,t):new RegExp(r)[e](bl(t))},function(e){var n=yl(this),i=bl(e),o=t(r,n,i);if(o.done)return o.value;if(!n.global)return Rl(n,i);var s=n.unicode;n.lastIndex=0;for(var c,a=[],u=0;null!==(c=Rl(n,i));){var l=bl(c[0]);a[u]=l,""===l&&(n.lastIndex=Sl(i,El(n.lastIndex),s)),u++}return 0===u?null:a}]}));var Il=tn,Al=Br,jl=function(e,r,t){return t.get&&Il(t.get,r,{getter:!0}),t.set&&Il(t.set,r,{setter:!0}),Al.f(e,r,t)},xl=le,kl=jl,Ll=w,_l=gr("species"),Tl=w,Nl=m,Pl=C,Fl=ni,Dl=Ba,Cl=it,Ml=un.f,Gl=fe,$l=ha,Ul=ds,Bl=Ea,Xl=Os,ql=Ma,Wl=an,Vl=b,Hl=tr,zl=Mt.enforce,Yl=function(e){var r=xl(e);Ll&&r&&!r[_l]&&kl(r,_l,{configurable:!0,get:function(){return this}})},Kl=Is,Jl=xs,Ql=gr("match"),Zl=Nl.RegExp,ef=Zl.prototype,rf=Nl.SyntaxError,tf=Pl(ef.exec),nf=Pl("".charAt),of=Pl("".replace),sf=Pl("".indexOf),cf=Pl("".slice),af=/^\?<[^\s\d!#%&*+<=>@^][^\s!#%&*+<=>@^]*>/,uf=/a/g,lf=/a/g,ff=new Zl(uf)!==uf,pf=Xl.MISSED_STICKY,hf=Xl.UNSUPPORTED_Y,df=Tl&&(!ff||pf||Kl||Jl||Vl((function(){return lf[Ql]=!1,Zl(uf)!=uf||Zl(lf)==lf||"/a/i"!=Zl(uf,"i")})));if(Fl("RegExp",df)){for(var vf=function(e,r){var t,n,i,o,s,c,a=Gl(ef,this),u=$l(e),l=void 0===r,f=[],p=e;if(!a&&u&&l&&e.constructor===vf)return e;if((u||Gl(ef,e))&&(e=e.source,l&&(r=Bl(p))),e=void 0===e?"":Ul(e),r=void 0===r?"":Ul(r),p=e,Kl&&"dotAll"in uf&&(n=!!r&&sf(r,"s")>-1)&&(r=of(r,/s/g,"")),t=r,pf&&"sticky"in uf&&(i=!!r&&sf(r,"y")>-1)&&hf&&(r=of(r,/y/g,"")),Jl&&(o=function(e){for(var r,t=e.length,n=0,i="",o=[],s={},c=!1,a=!1,u=0,l="";n<=t;n++){if("\\"===(r=nf(e,n)))r+=nf(e,++n);else if("]"===r)c=!1;else if(!c)switch(!0){case"["===r:c=!0;break;case"("===r:tf(af,cf(e,n+1))&&(n+=2,a=!0),i+=r,u++;continue;case">"===r&&a:if(""===l||Hl(s,l))throw new rf("Invalid capture group name");s[l]=!0,o[o.length]=[l,u],a=!1,l="";continue}a?l+=r:i+=r}return[i,o]}(e),e=o[0],f=o[1]),s=Dl(Zl(e,r),a?this:ef,vf),(n||i||f.length)&&(c=zl(s),n&&(c.dotAll=!0,c.raw=vf(function(e){for(var r,t=e.length,n=0,i="",o=!1;n<=t;n++)"\\"!==(r=nf(e,n))?o||"."!==r?("["===r?o=!0:"]"===r&&(o=!1),i+=r):i+="[\\s\\S]":i+=r+nf(e,++n);return i}(e),t)),i&&(c.sticky=!0),f.length&&(c.groups=f)),e!==p)try{Cl(s,"source",""===p?"(?:)":p)}catch(e){}return s},gf=Ml(Zl),yf=0;gf.length>yf;)ql(vf,Zl,gf[yf++]);ef.constructor=vf,vf.prototype=ef,Wl(Nl,"RegExp",vf,{constructor:!0})}Yl("RegExp");var mf=w,Ef=Is,bf=U,wf=jl,Of=Mt.get,Sf=RegExp.prototype,Rf=TypeError;mf&&Ef&&wf(Sf,"dotAll",{configurable:!0,get:function(){if(this!==Sf){if("RegExp"===bf(this))return!!Of(this).dotAll;throw Rf("Incompatible receiver, RegExp required")}}});var If=K,Af=ds,jf=C("".replace),xf=RegExp("^[\t\n\v\f\r                　\u2028\u2029\ufeff]+"),kf=RegExp("(^|[^\t\n\v\f\r                　\u2028\u2029\ufeff])[\t\n\v\f\r                　\u2028\u2029\ufeff]+$"),Lf=function(e){return function(r){var t=Af(If(r));return 1&e&&(t=jf(t,xf,"")),2&e&&(t=jf(t,kf,"$1")),t}},_f={start:Lf(1),end:Lf(2),trim:Lf(3)},Tf=ft.PROPER,Nf=b,Pf="\t\n\v\f\r                　\u2028\u2029\ufeff",Ff=function(e){return Nf((function(){return!!Pf[e]()||"​᠎"!=="​᠎"[e]()||Tf&&Pf[e].name!==e}))},Df=_f.trim;fi({target:"String",proto:!0,forced:Ff("trim")},{trim:function(){return Df(this)}});var Cf=C([].slice),Mf=Au,Gf=ne,$f=U,Uf=ds,Bf=C([].push),Xf=fi,qf=le,Wf=Zs,Vf=I,Hf=C,zf=b,Yf=ne,Kf=ke,Jf=Cf,Qf=function(e){if(Gf(e))return e;if(Mf(e)){for(var r=e.length,t=[],n=0;n<r;n++){var i=e[n];"string"==typeof i?Bf(t,i):"number"!=typeof i&&"Number"!=$f(i)&&"String"!=$f(i)||Bf(t,Uf(i))}var o=t.length,s=!0;return function(e,r){if(s)return s=!1,r;if(Mf(this))return r;for(var n=0;n<o;n++)if(t[n]===e)return r}}},Zf=Se,ep=String,rp=qf("JSON","stringify"),tp=Hf(/./.exec),np=Hf("".charAt),ip=Hf("".charCodeAt),op=Hf("".replace),sp=Hf(1..toString),cp=/[\uD800-\uDFFF]/g,ap=/^[\uD800-\uDBFF]$/,up=/^[\uDC00-\uDFFF]$/,lp=!Zf||zf((function(){var e=qf("Symbol")();return"[null]"!=rp([e])||"{}"!=rp({a:e})||"{}"!=rp(Object(e))})),fp=zf((function(){return'"\\udf06\\ud834"'!==rp("\udf06\ud834")||'"\\udead"'!==rp("\udead")})),pp=function(e,r){var t=Jf(arguments),n=Qf(r);if(Yf(n)||void 0!==e&&!Kf(e))return t[1]=function(e,r){if(Yf(n)&&(r=Vf(n,this,ep(e),r)),!Kf(r))return r},Wf(rp,null,t)},hp=function(e,r,t){var n=np(t,r-1),i=np(t,r+1);return tp(ap,e)&&!tp(up,i)||tp(up,e)&&!tp(ap,n)?"\\u"+sp(ip(e,0),16):e};rp&&Xf({target:"JSON",stat:!0,arity:3,forced:lp||fp},{stringify:function(e,r,t){var n=Jf(arguments),i=Wf(lp?pp:rp,null,n);return fp&&"string"==typeof i?op(i,cp,hp):i}});var dp=Ar,vp=Br,gp=T,yp=function(e,r,t){var n=dp(r);n in e?vp.f(e,n,gp(0,t)):e[n]=t},mp=fi,Ep=Au,bp=$u,wp=se,Op=yn,Sp=On,Rp=Z,Ip=yp,Ap=gr,jp=Cf,xp=ol("slice"),kp=Ap("species"),Lp=Array,_p=Math.max;mp({target:"Array",proto:!0,forced:!xp},{slice:function(e,r){var t,n,i,o=Rp(this),s=Sp(o),c=Op(e,s),a=Op(void 0===r?s:r,s);if(Ep(o)&&(t=o.constructor,(bp(t)&&(t===Lp||Ep(t.prototype))||wp(t)&&null===(t=t[kp]))&&(t=void 0),t===Lp||void 0===t))return jp(o,c,a);for(n=new(void 0===t?Lp:t)(_p(a-c,0)),i=0;c<a;c++,i++)c in o&&Ip(n,i,o[c]);return n.length=i,n}});var Tp=b,Np=function(e,r){var t=[][e];return!!t&&Tp((function(){t.call(null,r||function(){return 1},1)}))},Pp=fi,Fp=jn.indexOf,Dp=Np,Cp=tc([].indexOf),Mp=!!Cp&&1/Cp([1],1,-0)<0;Pp({target:"Array",proto:!0,forced:Mp||!Dp("indexOf")},{indexOf:function(e){var r=arguments.length>1?arguments[1]:void 0;return Mp?Cp(this,e,r)||0:Fp(this,e,r)}});var Gp=w,$p=Au,Up=TypeError,Bp=Object.getOwnPropertyDescriptor,Xp=Gp&&!function(){if(void 0!==this)return!0;try{Object.defineProperty([],"length",{writable:!1}).length=1}catch(e){return e instanceof TypeError}}()?function(e,r){if($p(e)&&!Bp(e,"length").writable)throw Up("Cannot set read only .length");return e.length=r}:function(e,r){return e.length=r},qp=TypeError,Wp=function(e){if(e>9007199254740991)throw qp("Maximum allowed index exceeded");return e},Vp=Ze,Hp=On,zp=Xp,Yp=Wp;fi({target:"Array",proto:!0,arity:1,forced:b((function(){return 4294967297!==[].push.call({length:4294967296},1)}))||!function(){try{Object.defineProperty([],"length",{writable:!1}).push()}catch(e){return e instanceof TypeError}}()},{push:function(e){var r=Vp(this),t=Hp(r),n=arguments.length;Yp(t+n);for(var i=0;i<n;i++)r[t]=arguments[i],t++;return zp(r,t),t}});var Kp=fi,Jp=b,Qp=Au,Zp=se,eh=Ze,rh=On,th=Wp,nh=yp,ih=Hu,oh=ol,sh=Ee,ch=gr("isConcatSpreadable"),ah=sh>=51||!Jp((function(){var e=[];return e[ch]=!1,e.concat()[0]!==e})),uh=function(e){if(!Zp(e))return!1;var r=e[ch];return void 0!==r?!!r:Qp(e)};Kp({target:"Array",proto:!0,arity:1,forced:!ah||!oh("concat")},{concat:function(e){var r,t,n,i,o,s=eh(this),c=ih(s,0),a=0;for(r=-1,n=arguments.length;r<n;r++)if(uh(o=-1===r?s:arguments[r]))for(i=rh(o),th(a+i),t=0;t<i;t++,a++)t in o&&nh(c,a,o[t]);else th(a+1),nh(c,a++,o);return c.length=a,c}});var lh=_e,fh=TypeError,ph=function(e,r){if(!delete e[r])throw fh("Cannot delete property "+lh(r)+" of "+lh(e))},hh=Ze,dh=On,vh=Xp,gh=ph,yh=Wp;fi({target:"Array",proto:!0,arity:1,forced:1!==[].unshift(0)||!function(){try{Object.defineProperty([],"length",{writable:!1}).unshift()}catch(e){return e instanceof TypeError}}()},{unshift:function(e){var r=hh(this),t=dh(r),n=arguments.length;if(n){yh(t+n);for(var i=t;i--;){var o=i+n;i in r?r[o]=r[i]:gh(r,o)}for(var s=0;s<n;s++)r[s]=arguments[s]}return vh(r,t+n)}});var mh,Eh,bh,wh={},Oh={};function Sh(){if(mh)return Oh;mh=1;var e=l.default,r="win32"===process.platform,t=u.default,n=process.env.NODE_DEBUG&&/fs/.test(process.env.NODE_DEBUG);function i(e){return"function"==typeof e?e:function(){var e;if(n){var r=new Error;e=function(e){e&&(r.message=e.message,t(e=r))}}else e=t;return e;function t(e){if(e){if(process.throwDeprecation)throw e;if(!process.noDeprecation){var r="fs: missing callback "+(e.stack||e.message);process.traceDeprecation?console.trace(r):console.error(r)}}}}()}if(e.normalize,r)var o=/(.*?)(?:[\/\\]+|$)/g;else o=/(.*?)(?:[\/]+|$)/g;if(r)var s=/^(?:[a-zA-Z]:|[\\\/]{2}[^\\\/]+[\\\/][^\\\/]+)?[\\\/]*/;else s=/^[\/]*/;return Oh.realpathSync=function(n,i){if(n=e.resolve(n),i&&Object.prototype.hasOwnProperty.call(i,n))return i[n];var c,a,u,l,f=n,p={},h={};function d(){var e=s.exec(n);c=e[0].length,a=e[0],u=e[0],l="",r&&!h[u]&&(t.lstatSync(u),h[u]=!0)}for(d();c<n.length;){o.lastIndex=c;var v=o.exec(n);if(l=a,a+=v[0],u=l+v[1],c=o.lastIndex,!(h[u]||i&&i[u]===u)){var g;if(i&&Object.prototype.hasOwnProperty.call(i,u))g=i[u];else{var y=t.lstatSync(u);if(!y.isSymbolicLink()){h[u]=!0,i&&(i[u]=u);continue}var m=null;if(!r){var E=y.dev.toString(32)+":"+y.ino.toString(32);p.hasOwnProperty(E)&&(m=p[E])}null===m&&(t.statSync(u),m=t.readlinkSync(u)),g=e.resolve(l,m),i&&(i[u]=g),r||(p[E]=m)}n=e.resolve(g,n.slice(c)),d()}}return i&&(i[f]=n),n},Oh.realpath=function(n,c,a){if("function"!=typeof a&&(a=i(c),c=null),n=e.resolve(n),c&&Object.prototype.hasOwnProperty.call(c,n))return process.nextTick(a.bind(null,null,c[n]));var u,l,f,p,h=n,d={},v={};function g(){var e=s.exec(n);u=e[0].length,l=e[0],f=e[0],p="",r&&!v[f]?t.lstat(f,(function(e){if(e)return a(e);v[f]=!0,y()})):process.nextTick(y)}function y(){if(u>=n.length)return c&&(c[h]=n),a(null,n);o.lastIndex=u;var e=o.exec(n);return p=l,l+=e[0],f=p+e[1],u=o.lastIndex,v[f]||c&&c[f]===f?process.nextTick(y):c&&Object.prototype.hasOwnProperty.call(c,f)?b(c[f]):t.lstat(f,m)}function m(e,n){if(e)return a(e);if(!n.isSymbolicLink())return v[f]=!0,c&&(c[f]=f),process.nextTick(y);if(!r){var i=n.dev.toString(32)+":"+n.ino.toString(32);if(d.hasOwnProperty(i))return E(null,d[i],f)}t.stat(f,(function(e){if(e)return a(e);t.readlink(f,(function(e,t){r||(d[i]=t),E(e,t)}))}))}function E(r,t,n){if(r)return a(r);var i=e.resolve(p,t);c&&(c[n]=i),b(i)}function b(r){n=e.resolve(r,n.slice(u)),g()}g()},Oh}function Rh(){if(bh)return Eh;bh=1,Eh=c,c.realpath=c,c.sync=a,c.realpathSync=a,c.monkeypatch=function(){e.realpath=c,e.realpathSync=a},c.unmonkeypatch=function(){e.realpath=r,e.realpathSync=t};var e=u.default,r=e.realpath,t=e.realpathSync,n=process.version,i=/^v[0-5]\./.test(n),o=Sh();function s(e){return e&&"realpath"===e.syscall&&("ELOOP"===e.code||"ENOMEM"===e.code||"ENAMETOOLONG"===e.code)}function c(e,t,n){if(i)return r(e,t,n);"function"==typeof t&&(n=t,t=null),r(e,t,(function(r,i){s(r)?o.realpath(e,t,n):n(r,i)}))}function a(e,r){if(i)return t(e,r);try{return t(e,r)}catch(t){if(s(t))return o.realpathSync(e,r);throw t}}return Eh}var Ih=rl.filter;fi({target:"Array",proto:!0,forced:!ol("filter")},{filter:function(e){return Ih(this,e,arguments.length>1?arguments[1]:void 0)}});var Ah,jh,xh,kh,Lh,_h,Th,Nh,Ph=$u,Fh=_e,Dh=TypeError,Ch=Hr,Mh=function(e){if(Ph(e))return e;throw Dh(Fh(e)+" is not a constructor")},Gh=H,$h=gr("species"),Uh=yn,Bh=On,Xh=yp,qh=Array,Wh=Math.max,Vh=function(e,r,t){for(var n=Bh(e),i=Uh(r,n),o=Uh(void 0===t?n:t,n),s=qh(Wh(o-i,0)),c=0;i<o;i++,c++)Xh(s,c,e[i]);return s.length=c,s},Hh=Zs,zh=I,Yh=C,Kh=fc,Jh=Hr,Qh=H,Zh=ha,ed=K,rd=function(e,r){var t,n=Ch(e).constructor;return void 0===n||Gh(t=Ch(n)[$h])?r:Mh(t)},td=wc,nd=bn,id=ds,od=Me,sd=Vh,cd=Cc,ad=zs,ud=b,ld=Os.UNSUPPORTED_Y,fd=Math.min,pd=[].push,hd=Yh(/./.exec),dd=Yh(pd),vd=Yh("".slice),gd=!ud((function(){var e=/(?:)/,r=e.exec;e.exec=function(){return r.apply(this,arguments)};var t="ab".split(e);return 2!==t.length||"a"!==t[0]||"b"!==t[1]}));function yd(){if(jh)return Ah;jh=1,Ah=function(r,t){for(var n=[],i=0;i<r.length;i++){var o=t(r[i],i);e(o)?n.push.apply(n,o):n.push(o)}return n};var e=Array.isArray||function(e){return"[object Array]"===Object.prototype.toString.call(e)};return Ah}function md(){if(_h)return Lh;_h=1;var e=yd(),r=function(){if(kh)return xh;function e(e,n,i){e instanceof RegExp&&(e=r(e,i)),n instanceof RegExp&&(n=r(n,i));var o=t(e,n,i);return o&&{start:o[0],end:o[1],pre:i.slice(0,o[0]),body:i.slice(o[0]+e.length,o[1]),post:i.slice(o[1]+n.length)}}function r(e,r){var t=r.match(e);return t?t[0]:null}function t(e,r,t){var n,i,o,s,c,a=t.indexOf(e),u=t.indexOf(r,a+1),l=a;if(a>=0&&u>0){for(n=[],o=t.length;l>=0&&!c;)l==a?(n.push(l),a=t.indexOf(e,l+1)):1==n.length?c=[n.pop(),u]:((i=n.pop())<o&&(o=i,s=u),u=t.indexOf(r,l+1)),l=a<u&&a>=0?a:u;n.length&&(c=[o,s])}return c}return kh=1,xh=e,e.range=t,xh}();Lh=function(e){if(!e)return[];"{}"===e.substr(0,2)&&(e="\\{\\}"+e.substr(2));return d(function(e){return e.split("\\\\").join(t).split("\\{").join(n).split("\\}").join(i).split("\\,").join(o).split("\\.").join(s)}(e),!0).map(a)};var t="\0SLASH"+Math.random()+"\0",n="\0OPEN"+Math.random()+"\0",i="\0CLOSE"+Math.random()+"\0",o="\0COMMA"+Math.random()+"\0",s="\0PERIOD"+Math.random()+"\0";function c(e){return parseInt(e,10)==e?parseInt(e,10):e.charCodeAt(0)}function a(e){return e.split(t).join("\\").split(n).join("{").split(i).join("}").split(o).join(",").split(s).join(".")}function u(e){if(!e)return[""];var t=[],n=r("{","}",e);if(!n)return e.split(",");var i=n.pre,o=n.body,s=n.post,c=i.split(",");c[c.length-1]+="{"+o+"}";var a=u(s);return s.length&&(c[c.length-1]+=a.shift(),c.push.apply(c,a)),t.push.apply(t,c),t}function l(e){return"{"+e+"}"}function f(e){return/^-?0\d/.test(e)}function p(e,r){return e<=r}function h(e,r){return e>=r}function d(t,n){var o=[],s=r("{","}",t);if(!s||/\$$/.test(s.pre))return[t];var a,v=/^-?\d+\.\.-?\d+(?:\.\.-?\d+)?$/.test(s.body),g=/^[a-zA-Z]\.\.[a-zA-Z](?:\.\.-?\d+)?$/.test(s.body),y=v||g,m=s.body.indexOf(",")>=0;if(!y&&!m)return s.post.match(/,.*\}/)?d(t=s.pre+"{"+s.body+i+s.post):[t];if(y)a=s.body.split(/\.\./);else if(1===(a=u(s.body)).length&&1===(a=d(a[0],!1).map(l)).length)return(w=s.post.length?d(s.post,!1):[""]).map((function(e){return s.pre+a[0]+e}));var E,b=s.pre,w=s.post.length?d(s.post,!1):[""];if(y){var O=c(a[0]),S=c(a[1]),R=Math.max(a[0].length,a[1].length),I=3==a.length?Math.abs(c(a[2])):1,A=p;S<O&&(I*=-1,A=h);var j=a.some(f);E=[];for(var x=O;A(x,S);x+=I){var k;if(g)"\\"===(k=String.fromCharCode(x))&&(k="");else if(k=String(x),j){var L=R-k.length;if(L>0){var _=new Array(L+1).join("0");k=x<0?"-"+_+k.slice(1):_+k}}E.push(k)}}else E=e(a,(function(e){return d(e,!1)}));for(var T=0;T<E.length;T++)for(var N=0;N<w.length;N++){var P=b+E[T]+w[N];(!n||y||P)&&o.push(P)}return o}return Lh}function Ed(){if(Nh)return Th;Nh=1,Th=u,u.Minimatch=l;var e=function(){try{return require("path")}catch(e){}}()||{sep:"/"};u.sep=e.sep;var r=u.GLOBSTAR=l.GLOBSTAR={},t=md(),n={"!":{open:"(?:(?!(?:",close:"))[^/]*?)"},"?":{open:"(?:",close:")?"},"+":{open:"(?:",close:")+"},"*":{open:"(?:",close:")*"},"@":{open:"(?:",close:")"}},i="[^/]",o="[^/]*?",s="().*{}+?[]^$\\!".split("").reduce((function(e,r){return e[r]=!0,e}),{});var c=/\/+/;function a(e,r){r=r||{};var t={};return Object.keys(e).forEach((function(r){t[r]=e[r]})),Object.keys(r).forEach((function(e){t[e]=r[e]})),t}function u(e,r,t){return p(r),t||(t={}),!(!t.nocomment&&"#"===r.charAt(0))&&new l(r,t).match(e)}function l(r,t){if(!(this instanceof l))return new l(r,t);p(r),t||(t={}),r=r.trim(),t.allowWindowsEscape||"/"===e.sep||(r=r.split(e.sep).join("/")),this.options=t,this.set=[],this.pattern=r,this.regexp=null,this.negate=!1,this.comment=!1,this.empty=!1,this.partial=!!t.partial,this.make()}function f(e,r){return r||(r=this instanceof l?this.options:{}),e=void 0===e?this.pattern:e,p(e),r.nobrace||!/\{(?:(?!\{).)*\}/.test(e)?[e]:t(e)}u.filter=function(e,r){return r=r||{},function(t,n,i){return u(t,e,r)}},u.defaults=function(e){if(!e||"object"!=typeof e||!Object.keys(e).length)return u;var r=u,t=function(t,n,i){return r(t,n,a(e,i))};return(t.Minimatch=function(t,n){return new r.Minimatch(t,a(e,n))}).defaults=function(t){return r.defaults(a(e,t)).Minimatch},t.filter=function(t,n){return r.filter(t,a(e,n))},t.defaults=function(t){return r.defaults(a(e,t))},t.makeRe=function(t,n){return r.makeRe(t,a(e,n))},t.braceExpand=function(t,n){return r.braceExpand(t,a(e,n))},t.match=function(t,n,i){return r.match(t,n,a(e,i))},t},l.defaults=function(e){return u.defaults(e).Minimatch},l.prototype.debug=function(){},l.prototype.make=function(){var e=this.pattern,r=this.options;if(!r.nocomment&&"#"===e.charAt(0))return void(this.comment=!0);if(!e)return void(this.empty=!0);this.parseNegate();var t=this.globSet=this.braceExpand();r.debug&&(this.debug=function(){console.error.apply(console,arguments)});this.debug(this.pattern,t),t=this.globParts=t.map((function(e){return e.split(c)})),this.debug(this.pattern,t),t=t.map((function(e,r,t){return e.map(this.parse,this)}),this),this.debug(this.pattern,t),t=t.filter((function(e){return-1===e.indexOf(!1)})),this.debug(this.pattern,t),this.set=t},l.prototype.parseNegate=function(){var e=this.pattern,r=!1,t=this.options,n=0;if(t.nonegate)return;for(var i=0,o=e.length;i<o&&"!"===e.charAt(i);i++)r=!r,n++;n&&(this.pattern=e.substr(n));this.negate=r},u.braceExpand=function(e,r){return f(e,r)},l.prototype.braceExpand=f;var p=function(e){if("string"!=typeof e)throw new TypeError("invalid pattern");if(e.length>65536)throw new TypeError("pattern is too long")};l.prototype.parse=function(e,t){p(e);var c=this.options;if("**"===e){if(!c.noglobstar)return r;e="*"}if(""===e)return"";var a,u="",l=!!c.nocase,f=!1,d=[],v=[],g=!1,y=-1,m=-1,E="."===e.charAt(0)?"":c.dot?"(?!(?:^|\\/)\\.{1,2}(?:$|\\/))":"(?!\\.)",b=this;function w(){if(a){switch(a){case"*":u+=o,l=!0;break;case"?":u+=i,l=!0;break;default:u+="\\"+a}b.debug("clearStateChar %j %j",a,u),a=!1}}for(var O,S=0,R=e.length;S<R&&(O=e.charAt(S));S++)if(this.debug("%s\t%s %s %j",e,S,u,O),f&&s[O])u+="\\"+O,f=!1;else switch(O){case"/":return!1;case"\\":w(),f=!0;continue;case"?":case"*":case"+":case"@":case"!":if(this.debug("%s\t%s %s %j <-- stateChar",e,S,u,O),g){this.debug("  in class"),"!"===O&&S===m+1&&(O="^"),u+=O;continue}b.debug("call clearStateChar %j",a),w(),a=O,c.noext&&w();continue;case"(":if(g){u+="(";continue}if(!a){u+="\\(";continue}d.push({type:a,start:S-1,reStart:u.length,open:n[a].open,close:n[a].close}),u+="!"===a?"(?:(?!(?:":"(?:",this.debug("plType %j %j",a,u),a=!1;continue;case")":if(g||!d.length){u+="\\)";continue}w(),l=!0;var I=d.pop();u+=I.close,"!"===I.type&&v.push(I),I.reEnd=u.length;continue;case"|":if(g||!d.length||f){u+="\\|",f=!1;continue}w(),u+="|";continue;case"[":if(w(),g){u+="\\"+O;continue}g=!0,m=S,y=u.length,u+=O;continue;case"]":if(S===m+1||!g){u+="\\"+O,f=!1;continue}var A=e.substring(m+1,S);try{RegExp("["+A+"]")}catch(e){var j=this.parse(A,h);u=u.substr(0,y)+"\\["+j[0]+"\\]",l=l||j[1],g=!1;continue}l=!0,g=!1,u+=O;continue;default:w(),f?f=!1:!s[O]||"^"===O&&g||(u+="\\"),u+=O}g&&(A=e.substr(m+1),j=this.parse(A,h),u=u.substr(0,y)+"\\["+j[0],l=l||j[1]);for(I=d.pop();I;I=d.pop()){var x=u.slice(I.reStart+I.open.length);this.debug("setting tail",u,I),x=x.replace(/((?:\\{2}){0,64})(\\?)\|/g,(function(e,r,t){return t||(t="\\"),r+r+t+"|"})),this.debug("tail=%j\n   %s",x,x,I,u);var k="*"===I.type?o:"?"===I.type?i:"\\"+I.type;l=!0,u=u.slice(0,I.reStart)+k+"\\("+x}w(),f&&(u+="\\\\");var L=!1;switch(u.charAt(0)){case"[":case".":case"(":L=!0}for(var _=v.length-1;_>-1;_--){var T=v[_],N=u.slice(0,T.reStart),P=u.slice(T.reStart,T.reEnd-8),F=u.slice(T.reEnd-8,T.reEnd),D=u.slice(T.reEnd);F+=D;var C=N.split("(").length-1,M=D;for(S=0;S<C;S++)M=M.replace(/\)[+*?]?/,"");var G="";""===(D=M)&&t!==h&&(G="$"),u=N+P+D+G+F}""!==u&&l&&(u="(?=.)"+u);L&&(u=E+u);if(t===h)return[u,l];if(!l)return function(e){return e.replace(/\\(.)/g,"$1")}(e);var $=c.nocase?"i":"";try{var U=new RegExp("^"+u+"$",$)}catch(e){return new RegExp("$.")}return U._glob=e,U._src=u,U};var h={};return u.makeRe=function(e,r){return new l(e,r||{}).makeRe()},l.prototype.makeRe=function(){if(this.regexp||!1===this.regexp)return this.regexp;var e=this.set;if(!e.length)return this.regexp=!1,this.regexp;var t=this.options,n=t.noglobstar?o:t.dot?"(?:(?!(?:\\/|^)(?:\\.{1,2})($|\\/)).)*?":"(?:(?!(?:\\/|^)\\.).)*?",i=t.nocase?"i":"",s=e.map((function(e){return e.map((function(e){return e===r?n:"string"==typeof e?function(e){return e.replace(/[-[\]{}()*+?.,\\^$|#\s]/g,"\\$&")}(e):e._src})).join("\\/")})).join("|");s="^(?:"+s+")$",this.negate&&(s="^(?!"+s+").*$");try{this.regexp=new RegExp(s,i)}catch(e){this.regexp=!1}return this.regexp},u.match=function(e,r,t){var n=new l(r,t=t||{});return e=e.filter((function(e){return n.match(e)})),n.options.nonull&&!e.length&&e.push(r),e},l.prototype.match=function(r,t){if(void 0===t&&(t=this.partial),this.debug("match",r,this.pattern),this.comment)return!1;if(this.empty)return""===r;if("/"===r&&t)return!0;var n=this.options;"/"!==e.sep&&(r=r.split(e.sep).join("/")),r=r.split(c),this.debug(this.pattern,"split",r);var i,o,s=this.set;for(this.debug(this.pattern,"set",s),o=r.length-1;o>=0&&!(i=r[o]);o--);for(o=0;o<s.length;o++){var a=s[o],u=r;if(n.matchBase&&1===a.length&&(u=[i]),this.matchOne(u,a,t))return!!n.flipNegate||!this.negate}return!n.flipNegate&&this.negate},l.prototype.matchOne=function(e,t,n){var i=this.options;this.debug("matchOne",{this:this,file:e,pattern:t}),this.debug("matchOne",e.length,t.length);for(var o=0,s=0,c=e.length,a=t.length;o<c&&s<a;o++,s++){this.debug("matchOne loop");var u,l=t[s],f=e[o];if(this.debug(t,l,f),!1===l)return!1;if(l===r){this.debug("GLOBSTAR",[t,l,f]);var p=o,h=s+1;if(h===a){for(this.debug("** at the end");o<c;o++)if("."===e[o]||".."===e[o]||!i.dot&&"."===e[o].charAt(0))return!1;return!0}for(;p<c;){var d=e[p];if(this.debug("\nglobstar while",e,p,t,h,d),this.matchOne(e.slice(p),t.slice(h),n))return this.debug("globstar found match!",p,c,d),!0;if("."===d||".."===d||!i.dot&&"."===d.charAt(0)){this.debug("dot detected!",e,p,t,h);break}this.debug("globstar swallow a segment, and continue"),p++}return!(!n||(this.debug("\n>>> no match, partial?",e,p,t,h),p!==c))}if("string"==typeof l?(u=f===l,this.debug("string match",l,f,u)):(u=f.match(l),this.debug("pattern match",l,f,u)),!u)return!1}if(o===c&&s===a)return!0;if(o===c)return n;if(s===a)return o===c-1&&""===e[o];throw new Error("wtf?")},Th}Kh("split",(function(e,r,t){var n;return n="c"=="abbc".split(/(b)*/)[1]||4!="test".split(/(?:)/,-1).length||2!="ab".split(/(?:ab)*/).length||4!=".".split(/(.?)(.?)/).length||".".split(/()()/).length>1||"".split(/.?/).length?function(e,t){var n=id(ed(this)),i=void 0===t?4294967295:t>>>0;if(0===i)return[];if(void 0===e)return[n];if(!Zh(e))return zh(r,n,e,i);for(var o,s,c,a=[],u=(e.ignoreCase?"i":"")+(e.multiline?"m":"")+(e.unicode?"u":"")+(e.sticky?"y":""),l=0,f=new RegExp(e.source,u+"g");(o=zh(ad,f,n))&&!((s=f.lastIndex)>l&&(dd(a,vd(n,l,o.index)),o.length>1&&o.index<n.length&&Hh(pd,a,sd(o,1)),c=o[0].length,l=s,a.length>=i));)f.lastIndex===o.index&&f.lastIndex++;return l===n.length?!c&&hd(f,"")||dd(a,""):dd(a,vd(n,l)),a.length>i?sd(a,0,i):a}:"0".split(void 0,0).length?function(e,t){return void 0===e&&0===t?[]:zh(r,this,e,t)}:r,[function(r,t){var i=ed(this),o=Qh(r)?void 0:od(r,e);return o?zh(o,r,i,t):zh(n,id(i),r,t)},function(e,i){var o=Jh(this),s=id(e),c=t(n,o,s,i,n!==r);if(c.done)return c.value;var a=rd(o,RegExp),u=o.unicode,l=(o.ignoreCase?"i":"")+(o.multiline?"m":"")+(o.unicode?"u":"")+(ld?"g":"y"),f=new a(ld?"^(?:"+o.source+")":o,l),p=void 0===i?4294967295:i>>>0;if(0===p)return[];if(0===s.length)return null===cd(f,s)?[s]:[];for(var h=0,d=0,v=[];d<s.length;){f.lastIndex=ld?0:d;var g,y=cd(f,ld?vd(s,d):s);if(null===y||(g=fd(nd(f.lastIndex+(ld?d:0)),s.length))===h)d=td(s,d,u);else{if(dd(v,vd(s,h,d)),v.length===p)return v;for(var m=1;m<=y.length-1;m++)if(dd(v,y[m]),v.length===p)return v;d=h=g}}return dd(v,vd(s,h)),v}]}),!gd,ld);var bd,wd,Od={exports:{}},Sd={exports:{}};function Rd(){if(wd)return Od.exports;wd=1;try{var e=require("util");if("function"!=typeof e.inherits)throw"";Od.exports=e.inherits}catch(e){Od.exports=(bd||(bd=1,"function"==typeof Object.create?Sd.exports=function(e,r){r&&(e.super_=r,e.prototype=Object.create(r.prototype,{constructor:{value:e,enumerable:!1,writable:!0,configurable:!0}}))}:Sd.exports=function(e,r){if(r){e.super_=r;var t=function(){};t.prototype=r.prototype,e.prototype=new t,e.prototype.constructor=e}}),Sd.exports)}return Od.exports}var Id,Ad={exports:{}};function jd(){if(Id)return Ad.exports;function e(e){return"/"===e.charAt(0)}function r(e){var r=/^([a-zA-Z]:|[\\\/]{2}[^\\\/]+[\\\/]+[^\\\/]+)?([\\\/])?([\s\S]*?)$/.exec(e),t=r[1]||"",n=Boolean(t&&":"!==t.charAt(1));return Boolean(r[2]||n)}return Id=1,Ad.exports="win32"===process.platform?r:e,Ad.exports.posix=e,Ad.exports.win32=r,Ad.exports}var xd=Vh,kd=Math.floor,Ld=function(e,r){var t=e.length,n=kd(t/2);return t<8?_d(e,r):Td(e,Ld(xd(e,0,n),r),Ld(xd(e,n),r),r)},_d=function(e,r){for(var t,n,i=e.length,o=1;o<i;){for(n=o,t=e[o];n&&r(e[n-1],t)>0;)e[n]=e[--n];n!==o++&&(e[n]=t)}return e},Td=function(e,r,t,n){for(var i=r.length,o=t.length,s=0,c=0;s<i||c<o;)e[s+c]=s<i&&c<o?n(r[s],t[c])<=0?r[s++]:t[c++]:s<i?r[s++]:t[c++];return e},Nd=Ld,Pd=pe.match(/firefox\/(\d+)/i),Fd=!!Pd&&+Pd[1],Dd=/MSIE|Trident/.test(pe),Cd=pe.match(/AppleWebKit\/(\d+)\./),Md=!!Cd&&+Cd[1],Gd=fi,$d=C,Ud=Fe,Bd=Ze,Xd=On,qd=ph,Wd=ds,Vd=b,Hd=Nd,zd=Np,Yd=Fd,Kd=Dd,Jd=Ee,Qd=Md,Zd=[],ev=$d(Zd.sort),rv=$d(Zd.push),tv=Vd((function(){Zd.sort(void 0)})),nv=Vd((function(){Zd.sort(null)})),iv=zd("sort"),ov=!Vd((function(){if(Jd)return Jd<70;if(!(Yd&&Yd>3)){if(Kd)return!0;if(Qd)return Qd<603;var e,r,t,n,i="";for(e=65;e<76;e++){switch(r=String.fromCharCode(e),e){case 66:case 69:case 70:case 72:t=3;break;case 68:case 71:t=4;break;default:t=2}for(n=0;n<47;n++)Zd.push({k:r+n,v:t})}for(Zd.sort((function(e,r){return r.v-e.v})),n=0;n<Zd.length;n++)r=Zd[n].k.charAt(0),i.charAt(i.length-1)!==r&&(i+=r);return"DGBEFHACIJK"!==i}}));Gd({target:"Array",proto:!0,forced:tv||!nv||!iv||!ov},{sort:function(e){void 0!==e&&Ud(e);var r=Bd(this);if(ov)return void 0===e?ev(r):ev(r,e);var t,n,i=[],o=Xd(r);for(n=0;n<o;n++)n in r&&rv(i,r[n]);for(Hd(i,function(e){return function(r,t){return void 0===t?-1:void 0===r?1:void 0!==e?+e(r,t)||0:Wd(r)>Wd(t)?1:-1}}(e)),t=Xd(i),n=0;n<t;)r[n]=i[n++];for(;n<o;)qd(r,n++);return r}});var sv,cv,av,uv={};function lv(){if(sv)return uv;function e(e,r){return Object.prototype.hasOwnProperty.call(e,r)}sv=1,uv.setopts=function(n,s,u){u||(u={});if(u.matchBase&&-1===s.indexOf("/")){if(u.noglobstar)throw new Error("base matching requires globstar");s="**/"+s}n.silent=!!u.silent,n.pattern=s,n.strict=!1!==u.strict,n.realpath=!!u.realpath,n.realpathCache=u.realpathCache||Object.create(null),n.follow=!!u.follow,n.dot=!!u.dot,n.mark=!!u.mark,n.nodir=!!u.nodir,n.nodir&&(n.mark=!0);n.sync=!!u.sync,n.nounique=!!u.nounique,n.nonull=!!u.nonull,n.nosort=!!u.nosort,n.nocase=!!u.nocase,n.stat=!!u.stat,n.noprocess=!!u.noprocess,n.absolute=!!u.absolute,n.fs=u.fs||r,n.maxLength=u.maxLength||1/0,n.cache=u.cache||Object.create(null),n.statCache=u.statCache||Object.create(null),n.symlinks=u.symlinks||Object.create(null),function(e,r){e.ignore=r.ignore||[],Array.isArray(e.ignore)||(e.ignore=[e.ignore]);e.ignore.length&&(e.ignore=e.ignore.map(c))}(n,u),n.changedCwd=!1;var l=process.cwd();e(u,"cwd")?(n.cwd=t.resolve(u.cwd),n.changedCwd=n.cwd!==l):n.cwd=l;n.root=u.root||t.resolve(n.cwd,"/"),n.root=t.resolve(n.root),"win32"===process.platform&&(n.root=n.root.replace(/\\/g,"/"));n.cwdAbs=i(n.cwd)?n.cwd:a(n,n.cwd),"win32"===process.platform&&(n.cwdAbs=n.cwdAbs.replace(/\\/g,"/"));n.nomount=!!u.nomount,u.nonegate=!0,u.nocomment=!0,u.allowWindowsEscape=!1,n.minimatch=new o(s,u),n.options=n.minimatch.options},uv.ownProp=e,uv.makeAbs=a,uv.finish=function(e){for(var r=e.nounique,t=r?[]:Object.create(null),n=0,i=e.matches.length;n<i;n++){var o=e.matches[n];if(o&&0!==Object.keys(o).length){var c=Object.keys(o);r?t.push.apply(t,c):c.forEach((function(e){t[e]=!0}))}else if(e.nonull){var u=e.minimatch.globSet[n];r?t.push(u):t[u]=!0}}r||(t=Object.keys(t));e.nosort||(t=t.sort(s));if(e.mark){for(n=0;n<t.length;n++)t[n]=e._mark(t[n]);e.nodir&&(t=t.filter((function(r){var t=!/\/$/.test(r),n=e.cache[r]||e.cache[a(e,r)];return t&&n&&(t="DIR"!==n&&!Array.isArray(n)),t})))}e.ignore.length&&(t=t.filter((function(r){return!f(e,r)})));e.found=t},uv.mark=function(e,r){var t=a(e,r),n=e.cache[t],i=r;if(n){var o="DIR"===n||Array.isArray(n),s="/"===r.slice(-1);if(o&&!s?i+="/":!o&&s&&(i=i.slice(0,-1)),i!==r){var c=a(e,i);e.statCache[c]=e.statCache[t],e.cache[c]=e.cache[t]}}return i},uv.isIgnored=f,uv.childrenIgnored=function(e,r){return!!e.ignore.length&&e.ignore.some((function(e){return!(!e.gmatcher||!e.gmatcher.match(r))}))};var r=u.default,t=l.default,n=Ed(),i=jd(),o=n.Minimatch;function s(e,r){return e.localeCompare(r,"en")}function c(e){var r=null;if("/**"===e.slice(-3)){var t=e.replace(/(\/\*\*)+$/,"");r=new o(t,{dot:!0})}return{matcher:new o(e,{dot:!0}),gmatcher:r}}function a(e,r){var n=r;return n="/"===r.charAt(0)?t.join(e.root,r):i(r)||""===r?r:e.changedCwd?t.resolve(e.cwd,r):t.resolve(r),"win32"===process.platform&&(n=n.replace(/\\/g,"/")),n}function f(e,r){return!!e.ignore.length&&e.ignore.some((function(e){return e.matcher.match(r)||!(!e.gmatcher||!e.gmatcher.match(r))}))}return uv}var fv,pv,hv=fi,dv=Ze,vv=yn,gv=hn,yv=On,mv=Xp,Ev=Wp,bv=Hu,wv=yp,Ov=ph,Sv=ol("splice"),Rv=Math.max,Iv=Math.min;function Av(){if(pv)return fv;return pv=1,fv=function e(r,t){if(r&&t)return e(r)(t);if("function"!=typeof r)throw new TypeError("need wrapper function");return Object.keys(r).forEach((function(e){n[e]=r[e]})),n;function n(){for(var e=new Array(arguments.length),t=0;t<e.length;t++)e[t]=arguments[t];var n=r.apply(this,e),i=e[e.length-1];return"function"==typeof n&&n!==i&&Object.keys(i).forEach((function(e){n[e]=i[e]})),n}},fv}hv({target:"Array",proto:!0,forced:!Sv},{splice:function(e,r){var t,n,i,o,s,c,a=dv(this),u=yv(a),l=vv(e,u),f=arguments.length;for(0===f?t=n=0:1===f?(t=0,n=u-l):(t=f-2,n=Iv(Rv(gv(r),0),u-l)),Ev(u+t-n),i=bv(a,n),o=0;o<n;o++)(s=l+o)in a&&wv(i,o,a[s]);if(i.length=n,t<n){for(o=l;o<u-n;o++)c=o+t,(s=o+n)in a?a[c]=a[s]:Ov(a,c);for(o=u;o>u-n+t;o--)Ov(a,o-1)}else if(t>n)for(o=u-n;o>l;o--)c=o+t-1,(s=o+n-1)in a?a[c]=a[s]:Ov(a,c);for(o=0;o<t;o++)a[o+l]=arguments[o+2];return mv(a,u-n+t),i}});var jv,xv,kv,Lv,_v,Tv,Nv,Pv,Fv,Dv,Cv,Mv,Gv,$v,Uv={exports:{}};function Bv(){if(jv)return Uv.exports;jv=1;var e=Av();function r(e){var r=function r(){return r.called?r.value:(r.called=!0,r.value=e.apply(this,arguments))};return r.called=!1,r}function t(e){var r=function r(){if(r.called)throw new Error(r.onceError);return r.called=!0,r.value=e.apply(this,arguments)},t=e.name||"Function wrapped with `once`";return r.onceError=t+" shouldn't be called more than once",r.called=!1,r}return Uv.exports=e(r),Uv.exports.strict=e(t),r.proto=r((function(){Object.defineProperty(Function.prototype,"once",{value:function(){return r(this)},configurable:!0}),Object.defineProperty(Function.prototype,"onceStrict",{value:function(){return t(this)},configurable:!0})})),Uv.exports}function Xv(){if(kv)return xv;kv=1;var e=Av(),r=Object.create(null),t=Bv();function n(e){for(var r=e.length,t=[],n=0;n<r;n++)t[n]=e[n];return t}return xv=e((function(e,i){return r[e]?(r[e].push(i),null):(r[e]=[i],function(e){return t((function t(){var i=r[e],o=i.length,s=n(arguments);try{for(var c=0;c<o;c++)i[c].apply(null,s)}finally{i.length>o?(i.splice(0,o),process.nextTick((function(){t.apply(null,s)}))):delete r[e]}}))}(e))})),xv}function qv(){if(_v)return Lv;_v=1,Lv=m;var e=Rh(),r=Ed();r.Minimatch;var t=Rd(),n=f.default.EventEmitter,i=l.default,o=p.default,s=jd(),c=function(){if(av)return cv;av=1,cv=f,f.GlobSync=h;var e=Rh(),r=Ed();r.Minimatch,qv().Glob;var t=l.default,n=p.default,i=jd(),o=lv(),s=o.setopts,c=o.ownProp,a=o.childrenIgnored,u=o.isIgnored;function f(e,r){if("function"==typeof r||3===arguments.length)throw new TypeError("callback provided to sync glob\nSee: https://github.com/isaacs/node-glob/issues/167");return new h(e,r).found}function h(e,r){if(!e)throw new Error("must provide pattern");if("function"==typeof r||3===arguments.length)throw new TypeError("callback provided to sync glob\nSee: https://github.com/isaacs/node-glob/issues/167");if(!(this instanceof h))return new h(e,r);if(s(this,e,r),this.noprocess)return this;var t=this.minimatch.set.length;this.matches=new Array(t);for(var n=0;n<t;n++)this._process(this.minimatch.set[n],n,!1);this._finish()}return h.prototype._finish=function(){if(n.ok(this instanceof h),this.realpath){var r=this;this.matches.forEach((function(t,n){var i=r.matches[n]=Object.create(null);for(var o in t)try{o=r._makeAbs(o),i[e.realpathSync(o,r.realpathCache)]=!0}catch(e){if("stat"!==e.syscall)throw e;i[r._makeAbs(o)]=!0}}))}o.finish(this)},h.prototype._process=function(e,t,o){n.ok(this instanceof h);for(var s,c=0;"string"==typeof e[c];)c++;switch(c){case e.length:return void this._processSimple(e.join("/"),t);case 0:s=null;break;default:s=e.slice(0,c).join("/")}var u,l=e.slice(c);null===s?u=".":i(s)||i(e.map((function(e){return"string"==typeof e?e:"[*]"})).join("/"))?(s&&i(s)||(s="/"+s),u=s):u=s;var f=this._makeAbs(u);a(this,u)||(l[0]===r.GLOBSTAR?this._processGlobStar(s,u,f,l,t,o):this._processReaddir(s,u,f,l,t,o))},h.prototype._processReaddir=function(e,r,n,i,o,s){var c=this._readdir(n,s);if(c){for(var a=i[0],u=!!this.minimatch.negate,l=a._glob,f=this.dot||"."===l.charAt(0),p=[],h=0;h<c.length;h++)("."!==(g=c[h]).charAt(0)||f)&&(u&&!e?!g.match(a):g.match(a))&&p.push(g);var d=p.length;if(0!==d)if(1!==i.length||this.mark||this.stat)for(i.shift(),h=0;h<d;h++){var v;g=p[h],v=e?[e,g]:[g],this._process(v.concat(i),o,s)}else{this.matches[o]||(this.matches[o]=Object.create(null));for(h=0;h<d;h++){var g=p[h];e&&(g="/"!==e.slice(-1)?e+"/"+g:e+g),"/"!==g.charAt(0)||this.nomount||(g=t.join(this.root,g)),this._emitMatch(o,g)}}}},h.prototype._emitMatch=function(e,r){if(!u(this,r)){var t=this._makeAbs(r);if(this.mark&&(r=this._mark(r)),this.absolute&&(r=t),!this.matches[e][r]){if(this.nodir){var n=this.cache[t];if("DIR"===n||Array.isArray(n))return}this.matches[e][r]=!0,this.stat&&this._stat(r)}}},h.prototype._readdirInGlobStar=function(e){if(this.follow)return this._readdir(e,!1);var r,t;try{t=this.fs.lstatSync(e)}catch(e){if("ENOENT"===e.code)return null}var n=t&&t.isSymbolicLink();return this.symlinks[e]=n,n||!t||t.isDirectory()?r=this._readdir(e,!1):this.cache[e]="FILE",r},h.prototype._readdir=function(e,r){if(r&&!c(this.symlinks,e))return this._readdirInGlobStar(e);if(c(this.cache,e)){var t=this.cache[e];if(!t||"FILE"===t)return null;if(Array.isArray(t))return t}try{return this._readdirEntries(e,this.fs.readdirSync(e))}catch(r){return this._readdirError(e,r),null}},h.prototype._readdirEntries=function(e,r){if(!this.mark&&!this.stat)for(var t=0;t<r.length;t++){var n=r[t];n="/"===e?e+n:e+"/"+n,this.cache[n]=!0}return this.cache[e]=r,r},h.prototype._readdirError=function(e,r){switch(r.code){case"ENOTSUP":case"ENOTDIR":var t=this._makeAbs(e);if(this.cache[t]="FILE",t===this.cwdAbs){var n=new Error(r.code+" invalid cwd "+this.cwd);throw n.path=this.cwd,n.code=r.code,n}break;case"ENOENT":case"ELOOP":case"ENAMETOOLONG":case"UNKNOWN":this.cache[this._makeAbs(e)]=!1;break;default:if(this.cache[this._makeAbs(e)]=!1,this.strict)throw r;this.silent||console.error("glob error",r)}},h.prototype._processGlobStar=function(e,r,t,n,i,o){var s=this._readdir(t,o);if(s){var c=n.slice(1),a=e?[e]:[],u=a.concat(c);this._process(u,i,!1);var l=s.length;if(!this.symlinks[t]||!o)for(var f=0;f<l;f++)if("."!==s[f].charAt(0)||this.dot){var p=a.concat(s[f],c);this._process(p,i,!0);var h=a.concat(s[f],n);this._process(h,i,!0)}}},h.prototype._processSimple=function(e,r){var n=this._stat(e);if(this.matches[r]||(this.matches[r]=Object.create(null)),n){if(e&&i(e)&&!this.nomount){var o=/[\/\\]$/.test(e);"/"===e.charAt(0)?e=t.join(this.root,e):(e=t.resolve(this.root,e),o&&(e+="/"))}"win32"===process.platform&&(e=e.replace(/\\/g,"/")),this._emitMatch(r,e)}},h.prototype._stat=function(e){var r=this._makeAbs(e),t="/"===e.slice(-1);if(e.length>this.maxLength)return!1;if(!this.stat&&c(this.cache,r)){var n=this.cache[r];if(Array.isArray(n)&&(n="DIR"),!t||"DIR"===n)return n;if(t&&"FILE"===n)return!1}var i=this.statCache[r];if(!i){var o;try{o=this.fs.lstatSync(r)}catch(e){if(e&&("ENOENT"===e.code||"ENOTDIR"===e.code))return this.statCache[r]=!1,!1}if(o&&o.isSymbolicLink())try{i=this.fs.statSync(r)}catch(e){i=o}else i=o}return this.statCache[r]=i,n=!0,i&&(n=i.isDirectory()?"DIR":"FILE"),this.cache[r]=this.cache[r]||n,(!t||"FILE"!==n)&&n},h.prototype._mark=function(e){return o.mark(this,e)},h.prototype._makeAbs=function(e){return o.makeAbs(this,e)},cv}(),a=lv(),u=a.setopts,h=a.ownProp,d=Xv(),v=a.childrenIgnored,g=a.isIgnored,y=Bv();function m(e,r,t){if("function"==typeof r&&(t=r,r={}),r||(r={}),r.sync){if(t)throw new TypeError("callback provided to sync glob");return c(e,r)}return new b(e,r,t)}m.sync=c;var E=m.GlobSync=c.GlobSync;function b(e,r,t){if("function"==typeof r&&(t=r,r=null),r&&r.sync){if(t)throw new TypeError("callback provided to sync glob");return new E(e,r)}if(!(this instanceof b))return new b(e,r,t);u(this,e,r),this._didRealPath=!1;var n=this.minimatch.set.length;this.matches=new Array(n),"function"==typeof t&&(t=y(t),this.on("error",t),this.on("end",(function(e){t(null,e)})));var i=this;if(this._processing=0,this._emitQueue=[],this._processQueue=[],this.paused=!1,this.noprocess)return this;if(0===n)return s();for(var o=0;o<n;o++)this._process(this.minimatch.set[o],o,!1,s);function s(){--i._processing,i._processing<=0&&i._finish()}}return m.glob=m,m.hasMagic=function(e,r){var t=function(e,r){if(null===r||"object"!=typeof r)return e;for(var t=Object.keys(r),n=t.length;n--;)e[t[n]]=r[t[n]];return e}({},r);t.noprocess=!0;var n=new b(e,t).minimatch.set;if(!e)return!1;if(n.length>1)return!0;for(var i=0;i<n[0].length;i++)if("string"!=typeof n[0][i])return!0;return!1},m.Glob=b,t(b,n),b.prototype._finish=function(){if(o(this instanceof b),!this.aborted){if(this.realpath&&!this._didRealpath)return this._realpath();a.finish(this),this.emit("end",this.found)}},b.prototype._realpath=function(){if(!this._didRealpath){this._didRealpath=!0;var e=this.matches.length;if(0===e)return this._finish();for(var r=this,t=0;t<this.matches.length;t++)this._realpathSet(t,n)}function n(){0==--e&&r._finish()}},b.prototype._realpathSet=function(r,t){var n=this.matches[r];if(!n)return t();var i=Object.keys(n),o=this,s=i.length;if(0===s)return t();var c=this.matches[r]=Object.create(null);i.forEach((function(n,i){n=o._makeAbs(n),e.realpath(n,o.realpathCache,(function(e,i){e?"stat"===e.syscall?c[n]=!0:o.emit("error",e):c[i]=!0,0==--s&&(o.matches[r]=c,t())}))}))},b.prototype._mark=function(e){return a.mark(this,e)},b.prototype._makeAbs=function(e){return a.makeAbs(this,e)},b.prototype.abort=function(){this.aborted=!0,this.emit("abort")},b.prototype.pause=function(){this.paused||(this.paused=!0,this.emit("pause"))},b.prototype.resume=function(){if(this.paused){if(this.emit("resume"),this.paused=!1,this._emitQueue.length){var e=this._emitQueue.slice(0);this._emitQueue.length=0;for(var r=0;r<e.length;r++){var t=e[r];this._emitMatch(t[0],t[1])}}if(this._processQueue.length){var n=this._processQueue.slice(0);this._processQueue.length=0;for(r=0;r<n.length;r++){var i=n[r];this._processing--,this._process(i[0],i[1],i[2],i[3])}}}},b.prototype._process=function(e,t,n,i){if(o(this instanceof b),o("function"==typeof i),!this.aborted)if(this._processing++,this.paused)this._processQueue.push([e,t,n,i]);else{for(var c,a=0;"string"==typeof e[a];)a++;switch(a){case e.length:return void this._processSimple(e.join("/"),t,i);case 0:c=null;break;default:c=e.slice(0,a).join("/")}var u,l=e.slice(a);null===c?u=".":s(c)||s(e.map((function(e){return"string"==typeof e?e:"[*]"})).join("/"))?(c&&s(c)||(c="/"+c),u=c):u=c;var f=this._makeAbs(u);if(v(this,u))return i();l[0]===r.GLOBSTAR?this._processGlobStar(c,u,f,l,t,n,i):this._processReaddir(c,u,f,l,t,n,i)}},b.prototype._processReaddir=function(e,r,t,n,i,o,s){var c=this;this._readdir(t,o,(function(a,u){return c._processReaddir2(e,r,t,n,i,o,u,s)}))},b.prototype._processReaddir2=function(e,r,t,n,o,s,c,a){if(!c)return a();for(var u=n[0],l=!!this.minimatch.negate,f=u._glob,p=this.dot||"."===f.charAt(0),h=[],d=0;d<c.length;d++){if("."!==(g=c[d]).charAt(0)||p)(l&&!e?!g.match(u):g.match(u))&&h.push(g)}var v=h.length;if(0===v)return a();if(1===n.length&&!this.mark&&!this.stat){this.matches[o]||(this.matches[o]=Object.create(null));for(d=0;d<v;d++){var g=h[d];e&&(g="/"!==e?e+"/"+g:e+g),"/"!==g.charAt(0)||this.nomount||(g=i.join(this.root,g)),this._emitMatch(o,g)}return a()}n.shift();for(d=0;d<v;d++){g=h[d];e&&(g="/"!==e?e+"/"+g:e+g),this._process([g].concat(n),o,s,a)}a()},b.prototype._emitMatch=function(e,r){if(!this.aborted&&!g(this,r))if(this.paused)this._emitQueue.push([e,r]);else{var t=s(r)?r:this._makeAbs(r);if(this.mark&&(r=this._mark(r)),this.absolute&&(r=t),!this.matches[e][r]){if(this.nodir){var n=this.cache[t];if("DIR"===n||Array.isArray(n))return}this.matches[e][r]=!0;var i=this.statCache[t];i&&this.emit("stat",r,i),this.emit("match",r)}}},b.prototype._readdirInGlobStar=function(e,r){if(!this.aborted){if(this.follow)return this._readdir(e,!1,r);var t=this,n=d("lstat\0"+e,(function(n,i){if(n&&"ENOENT"===n.code)return r();var o=i&&i.isSymbolicLink();t.symlinks[e]=o,o||!i||i.isDirectory()?t._readdir(e,!1,r):(t.cache[e]="FILE",r())}));n&&t.fs.lstat(e,n)}},b.prototype._readdir=function(e,r,t){if(!this.aborted&&(t=d("readdir\0"+e+"\0"+r,t))){if(r&&!h(this.symlinks,e))return this._readdirInGlobStar(e,t);if(h(this.cache,e)){var n=this.cache[e];if(!n||"FILE"===n)return t();if(Array.isArray(n))return t(null,n)}this.fs.readdir(e,function(e,r,t){return function(n,i){n?e._readdirError(r,n,t):e._readdirEntries(r,i,t)}}(this,e,t))}},b.prototype._readdirEntries=function(e,r,t){if(!this.aborted){if(!this.mark&&!this.stat)for(var n=0;n<r.length;n++){var i=r[n];i="/"===e?e+i:e+"/"+i,this.cache[i]=!0}return this.cache[e]=r,t(null,r)}},b.prototype._readdirError=function(e,r,t){if(!this.aborted){switch(r.code){case"ENOTSUP":case"ENOTDIR":var n=this._makeAbs(e);if(this.cache[n]="FILE",n===this.cwdAbs){var i=new Error(r.code+" invalid cwd "+this.cwd);i.path=this.cwd,i.code=r.code,this.emit("error",i),this.abort()}break;case"ENOENT":case"ELOOP":case"ENAMETOOLONG":case"UNKNOWN":this.cache[this._makeAbs(e)]=!1;break;default:this.cache[this._makeAbs(e)]=!1,this.strict&&(this.emit("error",r),this.abort()),this.silent||console.error("glob error",r)}return t()}},b.prototype._processGlobStar=function(e,r,t,n,i,o,s){var c=this;this._readdir(t,o,(function(a,u){c._processGlobStar2(e,r,t,n,i,o,u,s)}))},b.prototype._processGlobStar2=function(e,r,t,n,i,o,s,c){if(!s)return c();var a=n.slice(1),u=e?[e]:[],l=u.concat(a);this._process(l,i,!1,c);var f=this.symlinks[t],p=s.length;if(f&&o)return c();for(var h=0;h<p;h++){if("."!==s[h].charAt(0)||this.dot){var d=u.concat(s[h],a);this._process(d,i,!0,c);var v=u.concat(s[h],n);this._process(v,i,!0,c)}}c()},b.prototype._processSimple=function(e,r,t){var n=this;this._stat(e,(function(i,o){n._processSimple2(e,r,i,o,t)}))},b.prototype._processSimple2=function(e,r,t,n,o){if(this.matches[r]||(this.matches[r]=Object.create(null)),!n)return o();if(e&&s(e)&&!this.nomount){var c=/[\/\\]$/.test(e);"/"===e.charAt(0)?e=i.join(this.root,e):(e=i.resolve(this.root,e),c&&(e+="/"))}"win32"===process.platform&&(e=e.replace(/\\/g,"/")),this._emitMatch(r,e),o()},b.prototype._stat=function(e,r){var t=this._makeAbs(e),n="/"===e.slice(-1);if(e.length>this.maxLength)return r();if(!this.stat&&h(this.cache,t)){var i=this.cache[t];if(Array.isArray(i)&&(i="DIR"),!n||"DIR"===i)return r(null,i);if(n&&"FILE"===i)return r()}var o=this.statCache[t];if(void 0!==o){if(!1===o)return r(null,o);var s=o.isDirectory()?"DIR":"FILE";return n&&"FILE"===s?r():r(null,s,o)}var c=this,a=d("stat\0"+t,(function(n,i){if(i&&i.isSymbolicLink())return c.fs.stat(t,(function(n,o){n?c._stat2(e,t,null,i,r):c._stat2(e,t,n,o,r)}));c._stat2(e,t,n,i,r)}));a&&c.fs.lstat(t,a)},b.prototype._stat2=function(e,r,t,n,i){if(t&&("ENOENT"===t.code||"ENOTDIR"===t.code))return this.statCache[r]=!1,i();var o="/"===e.slice(-1);if(this.statCache[r]=n,"/"===r.slice(-1)&&n&&!n.isDirectory())return i(null,!1,n);var s=!0;return n&&(s=n.isDirectory()?"DIR":"FILE"),this.cache[r]=this.cache[r]||s,o&&"FILE"===s?i():i(null,s,n)},Lv}function Wv(){if(Tv)return wh;Tv=1;var e=a.default,r=u.default,t=qv(),n=Ty(),i=Object.create(n);wh.extend=Object.assign;var o=Boolean(process.versions.electron),s={fatal:!1,globOptions:{},maxdepth:255,noglob:!1,silent:!1,verbose:!1,execPath:null,bufLength:65536},c={reset:function(){Object.assign(this,s),o||(this.execPath=process.execPath)},resetForTesting:function(){this.reset(),this.silent=!0}};c.reset(),wh.config=c;var l={error:null,errorCode:0,currentCmd:"shell.js"};function f(e){return"object"==typeof e&&null!==e}function p(){c.silent||console.error.apply(console,arguments)}function h(e){if("string"!=typeof e)throw new TypeError("input must be a string");return e.replace(/\\/g,"/")}function d(e,r,t){if("string"!=typeof e)throw new Error("msg must be a string");var n={continue:!1,code:1,prefix:l.currentCmd+": ",silent:!1};"number"==typeof r&&f(t)?t.code=r:f(r)?t=r:"number"==typeof r?t={code:r}:"number"!=typeof r&&(t={}),t=Object.assign({},n,t),l.errorCode||(l.errorCode=t.code);var i=h(t.prefix+e);if(l.error=l.error?l.error+"\n":"",l.error+=i,c.fatal)throw new Error(i);if(e.length>0&&!t.silent&&p(i),!t.continue)throw{msg:"earlyExit",retValue:new v("",l.error,l.errorCode)}}function v(e,r,t){var n;return e instanceof Array?((n=e).stdout=e.join("\n"),e.length>0&&(n.stdout+="\n")):(n=new String(e)).stdout=e,n.stderr=r,n.code=t,w.forEach((function(e){n[e]=i[e].bind(n)})),n}function g(e,r,t){if("string"!=typeof e&&!f(e))throw new Error("options must be strings or key-value pairs");if(!f(r))throw new Error("parseOptions() internal error: map must be an object");if(t&&!f(t))throw new Error("parseOptions() internal error: errorOptions must be object");if("--"===e)return{};var n={};if(Object.keys(r).forEach((function(e){var t=r[e];"!"!==t[0]&&(n[t]=!1)})),""===e)return n;if("string"==typeof e){if("-"!==e[0])throw new Error("Options string must start with a '-'");e.slice(1).split("").forEach((function(e){if(e in r){var i=r[e];"!"===i[0]?n[i.slice(1)]=!1:n[i]=!0}else d("option not recognized: "+e,t||{})}))}else Object.keys(e).forEach((function(i){var o=i[1];if(o in r){var s=r[o];n[s]=e[i]}else d("option not recognized: "+o,t||{})}));return n}function y(e){if(!Array.isArray(e))throw new TypeError("must be an array");var r=[];return e.forEach((function(e){if("string"!=typeof e)r.push(e);else{var n;try{n=(n=t.sync(e,c.globOptions)).length>0?n:[e]}catch(r){n=[e]}r=r.concat(n)}})),r}wh.state=l,delete process.env.OLDPWD,wh.isObject=f,wh.log=p,wh.convertErrorOutput=h,wh.error=d,wh.ShellString=v,wh.parseOptions=g,wh.expand=y;var m="function"==typeof Buffer.alloc?function(e){return Buffer.alloc(e||c.bufLength)}:function(e){return new Buffer(e||c.bufLength)};function E(r,t,n){return n=n||{},function(){var i=null;l.currentCmd=r,l.error=null,l.errorCode=0;try{var o=[].slice.call(arguments,0);if(c.verbose&&console.error.apply(console,[r].concat(o)),l.pipedValue=this&&"string"==typeof this.stdout?this.stdout:"",!1===n.unix)i=t.apply(this,o);else{f(o[0])&&"Object"===o[0].constructor.name||(0===o.length||"string"!=typeof o[0]||o[0].length<=1||"-"!==o[0][0])&&o.unshift(""),o=(o=o.reduce((function(e,r){return Array.isArray(r)?e.concat(r):(e.push(r),e)}),[])).map((function(e){return f(e)&&"String"===e.constructor.name?e.toString():e}));var s=e.homedir();o=o.map((function(e){return"string"==typeof e&&"~/"===e.slice(0,2)||"~"===e?e.replace(/^~/,s):e})),c.noglob||!0!==n.allowGlobbing||(o=o.slice(0,n.globStart).concat(y(o.slice(n.globStart))));try{f(n.cmdOptions)&&(o[0]=g(o[0],n.cmdOptions)),i=t.apply(this,o)}catch(e){if("earlyExit"!==e.msg)throw e;i=e.retValue}}}catch(e){if(!l.error)throw e.name="ShellJSInternalError",e;if(c.fatal)throw e}return n.wrapOutput&&("string"==typeof i||Array.isArray(i))&&(i=new v(i,l.error,l.errorCode)),l.currentCmd="shell.js",i}}wh.buffer=m,wh.unlinkSync=function(e){try{r.unlinkSync(e)}catch(t){if("EPERM"!==t.code)throw t;r.chmodSync(e,"0666"),r.unlinkSync(e)}},wh.statFollowLinks=function(){return r.statSync.apply(r,arguments)},wh.statNoFollowLinks=function(){return r.lstatSync.apply(r,arguments)},wh.randomFileName=function(){return"shelljs_"+function e(r){if(1===r)return parseInt(16*Math.random(),10).toString(16);for(var t="",n=0;n<r;n++)t+=e(1);return t}(20)},wh.wrap=E,wh.readFromPipe=function(){return l.pipedValue};var b={allowGlobbing:!0,canReceivePipe:!1,cmdOptions:null,globStart:1,pipeOnly:!1,wrapOutput:!0,unix:!0},w=[];return wh.register=function(e,r,t){if(t=t||{},Object.keys(t).forEach((function(e){if(!b.hasOwnProperty(e))throw new Error("Unknown option '"+e+"'");if(typeof t[e]!=typeof b[e])throw new TypeError("Unsupported type '"+typeof t[e]+"' for option '"+e+"'")})),t=Object.assign({},b,t),n.hasOwnProperty(e))throw new Error("Command `"+e+"` already exists");t.pipeOnly?(t.canReceivePipe=!0,i[e]=E(e,r,t)):n[e]=E(e,r,t),t.canReceivePipe&&w.push(e)},wh}function Vv(){if(Pv)return Nv;Pv=1;var e=Wv(),r=u.default;function t(t,i){var o=e.readFromPipe();return i||o||e.error("no paths given"),(i=[].slice.call(arguments,1)).forEach((function(t){r.existsSync(t)?e.statFollowLinks(t).isDirectory()&&e.error(t+": Is a directory"):e.error("no such file or directory: "+t),o+=r.readFileSync(t,"utf8")})),t.number&&(o=n(o)),o}function n(e){var r=e.split("\n"),t=r.pop();return r=r.map((function(e,r){return i(r+1,e)})),t.length&&(t=i(r.length+1,t)),r.push(t),r.join("\n")}function i(e,r){return("     "+e).slice(-6)+"\t"+r}return e.register("cat",t,{canReceivePipe:!0,cmdOptions:{n:"number"}}),Nv=t}function Hv(){if(Dv)return Fv;Dv=1;var e=a.default,r=Wv();function t(t,n){n||(n=e.homedir()),"-"===n&&(process.env.OLDPWD?n=process.env.OLDPWD:r.error("could not find previous directory"));try{var i=process.cwd();process.chdir(n),process.env.OLDPWD=i}catch(e){var o;try{r.statFollowLinks(n),o="not a directory: "+n}catch(e){o="no such file or directory: "+n}o&&r.error(o)}return""}return r.register("cd",t,{}),Fv=t}function zv(){if(Mv)return Cv;Mv=1;var e,r=Wv(),t=u.default,n=l.default,i={OTHER_EXEC:(e={EXEC:1,WRITE:2,READ:4}).EXEC,OTHER_WRITE:e.WRITE,OTHER_READ:e.READ,GROUP_EXEC:e.EXEC<<3,GROUP_WRITE:e.WRITE<<3,GROUP_READ:e.READ<<3,OWNER_EXEC:e.EXEC<<6,OWNER_WRITE:e.WRITE<<6,OWNER_READ:e.READ<<6,STICKY:parseInt("01000",8),SETGID:parseInt("02000",8),SETUID:parseInt("04000",8),TYPE_MASK:parseInt("0770000",8)};function o(e,o,s){var c;return s||(e.length>0&&"-"===e.charAt(0)?[].unshift.call(arguments,""):r.error("You must specify a file.")),e=r.parseOptions(e,{R:"recursive",c:"changes",v:"verbose"}),s=[].slice.call(arguments,2),e.recursive?(c=[],s.forEach((function e(n){var i=r.statNoFollowLinks(n);i.isSymbolicLink()||(c.push(n),i.isDirectory()&&t.readdirSync(n).forEach((function(r){e(n+"/"+r)})))}))):c=s,c.forEach((function(s){if(s=n.resolve(s),t.existsSync(s)||r.error("File not found: "+s),!e.recursive||!r.statNoFollowLinks(s).isSymbolicLink()){var c=r.statFollowLinks(s),a=c.isDirectory(),u=c.mode,l=u&i.TYPE_MASK,f=u;isNaN(parseInt(o,8))?o.split(",").forEach((function(n){var o=/([ugoa]*)([=\+-])([rwxXst]*)/i.exec(n);if(o){var c=o[1],p=o[2],h=o[3],d=-1!==c.indexOf("u")||"a"===c||""===c,v=-1!==c.indexOf("g")||"a"===c||""===c,g=-1!==c.indexOf("o")||"a"===c||""===c,y=-1!==h.indexOf("r"),m=-1!==h.indexOf("w"),E=-1!==h.indexOf("x"),b=-1!==h.indexOf("X"),w=-1!==h.indexOf("t"),O=-1!==h.indexOf("s");b&&a&&(E=!0);var S=0;switch(d&&(S|=(y?i.OWNER_READ:0)+(m?i.OWNER_WRITE:0)+(E?i.OWNER_EXEC:0)+(O?i.SETUID:0)),v&&(S|=(y?i.GROUP_READ:0)+(m?i.GROUP_WRITE:0)+(E?i.GROUP_EXEC:0)+(O?i.SETGID:0)),g&&(S|=(y?i.OTHER_READ:0)+(m?i.OTHER_WRITE:0)+(E?i.OTHER_EXEC:0)),w&&(S|=i.STICKY),p){case"+":f|=S;break;case"-":f&=~S;break;case"=":f=l+S,r.statFollowLinks(s).isDirectory()&&(f|=i.SETUID+i.SETGID&u);break;default:r.error("Could not recognize operator: `"+p+"`")}e.verbose&&console.log(s+" -> "+f.toString(8)),u!==f&&(!e.verbose&&e.changes&&console.log(s+" -> "+f.toString(8)),t.chmodSync(s,f),u=f)}else r.error("Invalid symbolic mode change: "+n)})):(f=l+parseInt(o,8),r.statFollowLinks(s).isDirectory()&&(f|=i.SETUID+i.SETGID&u),t.chmodSync(s,f))}})),""}return r.register("chmod",o,{}),Cv=o}function Yv(){if($v)return Gv;$v=1;var e=u.default,r=l.default,t=Wv();function n(r,n,i){e.existsSync(r)||t.error("copyFileSync: no such file or directory: "+r);var o="win32"===process.platform;try{if(i.update&&t.statFollowLinks(r).mtime<e.statSync(n).mtime)return}catch(e){}if(t.statNoFollowLinks(r).isSymbolicLink()&&!i.followsymlink){try{t.statNoFollowLinks(n),t.unlinkSync(n)}catch(e){}var s=e.readlinkSync(r);e.symlinkSync(s,n,o?"junction":null)}else{var c=t.buffer(),a=c.length,u=a,l=0,f=null,p=null;try{f=e.openSync(r,"r")}catch(e){t.error("copyFileSync: could not read src file ("+r+")")}try{p=e.openSync(n,"w")}catch(e){t.error("copyFileSync: could not write to dest file (code="+e.code+"):"+n)}for(;u===a;)u=e.readSync(f,c,0,a,l),e.writeSync(p,c,0,u),l+=u;e.closeSync(f),e.closeSync(p),e.chmodSync(n,t.statFollowLinks(r).mode)}}function i(r,o,c,a){if(a||(a={}),!(c>=t.config.maxdepth)){c++;var u="win32"===process.platform;try{e.mkdirSync(o)}catch(e){if("EEXIST"!==e.code)throw e}for(var l=e.readdirSync(r),f=0;f<l.length;f++){var p,h=r+"/"+l[f],d=o+"/"+l[f],v=t.statNoFollowLinks(h);if(a.followsymlink&&s(r,h))console.error("Cycle link found."),p=e.readlinkSync(h),e.symlinkSync(p,d,u?"junction":null);else if(v.isDirectory())i(h,d,c,a);else if(v.isSymbolicLink()&&!a.followsymlink){p=e.readlinkSync(h);try{t.statNoFollowLinks(d),t.unlinkSync(d)}catch(e){}e.symlinkSync(p,d,u?"junction":null)}else v.isSymbolicLink()&&a.followsymlink?(v=t.statFollowLinks(h)).isDirectory()?i(h,d,c,a):n(h,d,a):e.existsSync(d)&&a.no_force?t.log("skipping existing file: "+l[f]):n(h,d,a)}var g=t.statFollowLinks(r);e.chmodSync(o,g.mode)}}function o(e,t){var n=e[t];return e.slice(0,t).some((function(e){return r.basename(e)===r.basename(n)}))}function s(r,n){if(t.statNoFollowLinks(n).isSymbolicLink()&&t.statFollowLinks(n).isDirectory()){var i=e.realpathSync(r),o=e.realpathSync(n);if(new RegExp(o).test(i))return!0}return!1}function c(s,c,a){s.followsymlink&&(s.noFollowsymlink=!1),s.recursive||s.noFollowsymlink||(s.followsymlink=!0),arguments.length<3?t.error("missing <source> and/or <dest>"):(c=[].slice.call(arguments,1,arguments.length-1),a=arguments[arguments.length-1]);var u=e.existsSync(a),l=u&&t.statFollowLinks(a);return u&&l.isDirectory()||!(c.length>1)||t.error("dest is not a directory (too many sources)"),u&&l.isFile()&&s.no_force?new t.ShellString("","",0):(c.forEach((function(u,f){if(!e.existsSync(u))return""===u&&(u="''"),void t.error("no such file or directory: "+u,{continue:!0});var p=t.statFollowLinks(u);if(!s.noFollowsymlink&&p.isDirectory())if(s.recursive){var h=l&&l.isDirectory()?r.join(a,r.basename(u)):a;try{t.statFollowLinks(r.dirname(a)),i(u,h,0,{no_force:s.no_force,followsymlink:s.followsymlink})}catch(e){t.error("cannot create directory '"+a+"': No such file or directory")}}else t.error("omitting directory '"+u+"'",{continue:!0});else{var d=a;l&&l.isDirectory()&&(d=r.normalize(a+"/"+r.basename(u)));var v=e.existsSync(d);if(v&&o(c,f))return void(s.no_force||t.error("will not overwrite just-created '"+d+"' with '"+u+"'",{continue:!0}));if(v&&s.no_force)return;if(""===r.relative(u,d))return void t.error("'"+d+"' and '"+u+"' are the same file",{continue:!0});n(u,d,s)}})),new t.ShellString("",t.state.error,t.state.errorCode))}return t.register("cp",c,{cmdOptions:{f:"!no_force",n:"no_force",u:"update",R:"recursive",r:"recursive",L:"followsymlink",P:"noFollowsymlink"},wrapOutput:!1}),Gv=c}var Kv,Jv,Qv,Zv,eg,rg={};function tg(){if(Kv)return rg;Kv=1;var e=Wv(),r=Hv(),t=l.default;e.register("dirs",u,{wrapOutput:!1}),e.register("pushd",c,{wrapOutput:!1}),e.register("popd",a,{wrapOutput:!1});var n=[];function i(e){return/^[\-+]\d+$/.test(e)}function o(r){if(i(r)){if(Math.abs(r)<n.length+1)return/^-/.test(r)?Number(r)-1:Number(r);e.error(r+": directory stack index out of range")}else e.error(r+": invalid number")}function s(){return[process.cwd()].concat(n)}function c(c,a){i(c)&&(a=c,c=""),c=e.parseOptions(c,{n:"no-cd",q:"quiet"});var l=s();if("+0"===a)return l;if(a)if(i(a)){var f=o(a);l=l.slice(f).concat(l.slice(0,f))}else c["no-cd"]?l.splice(1,0,a):l.unshift(a);else{if(!(l.length>1))return e.error("no other directory");l=l.splice(1,1).concat(l)}return c["no-cd"]?l=l.slice(1):(a=t.resolve(l.shift()),r("",a)),n=l,u(c.quiet?"-q":"")}function a(s,c){if(i(s)&&(c=s,s=""),s=e.parseOptions(s,{n:"no-cd",q:"quiet"}),!n.length)return e.error("directory stack empty");if(c=o(c||"+0"),s["no-cd"]||c>0||n.length+c===0)c=c>0?c-1:c,n.splice(c,1);else{var a=t.resolve(n.shift());r("",a)}return u(s.quiet?"-q":"")}function u(r,t){if(i(r)&&(t=r,r=""),(r=e.parseOptions(r,{c:"clear",q:"quiet"})).clear)return n=[];var c=s();return t?((t=o(t))<0&&(t=c.length+t),r.quiet||e.log(c[t]),c[t]):(r.quiet||e.log(c.join(" ")),c)}return rg.pushd=c,rg.popd=a,rg.dirs=u,rg}function ng(){if(Qv)return Jv;Qv=1;var e=h.default.format,r=Wv();function t(t){var n=[].slice.call(arguments,t?0:1),i={};try{i=r.parseOptions(n[0],{e:"escapes",n:"no_newline"},{silent:!0}),n[0]&&n.shift()}catch(e){r.state.error=null}var o=e.apply(null,n);return i.no_newline||(o+="\n"),process.stdout.write(o),o}return r.register("echo",t,{allowGlobbing:!1}),Jv=t}function ig(){if(eg)return Zv;eg=1;var e=Wv();return Zv=function(){return e.state.error}}var og,sg={exports:{}};function cg(){return og||(og=1,function(e){if(require.main!==e)throw new Error("This file should not be required");var r=d.default,t=u.default,n=process.argv[2],i=t.readFileSync(n,"utf8"),o=JSON.parse(i),s=o.command,c=o.execOptions,a=o.pipe,l=o.stdoutFile,f=o.stderrFile,p=r.exec(s,c,(function(e){e?void 0===e.code?process.exitCode=1:process.exitCode=e.code:process.exitCode=0})),h=t.createWriteStream(l),v=t.createWriteStream(f);p.stdout.pipe(h),p.stderr.pipe(v),p.stdout.pipe(process.stdout),p.stderr.pipe(process.stderr),a&&p.stdin.end(a)}(sg)),sg.exports}var ag,ug,lg,fg,pg,hg,dg,vg,gg,yg,mg,Eg,bg,wg,Og,Sg,Rg,Ig,Ag,jg,xg,kg={};function Lg(){if(ag)return kg;ag=1;var e,r=Wv(),t=a.default,n=u.default;function i(e){if(!e||!n.existsSync(e))return!1;if(!r.statFollowLinks(e).isDirectory())return!1;var t=e+"/"+r.randomFileName();try{return n.writeFileSync(t," "),r.unlinkSync(t),e}catch(e){return!1}}function o(){return e||(e=i(t.tmpdir())||i(process.env.TMPDIR)||i(process.env.TEMP)||i(process.env.TMP)||i(process.env.Wimp$ScrapDir)||i("C:\\TEMP")||i("C:\\TMP")||i("\\TEMP")||i("\\TMP")||i("/tmp")||i("/var/tmp")||i("/usr/tmp")||i("."))}return r.register("tempdir",o,{allowGlobbing:!1,wrapOutput:!1}),kg.tempDir=o,kg.isCached=function(){return e},kg.clearCache=function(){e=void 0},kg}function _g(){if(lg)return ug;lg=1;var e=l.default;function r(){return e.resolve(process.cwd())}return Wv().register("pwd",r,{allowGlobbing:!1}),ug=r}function Tg(){if(pg)return fg;pg=1;var e=Wv(),r=Lg().tempDir,t=_g(),n=l.default,i=u.default,o=d.default,s=20971520;function c(c,a,u){a=a||{},c||e.error("must specify command");var l=e.readFromPipe();return"function"==typeof a&&(u=a,a={async:!0}),"object"==typeof a&&"function"==typeof u&&(a.async=!0),(a=e.extend({silent:e.config.silent,async:!1},a)).async?function(r,n,i,c){n=e.extend({silent:e.config.silent,cwd:t().toString(),env:process.env,maxBuffer:s,encoding:"utf8"},n);var a=o.exec(r,n,(function(e,r,t){c&&(e?void 0===e.code?c(1,r,t):c(e.code,r,t):c(0,r,t))}));return i&&a.stdin.end(i),n.silent||(a.stdout.pipe(process.stdout),a.stderr.pipe(process.stderr)),a}(c,a,l,u):function(c,a,u){e.config.execPath||e.error("Unable to find a path to the node binary. Please manually set config.execPath");var l=r(),f=n.resolve(l+"/"+e.randomFileName()),p=n.resolve(l+"/"+e.randomFileName()),h=n.resolve(l+"/"+e.randomFileName());a=e.extend({silent:e.config.silent,cwd:t().toString(),env:process.env,maxBuffer:s,encoding:"utf8"},a),i.existsSync(f)&&e.unlinkSync(f),i.existsSync(p)&&e.unlinkSync(p),i.existsSync(h)&&e.unlinkSync(h),a.cwd=n.resolve(a.cwd);var d={command:c,execOptions:a,pipe:u,stdoutFile:h,stderrFile:p};function v(e,r){i.writeFileSync(e,r,{encoding:"utf8",mode:parseInt("600",8)})}v(h,""),v(p,""),v(f,JSON.stringify(d));var g=[n.join(__dirname,"exec-child.js"),f];a.silent?a.stdio="ignore":a.stdio=[0,1,2];var y=0;try{delete a.shell,o.execFileSync(e.config.execPath,g,a)}catch(e){y=e.status||1}var m="",E="";"buffer"===a.encoding?(m=i.readFileSync(h),E=i.readFileSync(p)):(m=i.readFileSync(h,a.encoding),E=i.readFileSync(p,a.encoding));try{e.unlinkSync(f)}catch(e){}try{e.unlinkSync(p)}catch(e){}try{e.unlinkSync(h)}catch(e){}return 0!==y&&e.error(E,y,{continue:!0,silent:!0}),e.ShellString(m,E,y)}(c,a,l)}return e.register("exec",c,{unix:!1,canReceivePipe:!0,wrapOutput:!1}),fg=c}function Ng(){if(dg)return hg;dg=1;var e=l.default,r=u.default,t=Wv(),n=qv(),i=e.sep+"**";function o(o,c){o.all_deprecated&&(t.log("ls: Option -a is deprecated. Use -A instead"),o.all=!0),c=c?[].slice.call(arguments,1):["."];var a=[];function u(e,r,n){"win32"===process.platform&&(r=r.replace(/\\/g,"/")),o.long?(n=n||(o.link?t.statFollowLinks(e):t.statNoFollowLinks(e)),a.push(s(r,n))):a.push(r)}return c.forEach((function(s){var c;try{if((c=o.link?t.statFollowLinks(s):t.statNoFollowLinks(s)).isSymbolicLink())try{var a=t.statFollowLinks(s);a.isDirectory()&&(c=a)}catch(e){}}catch(e){return void t.error("no such file or directory: "+s,2,{continue:!0})}c.isDirectory()&&!o.directory?o.recursive?n.sync(s+i,{dot:o.all,follow:o.link}).forEach((function(r){e.relative(s,r)&&u(r,e.relative(s,r))})):o.all?r.readdirSync(s).forEach((function(r){u(e.join(s,r),r)})):r.readdirSync(s).forEach((function(r){"."!==r[0]&&u(e.join(s,r),r)})):u(s,s,c)})),a}function s(e,r){return r.name=e,r.toString=function(){return[this.mode,this.nlink,this.uid,this.gid,this.size,this.mtime,this.name].join(" ")},r}return t.register("ls",o,{cmdOptions:{R:"recursive",A:"all",L:"link",a:"all_deprecated",d:"directory",l:"long"}}),hg=o}function Pg(){if(gg)return vg;gg=1;var e=l.default,r=Wv(),t=Ng();function n(n,i){i?"string"==typeof i&&(i=[].slice.call(arguments,1)):r.error("no path specified");var o=[];function s(e){"win32"===process.platform&&(e=e.replace(/\\/g,"/")),o.push(e)}return i.forEach((function(n){var i;try{i=r.statFollowLinks(n)}catch(e){r.error("no such file or directory: "+n)}s(n),i.isDirectory()&&t({recursive:!0,all:!0},n).forEach((function(r){s(e.join(n,r))}))})),o}return r.register("find",n,{}),vg=n}function Fg(){if(mg)return yg;mg=1;var e=Wv(),r=u.default;function t(t,n,i){var o=e.readFromPipe();i||o||e.error("no paths given",2),i=[].slice.call(arguments,2),o&&i.unshift("-");var s=[];return t.ignoreCase&&(n=new RegExp(n,"i")),i.forEach((function(i){if(r.existsSync(i)||"-"===i){var c="-"===i?o:r.readFileSync(i,"utf8");if(t.nameOnly)c.match(n)&&s.push(i);else c.split("\n").forEach((function(e){var r=e.match(n);(t.inverse&&!r||!t.inverse&&r)&&s.push(e)}))}else e.error("no such file or directory: "+i,2,{continue:!0})})),s.join("\n")+"\n"}return e.register("grep",t,{globStart:2,canReceivePipe:!0,cmdOptions:{v:"inverse",l:"nameOnly",i:"ignoreCase"}}),yg=t}function Dg(){if(bg)return Eg;bg=1;var e=Wv(),r=u.default;function t(t,n){for(var i=e.buffer(),o=i.length,s=o,c=0,a=r.openSync(t,"r"),u=0,l="";s===o&&u<n;){s=r.readSync(a,i,0,o,c);var f=i.toString("utf8",0,s);u+=f.split("\n").length-1,l+=f,c+=s}return r.closeSync(a),l}function n(n,i){var o=[],s=e.readFromPipe();i||s||e.error("no paths given");var c=1;!0===n.numLines?(c=2,n.numLines=Number(arguments[1])):!1===n.numLines&&(n.numLines=10),i=[].slice.call(arguments,c),s&&i.unshift("-");var a=!1;return i.forEach((function(i){if("-"!==i){if(!r.existsSync(i))return void e.error("no such file or directory: "+i,{continue:!0});if(e.statFollowLinks(i).isDirectory())return void e.error("error reading '"+i+"': Is a directory",{continue:!0})}var c=("-"===i?s:n.numLines<0?r.readFileSync(i,"utf8"):t(i,n.numLines)).split("\n"),u=""===c[c.length-1];u&&c.pop(),a=u||n.numLines<c.length,o=o.concat(c.slice(0,n.numLines))})),a&&o.push(""),o.join("\n")}return e.register("head",n,{canReceivePipe:!0,cmdOptions:{n:"numLines"}}),Eg=n}function Cg(){if(Og)return wg;Og=1;var e=u.default,r=l.default,t=Wv();function n(n,i,o){i&&o||t.error("Missing <source> and/or <dest>"),i=String(i);var s=r.normalize(i).replace(RegExp(r.sep+"$"),""),c=r.resolve(i)===s;if(o=r.resolve(process.cwd(),String(o)),e.existsSync(o)&&(n.force||t.error("Destination file exists",{continue:!0}),e.unlinkSync(o)),n.symlink){var a="win32"===process.platform,u=a?"file":null,l=c?s:r.resolve(process.cwd(),r.dirname(o),i);e.existsSync(l)?a&&t.statFollowLinks(l).isDirectory()&&(u="junction"):t.error("Source file does not exist",{continue:!0});try{e.symlinkSync("junction"===u?l:i,o,u)}catch(e){t.error(e.message)}}else{e.existsSync(i)||t.error("Source file does not exist",{continue:!0});try{e.linkSync(i,o)}catch(e){t.error(e.message)}}return""}return t.register("ln",n,{cmdOptions:{s:"symlink",f:"force"}}),wg=n}function Mg(){if(Rg)return Sg;Rg=1;var e=Wv(),r=u.default,t=l.default;function n(i){var o=t.dirname(i);o===i&&e.error("dirname() failed: ["+i+"]"),r.existsSync(o)||n(o),r.mkdirSync(i,parseInt("0777",8))}function i(i,o){return o||e.error("no paths given"),"string"==typeof o&&(o=[].slice.call(arguments,1)),o.forEach((function(o){try{var s=e.statNoFollowLinks(o);return void(i.fullpath?s.isFile()&&e.error("cannot create directory "+o+": File exists",{continue:!0}):e.error("path already exists: "+o,{continue:!0}))}catch(e){}var c=t.dirname(o);if(r.existsSync(c)||i.fullpath)try{i.fullpath?n(t.resolve(o)):r.mkdirSync(o,parseInt("0777",8))}catch(r){var a;if("EACCES"===r.code)a="Permission denied";else{if("ENOTDIR"!==r.code&&"ENOENT"!==r.code)throw r;a="Not a directory"}e.error("cannot create directory "+o+": "+a,{continue:!0})}else e.error("no such file or directory: "+c,{continue:!0})})),""}return e.register("mkdir",i,{cmdOptions:{p:"fullpath"}}),Sg=i}function Gg(){if(Ag)return Ig;Ag=1;var e=Wv(),r=u.default;function t(i,o,s){var c;c=r.readdirSync(i);for(var a=0;a<c.length;a++){var u=i+"/"+c[a];if(e.statNoFollowLinks(u).isDirectory())t(u,o);else if(o||n(u))try{e.unlinkSync(u)}catch(r){e.error("could not remove file (code "+r.code+"): "+u,{continue:!0})}}if(!s){var l;try{for(var f=Date.now();;)try{if(l=r.rmdirSync(i),r.existsSync(i))throw{code:"EAGAIN"};break}catch(e){if("win32"!==process.platform||"ENOTEMPTY"!==e.code&&"EBUSY"!==e.code&&"EPERM"!==e.code&&"EAGAIN"!==e.code){if("ENOENT"===e.code)break;throw e}if(Date.now()-f>1e3)throw e}}catch(r){e.error("could not remove directory (code "+r.code+"): "+i,{continue:!0})}return l}}function n(e){var t=!0;try{var n=r.openSync(e,"a");r.closeSync(n)}catch(e){t=!1}return t}function i(r,t){t.force||n(r)?e.unlinkSync(r):e.error("permission denied: "+r,{continue:!0})}function o(r,n){n.recursive?t(r,n.force):e.error("path is a directory",{continue:!0})}function s(r,n){var i;try{i=e.statFollowLinks(r)}catch(t){return void e.unlinkSync(r)}if(i.isFile())e.unlinkSync(r);else if(i.isDirectory())if("/"===r[r.length-1])if(n.recursive){t(r,n.force,!0)}else e.error("path is a directory",{continue:!0});else e.unlinkSync(r)}function c(r){e.unlinkSync(r)}function a(r,t){return t||e.error("no paths given"),(t=[].slice.call(arguments,1)).forEach((function(t){var n;try{var a="/"===t[t.length-1]?t.slice(0,-1):t;n=e.statNoFollowLinks(a)}catch(n){return void(r.force||e.error("no such file or directory: "+t,{continue:!0}))}n.isFile()?i(t,r):n.isDirectory()?o(t,r):n.isSymbolicLink()?s(t,r):n.isFIFO()&&c(t)})),""}return e.register("rm",a,{cmdOptions:{f:"force",r:"recursive",R:"recursive"}}),Ig=a}function $g(){if(xg)return jg;xg=1;var e=u.default,r=l.default,t=Wv(),n=Yv(),i=Gg();function o(e,t){var n=e[t];return e.slice(0,t).some((function(e){return r.basename(e)===r.basename(n)}))}function s(s,c,a){arguments.length<3?t.error("missing <source> and/or <dest>"):arguments.length>3?(c=[].slice.call(arguments,1,arguments.length-1),a=arguments[arguments.length-1]):"string"==typeof c?c=[c]:t.error("invalid arguments");var u=e.existsSync(a),l=u&&t.statFollowLinks(a);return u&&l.isDirectory()||!(c.length>1)||t.error("dest is not a directory (too many sources)"),u&&l.isFile()&&s.no_force&&t.error("dest file already exists: "+a),c.forEach((function(u,l){if(e.existsSync(u)){var f=a;if(e.existsSync(a)&&t.statFollowLinks(a).isDirectory()&&(f=r.normalize(a+"/"+r.basename(u))),e.existsSync(f)&&o(c,l))s.no_force||t.error("will not overwrite just-created '"+f+"' with '"+u+"'",{continue:!0});else if(e.existsSync(f)&&s.no_force)t.error("dest file already exists: "+f,{continue:!0});else if(r.resolve(u)!==r.dirname(r.resolve(f)))try{e.renameSync(u,f)}catch(e){"EXDEV"===e.code&&(n("-r",u,f),i("-rf",u))}else t.error("cannot move to self: "+u,{continue:!0})}else t.error("no such file or directory: "+u,{continue:!0})})),""}return t.register("mv",s,{cmdOptions:{f:"!no_force",n:"no_force"}}),jg=s}var Ug,Bg={};function Xg(){return Ug||(Ug=1),Bg}var qg,Wg,Vg,Hg,zg,Yg={};function Kg(){return qg||(qg=1),Yg}function Jg(){if(Vg)return Wg;Vg=1;var e=Wv(),r=u.default;function t(t,n,i,o){var s=e.readFromPipe();"string"!=typeof i&&"function"!=typeof i&&("number"==typeof i?i=i.toString():e.error("invalid replacement string")),"string"==typeof n&&(n=RegExp(n)),o||s||e.error("no files given"),o=[].slice.call(arguments,3),s&&o.unshift("-");var c=[];return o.forEach((function(o){if(r.existsSync(o)||"-"===o){var a=("-"===o?s:r.readFileSync(o,"utf8")).split("\n").map((function(e){return e.replace(n,i)})).join("\n");c.push(a),t.inplace&&r.writeFileSync(o,a,"utf8")}else e.error("no such file or directory: "+o,2,{continue:!0})})),c.join("\n")}return e.register("sed",t,{globStart:3,canReceivePipe:!0,cmdOptions:{i:"inplace"}}),Wg=t}function Qg(){if(zg)return Hg;zg=1;var e=Wv();function r(r){if(!r){var t=[].slice.call(arguments,0);t.length<2&&e.error("must provide an argument"),r=t[1]}var n="+"===r[0];n&&(r="-"+r.slice(1)),r=e.parseOptions(r,{e:"fatal",v:"verbose",f:"noglob"}),n&&Object.keys(r).forEach((function(e){r[e]=!r[e]})),Object.keys(r).forEach((function(t){n!==r[t]&&(e.config[t]=r[t])}))}return e.register("set",r,{allowGlobbing:!1,wrapOutput:!1}),Hg=r}var Zg=_f.end,ey=Ff("trimEnd")?function(){return Zg(this)}:"".trimEnd;fi({target:"String",proto:!0,name:"trimEnd",forced:"".trimRight!==ey},{trimRight:ey});var ry,ty,ny,iy,oy,sy,cy,ay,uy,ly,fy,py,hy,dy,vy,gy,yy;function my(){if(ty)return ry;ty=1;var e=Wv(),r=u.default;function t(e){var r=e.match(/^\s*(\d*)\s*(.*)$/);return{num:Number(r[1]),value:r[2]}}function n(e,r){var t=e.toLowerCase(),n=r.toLowerCase();return t===n?-1*e.localeCompare(r):t.localeCompare(n)}function i(e,r){var i=t(e),o=t(r);return i.hasOwnProperty("num")&&o.hasOwnProperty("num")&&i.num!==o.num?i.num-o.num:n(i.value,o.value)}function o(t,o){var s=e.readFromPipe();o||s||e.error("no files given"),o=[].slice.call(arguments,1),s&&o.unshift("-");var c=o.reduce((function(t,n){if("-"!==n){if(!r.existsSync(n))return e.error("no such file or directory: "+n,{continue:!0}),t;if(e.statFollowLinks(n).isDirectory())return e.error("read failed: "+n+": Is a directory",{continue:!0}),t}var i="-"===n?s:r.readFileSync(n,"utf8");return t.concat(i.trimRight().split("\n"))}),[]),a=c.sort(t.numerical?i:n);return t.reverse&&(a=a.reverse()),a.join("\n")+"\n"}return e.register("sort",o,{canReceivePipe:!0,cmdOptions:{r:"reverse",n:"numerical"}}),ry=o}function Ey(){if(iy)return ny;iy=1;var e=Wv(),r=u.default;function t(t,n){var i=[],o=e.readFromPipe();n||o||e.error("no paths given");var s=1;!0===t.numLines?(s=2,t.numLines=Number(arguments[1])):!1===t.numLines&&(t.numLines=10),t.numLines=-1*Math.abs(t.numLines),n=[].slice.call(arguments,s),o&&n.unshift("-");var c=!1;return n.forEach((function(n){if("-"!==n){if(!r.existsSync(n))return void e.error("no such file or directory: "+n,{continue:!0});if(e.statFollowLinks(n).isDirectory())return void e.error("error reading '"+n+"': Is a directory",{continue:!0})}var s=("-"===n?o:r.readFileSync(n,"utf8")).split("\n");""===s[s.length-1]?(s.pop(),c=!0):c=!1,i=i.concat(s.slice(t.numLines))})),c&&i.push(""),i.join("\n")}return e.register("tail",t,{canReceivePipe:!0,cmdOptions:{n:"numLines"}}),ny=t}function by(){if(sy)return oy;sy=1;var e=Wv(),r=u.default;function t(t,n){n||e.error("no path given");var i=!1;if(Object.keys(t).forEach((function(e){!0===t[e]&&(i=!0)})),i||e.error("could not interpret expression"),t.link)try{return e.statNoFollowLinks(n).isSymbolicLink()}catch(e){return!1}if(!r.existsSync(n))return!1;if(t.exists)return!0;var o=e.statFollowLinks(n);return t.block?o.isBlockDevice():t.character?o.isCharacterDevice():t.directory?o.isDirectory():t.file?o.isFile():t.pipe?o.isFIFO():!!t.socket&&o.isSocket()}return e.register("test",t,{cmdOptions:{b:"block",c:"character",d:"directory",e:"exists",f:"file",L:"link",p:"pipe",S:"socket"},wrapOutput:!1,allowGlobbing:!1}),oy=t}function wy(){if(ay)return cy;ay=1;var e=Wv(),r=u.default,t=l.default;function n(n,i){i||e.error("wrong arguments"),r.existsSync(t.dirname(i))||e.error("no such file or directory: "+t.dirname(i));try{return r.writeFileSync(i,this.stdout||this.toString(),"utf8"),this}catch(r){e.error("could not write to file (code "+r.code+"): "+i,{continue:!0})}}return e.register("to",n,{pipeOnly:!0,wrapOutput:!1}),cy=n}function Oy(){if(ly)return uy;ly=1;var e=Wv(),r=u.default,t=l.default;function n(n,i){i||e.error("wrong arguments"),r.existsSync(t.dirname(i))||e.error("no such file or directory: "+t.dirname(i));try{return r.appendFileSync(i,this.stdout||this.toString(),"utf8"),this}catch(r){e.error("could not append to file (code "+r.code+"): "+i,{continue:!0})}}return e.register("toEnd",n,{pipeOnly:!0,wrapOutput:!1}),uy=n}function Sy(){if(py)return fy;py=1;var e=Wv(),r=u.default;function t(r,t){return t?"string"==typeof t?t=[].slice.call(arguments,1):e.error("file arg should be a string file path or an Array of string file paths"):e.error("no files given"),t.forEach((function(e){n(r,e)})),""}function n(t,n){var o=i(n);if((!o||!o.isDirectory())&&(o||!t.no_create)){r.closeSync(r.openSync(n,"a"));var s=new Date,c=t.date||s,a=t.date||s;if(t.reference){var u=i(t.reference);u||e.error("failed to get attributess of "+t.reference),c=u.mtime,a=u.atime}else t.date&&(c=t.date,a=t.date);t.atime_only&&t.mtime_only||(t.atime_only?c=o.mtime:t.mtime_only&&(a=o.atime)),r.utimesSync(n,a,c)}}function i(r){try{return e.statFollowLinks(r)}catch(e){return null}}return e.register("touch",t,{cmdOptions:{a:"atime_only",c:"no_create",d:"date",m:"mtime_only",r:"reference"}}),fy=t}function Ry(){if(dy)return hy;dy=1;var e=Wv(),r=u.default;function t(t,n,i){var o=e.readFromPipe();o||(n||e.error("no input given"),r.existsSync(n)?e.statFollowLinks(n).isDirectory()&&e.error("error reading '"+n+"'"):e.error(n+": No such file or directory")),i&&r.existsSync(i)&&e.statFollowLinks(i).isDirectory()&&e.error(i+": Is a directory");var s=(n?r.readFileSync(n,"utf8"):o).trimRight().split("\n"),c=s.reduceRight((function(e,r){return 0===e.length?[{count:1,ln:r}]:0===(n=e[0].ln,i=r,t.ignoreCase?n.toLocaleLowerCase().localeCompare(i.toLocaleLowerCase()):n.localeCompare(i))?[{count:e[0].count+1,ln:r}].concat(e.slice(1)):[{count:1,ln:r}].concat(e);var n,i}),[]).filter((function(e){return!t.duplicates||e.count>1})).map((function(e){return(t.count?(r=7,n=e.count,(i=""+n).length<r&&(i=Array(r-i.length+1).join(" ")+i),i+" "):"")+e.ln;var r,n,i})).join("\n")+"\n";return i?(new e.ShellString(c).to(i),""):c}return e.register("uniq",t,{canReceivePipe:!0,cmdOptions:{i:"ignoreCase",c:"count",d:"duplicates"}}),hy=t}function Iy(){if(gy)return vy;gy=1;var e=Wv(),r=u.default,t=l.default;e.register("which",s,{allowGlobbing:!1,cmdOptions:{a:"all"}});function n(){return"win32"===process.platform}function i(e){return e?e.split(t.delimiter):[]}function o(t){return r.existsSync(t)&&!e.statFollowLinks(t).isDirectory()&&(n()||function(e){try{r.accessSync(e,1)}catch(e){return!1}return!0}(t))}function s(r,s){s||e.error("must specify command");var c=n(),a=i(process.env.PATH),u=[];if(-1===s.indexOf("/")){var l=[""];if(c)l=i((process.env.PATHEXT||".com;.exe;.bat;.cmd;.vbs;.vbe;.js;.jse;.wsf;.wsh").toUpperCase());for(var f=0;f<a.length&&(!(u.length>0)||r.all);f++){var p=t.resolve(a[f],s);c&&(p=p.toUpperCase());var h=p.match(/\.[^<>:"/\|?*.]+$/);if(h&&l.indexOf(h[0])>=0){if(o(p)){u.push(p);break}}else for(var d=0;d<l.length;d++){var v=p+l[d];if(o(v)){u.push(v);break}}}}else o(s)&&u.push(t.resolve(s));return u.length>0?r.all?u:u[0]:r.all?[]:null}return vy=s}function Ay(){return yy||(yy={"/node_modules/shelljs/src/cat.js":Vv,"/node_modules/shelljs/src/cd.js":Hv,"/node_modules/shelljs/src/chmod.js":zv,"/node_modules/shelljs/src/common.js":Wv,"/node_modules/shelljs/src/cp.js":Yv,"/node_modules/shelljs/src/dirs.js":tg,"/node_modules/shelljs/src/echo.js":ng,"/node_modules/shelljs/src/error.js":ig,"/node_modules/shelljs/src/exec-child.js":cg,"/node_modules/shelljs/src/exec.js":Tg,"/node_modules/shelljs/src/find.js":Pg,"/node_modules/shelljs/src/grep.js":Fg,"/node_modules/shelljs/src/head.js":Dg,"/node_modules/shelljs/src/ln.js":Cg,"/node_modules/shelljs/src/ls.js":Ng,"/node_modules/shelljs/src/mkdir.js":Mg,"/node_modules/shelljs/src/mv.js":$g,"/node_modules/shelljs/src/popd.js":Xg,"/node_modules/shelljs/src/pushd.js":Kg,"/node_modules/shelljs/src/pwd.js":_g,"/node_modules/shelljs/src/rm.js":Gg,"/node_modules/shelljs/src/sed.js":Jg,"/node_modules/shelljs/src/set.js":Qg,"/node_modules/shelljs/src/sort.js":my,"/node_modules/shelljs/src/tail.js":Ey,"/node_modules/shelljs/src/tempdir.js":Lg,"/node_modules/shelljs/src/test.js":by,"/node_modules/shelljs/src/to.js":wy,"/node_modules/shelljs/src/toEnd.js":Oy,"/node_modules/shelljs/src/touch.js":Sy,"/node_modules/shelljs/src/uniq.js":Ry,"/node_modules/shelljs/src/which.js":Iy})}function jy(e,r){var t,n=function(e){var r=e[0];if("/"===r||"\\"===r)return!1;var t=e[1],n=e[2];return!(!("."!==r||t&&"/"!==t&&"\\"!==t)||!("."!==r||"."!==t||n&&"/"!==n&&"\\"!==n))&&(":"!==t||"/"!==n&&"\\"!==n)}(e);"/"===(e=xy(e))[0]&&(r="");for(var i=Ay(),o=["",".js",".json"];!(t=xy(n?r+"/node_modules/"+e:r+"/"+e)).endsWith("/..");){for(var s=0;s<o.length;s++){var c=t+o[s];if(i[c])return c}if(!n)break;var a=xy(r+"/..");if(a===r)break;r=a}return null}function xy(e){for(var r=(e=e.replace(/\\/g,"/")).split("/"),t=""===r[0],n=1;n<r.length;n++)"."!==r[n]&&""!==r[n]||r.splice(n--,1);for(n=1;n<r.length;n++)".."===r[n]&&n>0&&".."!==r[n-1]&&"."!==r[n-1]&&(r.splice(--n,2),n--);return e=r.join("/"),t&&"/"!==e[0]?e="/"+e:0===e.length&&(e="."),e}fi({target:"String",proto:!0,name:"trimEnd",forced:"".trimEnd!==ey},{trimEnd:ey});var ky,Ly={},_y=["cat","cd","chmod","cp","dirs","echo","exec","find","grep","head","ln","ls","mkdir","mv","pwd","rm","sed","set","sort","tail","tempdir","test","to","toEnd","touch","uniq","which"];function Ty(){if(ky)return Ly;ky=1;var e=Wv();return _y.forEach((function(e){!function(e){function r(r){var t=jy(r,e);if(null!==t)return Ay()[t]();throw new Error('Could not dynamically require "'+r+'". Please configure the dynamicRequireTargets or/and ignoreDynamicRequires option of @rollup/plugin-commonjs appropriately for this require call to work.')}return r.resolve=function(r){var t=jy(r,e);return null!==t?t:require.resolve(r)},r}("/node_modules/shelljs")("./src/"+e)})),Ly.exit=process.exit,Ly.error=ig(),Ly.ShellString=e.ShellString,Ly.env=process.env,Ly.config=e.config,Ly}g(_y);var Ny=Ty(),Py=Wv();Object.keys(Ny).forEach((function(e){v[e]=Ny[e]}));var Fy=wy();String.prototype.to=Py.wrap("to",Fy);var Dy=Oy();String.prototype.toEnd=Py.wrap("toEnd",Dy),v.config.fatal=!0,v.target={};var Cy,My=process.argv.slice(2),Gy=My.indexOf("--");Gy>-1&&(Cy=My.slice(Gy+1,My.length),My=My.slice(0,Gy)),setTimeout((function(){var e;if(1!==My.length||"--help"!==My[0]){for(e in v.target)!function(e,r){v.target[e]=function(){return r.done||(r.done=!0,r.result=r.apply(r,arguments)),r.result}}(e,v.target[e]);My.length>0?My.forEach((function(e){e in v.target?v.target[e](Cy):console.log("no such target: "+e)})):"all"in v.target&&v.target.all(Cy)}else for(e in console.log("Available targets:"),v.target)console.log("  "+e)}),0);var $y={exports:{}};!function(e,r){var t;r=$y.exports=f,t="object"==typeof process&&process.env&&process.env.NODE_DEBUG&&/\bsemver\b/i.test(process.env.NODE_DEBUG)?function(){var e=Array.prototype.slice.call(arguments,0);e.unshift("SEMVER"),console.log.apply(console,e)}:function(){},r.SEMVER_SPEC_VERSION="2.0.0";var n=Number.MAX_SAFE_INTEGER||9007199254740991,i=r.re=[],o=r.src=[],s=r.tokens={},c=0;function a(e){s[e]=c++}a("NUMERICIDENTIFIER"),o[s.NUMERICIDENTIFIER]="0|[1-9]\\d*",a("NUMERICIDENTIFIERLOOSE"),o[s.NUMERICIDENTIFIERLOOSE]="[0-9]+",a("NONNUMERICIDENTIFIER"),o[s.NONNUMERICIDENTIFIER]="\\d*[a-zA-Z-][a-zA-Z0-9-]*",a("MAINVERSION"),o[s.MAINVERSION]="("+o[s.NUMERICIDENTIFIER]+")\\.("+o[s.NUMERICIDENTIFIER]+")\\.("+o[s.NUMERICIDENTIFIER]+")",a("MAINVERSIONLOOSE"),o[s.MAINVERSIONLOOSE]="("+o[s.NUMERICIDENTIFIERLOOSE]+")\\.("+o[s.NUMERICIDENTIFIERLOOSE]+")\\.("+o[s.NUMERICIDENTIFIERLOOSE]+")",a("PRERELEASEIDENTIFIER"),o[s.PRERELEASEIDENTIFIER]="(?:"+o[s.NUMERICIDENTIFIER]+"|"+o[s.NONNUMERICIDENTIFIER]+")",a("PRERELEASEIDENTIFIERLOOSE"),o[s.PRERELEASEIDENTIFIERLOOSE]="(?:"+o[s.NUMERICIDENTIFIERLOOSE]+"|"+o[s.NONNUMERICIDENTIFIER]+")",a("PRERELEASE"),o[s.PRERELEASE]="(?:-("+o[s.PRERELEASEIDENTIFIER]+"(?:\\."+o[s.PRERELEASEIDENTIFIER]+")*))",a("PRERELEASELOOSE"),o[s.PRERELEASELOOSE]="(?:-?("+o[s.PRERELEASEIDENTIFIERLOOSE]+"(?:\\."+o[s.PRERELEASEIDENTIFIERLOOSE]+")*))",a("BUILDIDENTIFIER"),o[s.BUILDIDENTIFIER]="[0-9A-Za-z-]+",a("BUILD"),o[s.BUILD]="(?:\\+("+o[s.BUILDIDENTIFIER]+"(?:\\."+o[s.BUILDIDENTIFIER]+")*))",a("FULL"),a("FULLPLAIN"),o[s.FULLPLAIN]="v?"+o[s.MAINVERSION]+o[s.PRERELEASE]+"?"+o[s.BUILD]+"?",o[s.FULL]="^"+o[s.FULLPLAIN]+"$",a("LOOSEPLAIN"),o[s.LOOSEPLAIN]="[v=\\s]*"+o[s.MAINVERSIONLOOSE]+o[s.PRERELEASELOOSE]+"?"+o[s.BUILD]+"?",a("LOOSE"),o[s.LOOSE]="^"+o[s.LOOSEPLAIN]+"$",a("GTLT"),o[s.GTLT]="((?:<|>)?=?)",a("XRANGEIDENTIFIERLOOSE"),o[s.XRANGEIDENTIFIERLOOSE]=o[s.NUMERICIDENTIFIERLOOSE]+"|x|X|\\*",a("XRANGEIDENTIFIER"),o[s.XRANGEIDENTIFIER]=o[s.NUMERICIDENTIFIER]+"|x|X|\\*",a("XRANGEPLAIN"),o[s.XRANGEPLAIN]="[v=\\s]*("+o[s.XRANGEIDENTIFIER]+")(?:\\.("+o[s.XRANGEIDENTIFIER]+")(?:\\.("+o[s.XRANGEIDENTIFIER]+")(?:"+o[s.PRERELEASE]+")?"+o[s.BUILD]+"?)?)?",a("XRANGEPLAINLOOSE"),o[s.XRANGEPLAINLOOSE]="[v=\\s]*("+o[s.XRANGEIDENTIFIERLOOSE]+")(?:\\.("+o[s.XRANGEIDENTIFIERLOOSE]+")(?:\\.("+o[s.XRANGEIDENTIFIERLOOSE]+")(?:"+o[s.PRERELEASELOOSE]+")?"+o[s.BUILD]+"?)?)?",a("XRANGE"),o[s.XRANGE]="^"+o[s.GTLT]+"\\s*"+o[s.XRANGEPLAIN]+"$",a("XRANGELOOSE"),o[s.XRANGELOOSE]="^"+o[s.GTLT]+"\\s*"+o[s.XRANGEPLAINLOOSE]+"$",a("COERCE"),o[s.COERCE]="(^|[^\\d])(\\d{1,16})(?:\\.(\\d{1,16}))?(?:\\.(\\d{1,16}))?(?:$|[^\\d])",a("COERCERTL"),i[s.COERCERTL]=new RegExp(o[s.COERCE],"g"),a("LONETILDE"),o[s.LONETILDE]="(?:~>?)",a("TILDETRIM"),o[s.TILDETRIM]="(\\s*)"+o[s.LONETILDE]+"\\s+",i[s.TILDETRIM]=new RegExp(o[s.TILDETRIM],"g");a("TILDE"),o[s.TILDE]="^"+o[s.LONETILDE]+o[s.XRANGEPLAIN]+"$",a("TILDELOOSE"),o[s.TILDELOOSE]="^"+o[s.LONETILDE]+o[s.XRANGEPLAINLOOSE]+"$",a("LONECARET"),o[s.LONECARET]="(?:\\^)",a("CARETTRIM"),o[s.CARETTRIM]="(\\s*)"+o[s.LONECARET]+"\\s+",i[s.CARETTRIM]=new RegExp(o[s.CARETTRIM],"g");a("CARET"),o[s.CARET]="^"+o[s.LONECARET]+o[s.XRANGEPLAIN]+"$",a("CARETLOOSE"),o[s.CARETLOOSE]="^"+o[s.LONECARET]+o[s.XRANGEPLAINLOOSE]+"$",a("COMPARATORLOOSE"),o[s.COMPARATORLOOSE]="^"+o[s.GTLT]+"\\s*("+o[s.LOOSEPLAIN]+")$|^$",a("COMPARATOR"),o[s.COMPARATOR]="^"+o[s.GTLT]+"\\s*("+o[s.FULLPLAIN]+")$|^$",a("COMPARATORTRIM"),o[s.COMPARATORTRIM]="(\\s*)"+o[s.GTLT]+"\\s*("+o[s.LOOSEPLAIN]+"|"+o[s.XRANGEPLAIN]+")",i[s.COMPARATORTRIM]=new RegExp(o[s.COMPARATORTRIM],"g");a("HYPHENRANGE"),o[s.HYPHENRANGE]="^\\s*("+o[s.XRANGEPLAIN]+")\\s+-\\s+("+o[s.XRANGEPLAIN]+")\\s*$",a("HYPHENRANGELOOSE"),o[s.HYPHENRANGELOOSE]="^\\s*("+o[s.XRANGEPLAINLOOSE]+")\\s+-\\s+("+o[s.XRANGEPLAINLOOSE]+")\\s*$",a("STAR"),o[s.STAR]="(<|>)?=?\\s*\\*";for(var u=0;u<c;u++)t(u,o[u]),i[u]||(i[u]=new RegExp(o[u]));function l(e,r){if(r&&"object"==typeof r||(r={loose:!!r,includePrerelease:!1}),e instanceof f)return e;if("string"!=typeof e)return null;if(e.length>256)return null;if(!(r.loose?i[s.LOOSE]:i[s.FULL]).test(e))return null;try{return new f(e,r)}catch(e){return null}}function f(e,r){if(r&&"object"==typeof r||(r={loose:!!r,includePrerelease:!1}),e instanceof f){if(e.loose===r.loose)return e;e=e.version}else if("string"!=typeof e)throw new TypeError("Invalid Version: "+e);if(e.length>256)throw new TypeError("version is longer than 256 characters");if(!(this instanceof f))return new f(e,r);t("SemVer",e,r),this.options=r,this.loose=!!r.loose;var o=e.trim().match(r.loose?i[s.LOOSE]:i[s.FULL]);if(!o)throw new TypeError("Invalid Version: "+e);if(this.raw=e,this.major=+o[1],this.minor=+o[2],this.patch=+o[3],this.major>n||this.major<0)throw new TypeError("Invalid major version");if(this.minor>n||this.minor<0)throw new TypeError("Invalid minor version");if(this.patch>n||this.patch<0)throw new TypeError("Invalid patch version");o[4]?this.prerelease=o[4].split(".").map((function(e){if(/^[0-9]+$/.test(e)){var r=+e;if(r>=0&&r<n)return r}return e})):this.prerelease=[],this.build=o[5]?o[5].split("."):[],this.format()}r.parse=l,r.valid=function(e,r){var t=l(e,r);return t?t.version:null},r.clean=function(e,r){var t=l(e.trim().replace(/^[=v]+/,""),r);return t?t.version:null},r.SemVer=f,f.prototype.format=function(){return this.version=this.major+"."+this.minor+"."+this.patch,this.prerelease.length&&(this.version+="-"+this.prerelease.join(".")),this.version},f.prototype.toString=function(){return this.version},f.prototype.compare=function(e){return t("SemVer.compare",this.version,this.options,e),e instanceof f||(e=new f(e,this.options)),this.compareMain(e)||this.comparePre(e)},f.prototype.compareMain=function(e){return e instanceof f||(e=new f(e,this.options)),h(this.major,e.major)||h(this.minor,e.minor)||h(this.patch,e.patch)},f.prototype.comparePre=function(e){if(e instanceof f||(e=new f(e,this.options)),this.prerelease.length&&!e.prerelease.length)return-1;if(!this.prerelease.length&&e.prerelease.length)return 1;if(!this.prerelease.length&&!e.prerelease.length)return 0;var r=0;do{var n=this.prerelease[r],i=e.prerelease[r];if(t("prerelease compare",r,n,i),void 0===n&&void 0===i)return 0;if(void 0===i)return 1;if(void 0===n)return-1;if(n!==i)return h(n,i)}while(++r)},f.prototype.compareBuild=function(e){e instanceof f||(e=new f(e,this.options));var r=0;do{var n=this.build[r],i=e.build[r];if(t("prerelease compare",r,n,i),void 0===n&&void 0===i)return 0;if(void 0===i)return 1;if(void 0===n)return-1;if(n!==i)return h(n,i)}while(++r)},f.prototype.inc=function(e,r){switch(e){case"premajor":this.prerelease.length=0,this.patch=0,this.minor=0,this.major++,this.inc("pre",r);break;case"preminor":this.prerelease.length=0,this.patch=0,this.minor++,this.inc("pre",r);break;case"prepatch":this.prerelease.length=0,this.inc("patch",r),this.inc("pre",r);break;case"prerelease":0===this.prerelease.length&&this.inc("patch",r),this.inc("pre",r);break;case"major":0===this.minor&&0===this.patch&&0!==this.prerelease.length||this.major++,this.minor=0,this.patch=0,this.prerelease=[];break;case"minor":0===this.patch&&0!==this.prerelease.length||this.minor++,this.patch=0,this.prerelease=[];break;case"patch":0===this.prerelease.length&&this.patch++,this.prerelease=[];break;case"pre":if(0===this.prerelease.length)this.prerelease=[0];else{for(var t=this.prerelease.length;--t>=0;)"number"==typeof this.prerelease[t]&&(this.prerelease[t]++,t=-2);-1===t&&this.prerelease.push(0)}r&&(this.prerelease[0]===r?isNaN(this.prerelease[1])&&(this.prerelease=[r,0]):this.prerelease=[r,0]);break;default:throw new Error("invalid increment argument: "+e)}return this.format(),this.raw=this.version,this},r.inc=function(e,r,t,n){"string"==typeof t&&(n=t,t=void 0);try{return new f(e,t).inc(r,n).version}catch(e){return null}},r.diff=function(e,r){if(y(e,r))return null;var t=l(e),n=l(r),i="";if(t.prerelease.length||n.prerelease.length){i="pre";var o="prerelease"}for(var s in t)if(("major"===s||"minor"===s||"patch"===s)&&t[s]!==n[s])return i+s;return o},r.compareIdentifiers=h;var p=/^[0-9]+$/;function h(e,r){var t=p.test(e),n=p.test(r);return t&&n&&(e=+e,r=+r),e===r?0:t&&!n?-1:n&&!t?1:e<r?-1:1}function d(e,r,t){return new f(e,t).compare(new f(r,t))}function v(e,r,t){return d(e,r,t)>0}function g(e,r,t){return d(e,r,t)<0}function y(e,r,t){return 0===d(e,r,t)}function m(e,r,t){return 0!==d(e,r,t)}function E(e,r,t){return d(e,r,t)>=0}function b(e,r,t){return d(e,r,t)<=0}function w(e,r,t,n){switch(r){case"===":return"object"==typeof e&&(e=e.version),"object"==typeof t&&(t=t.version),e===t;case"!==":return"object"==typeof e&&(e=e.version),"object"==typeof t&&(t=t.version),e!==t;case"":case"=":case"==":return y(e,t,n);case"!=":return m(e,t,n);case">":return v(e,t,n);case">=":return E(e,t,n);case"<":return g(e,t,n);case"<=":return b(e,t,n);default:throw new TypeError("Invalid operator: "+r)}}function O(e,r){if(r&&"object"==typeof r||(r={loose:!!r,includePrerelease:!1}),e instanceof O){if(e.loose===!!r.loose)return e;e=e.value}if(!(this instanceof O))return new O(e,r);t("comparator",e,r),this.options=r,this.loose=!!r.loose,this.parse(e),this.semver===S?this.value="":this.value=this.operator+this.semver.version,t("comp",this)}r.rcompareIdentifiers=function(e,r){return h(r,e)},r.major=function(e,r){return new f(e,r).major},r.minor=function(e,r){return new f(e,r).minor},r.patch=function(e,r){return new f(e,r).patch},r.compare=d,r.compareLoose=function(e,r){return d(e,r,!0)},r.compareBuild=function(e,r,t){var n=new f(e,t),i=new f(r,t);return n.compare(i)||n.compareBuild(i)},r.rcompare=function(e,r,t){return d(r,e,t)},r.sort=function(e,t){return e.sort((function(e,n){return r.compareBuild(e,n,t)}))},r.rsort=function(e,t){return e.sort((function(e,n){return r.compareBuild(n,e,t)}))},r.gt=v,r.lt=g,r.eq=y,r.neq=m,r.gte=E,r.lte=b,r.cmp=w,r.Comparator=O;var S={};function R(e,r){if(r&&"object"==typeof r||(r={loose:!!r,includePrerelease:!1}),e instanceof R)return e.loose===!!r.loose&&e.includePrerelease===!!r.includePrerelease?e:new R(e.raw,r);if(e instanceof O)return new R(e.value,r);if(!(this instanceof R))return new R(e,r);if(this.options=r,this.loose=!!r.loose,this.includePrerelease=!!r.includePrerelease,this.raw=e,this.set=e.split(/\s*\|\|\s*/).map((function(e){return this.parseRange(e.trim())}),this).filter((function(e){return e.length})),!this.set.length)throw new TypeError("Invalid SemVer Range: "+e);this.format()}function I(e,r){for(var t=!0,n=e.slice(),i=n.pop();t&&n.length;)t=n.every((function(e){return i.intersects(e,r)})),i=n.pop();return t}function A(e){return!e||"x"===e.toLowerCase()||"*"===e}function j(e,r,t,n,i,o,s,c,a,u,l,f,p){return((r=A(t)?"":A(n)?">="+t+".0.0":A(i)?">="+t+"."+n+".0":">="+r)+" "+(c=A(a)?"":A(u)?"<"+(+a+1)+".0.0":A(l)?"<"+a+"."+(+u+1)+".0":f?"<="+a+"."+u+"."+l+"-"+f:"<="+c)).trim()}function x(e,r,n){for(var i=0;i<e.length;i++)if(!e[i].test(r))return!1;if(r.prerelease.length&&!n.includePrerelease){for(i=0;i<e.length;i++)if(t(e[i].semver),e[i].semver!==S&&e[i].semver.prerelease.length>0){var o=e[i].semver;if(o.major===r.major&&o.minor===r.minor&&o.patch===r.patch)return!0}return!1}return!0}function k(e,r,t){try{r=new R(r,t)}catch(e){return!1}return r.test(e)}function L(e,r,t,n){var i,o,s,c,a;switch(e=new f(e,n),r=new R(r,n),t){case">":i=v,o=b,s=g,c=">",a=">=";break;case"<":i=g,o=E,s=v,c="<",a="<=";break;default:throw new TypeError('Must provide a hilo val of "<" or ">"')}if(k(e,r,n))return!1;for(var u=0;u<r.set.length;++u){var l=r.set[u],p=null,h=null;if(l.forEach((function(e){e.semver===S&&(e=new O(">=0.0.0")),p=p||e,h=h||e,i(e.semver,p.semver,n)?p=e:s(e.semver,h.semver,n)&&(h=e)})),p.operator===c||p.operator===a)return!1;if((!h.operator||h.operator===c)&&o(e,h.semver))return!1;if(h.operator===a&&s(e,h.semver))return!1}return!0}O.prototype.parse=function(e){var r=this.options.loose?i[s.COMPARATORLOOSE]:i[s.COMPARATOR],t=e.match(r);if(!t)throw new TypeError("Invalid comparator: "+e);this.operator=void 0!==t[1]?t[1]:"","="===this.operator&&(this.operator=""),t[2]?this.semver=new f(t[2],this.options.loose):this.semver=S},O.prototype.toString=function(){return this.value},O.prototype.test=function(e){if(t("Comparator.test",e,this.options.loose),this.semver===S||e===S)return!0;if("string"==typeof e)try{e=new f(e,this.options)}catch(e){return!1}return w(e,this.operator,this.semver,this.options)},O.prototype.intersects=function(e,r){if(!(e instanceof O))throw new TypeError("a Comparator is required");var t;if(r&&"object"==typeof r||(r={loose:!!r,includePrerelease:!1}),""===this.operator)return""===this.value||(t=new R(e.value,r),k(this.value,t,r));if(""===e.operator)return""===e.value||(t=new R(this.value,r),k(e.semver,t,r));var n=!(">="!==this.operator&&">"!==this.operator||">="!==e.operator&&">"!==e.operator),i=!("<="!==this.operator&&"<"!==this.operator||"<="!==e.operator&&"<"!==e.operator),o=this.semver.version===e.semver.version,s=!(">="!==this.operator&&"<="!==this.operator||">="!==e.operator&&"<="!==e.operator),c=w(this.semver,"<",e.semver,r)&&(">="===this.operator||">"===this.operator)&&("<="===e.operator||"<"===e.operator),a=w(this.semver,">",e.semver,r)&&("<="===this.operator||"<"===this.operator)&&(">="===e.operator||">"===e.operator);return n||i||o&&s||c||a},r.Range=R,R.prototype.format=function(){return this.range=this.set.map((function(e){return e.join(" ").trim()})).join("||").trim(),this.range},R.prototype.toString=function(){return this.range},R.prototype.parseRange=function(e){var r=this.options.loose;e=e.trim();var n=r?i[s.HYPHENRANGELOOSE]:i[s.HYPHENRANGE];e=e.replace(n,j),t("hyphen replace",e),e=e.replace(i[s.COMPARATORTRIM],"$1$2$3"),t("comparator trim",e,i[s.COMPARATORTRIM]),e=(e=(e=e.replace(i[s.TILDETRIM],"$1~")).replace(i[s.CARETTRIM],"$1^")).split(/\s+/).join(" ");var o=r?i[s.COMPARATORLOOSE]:i[s.COMPARATOR],c=e.split(" ").map((function(e){return function(e,r){return t("comp",e,r),e=function(e,r){return e.trim().split(/\s+/).map((function(e){return function(e,r){t("caret",e,r);var n=r.loose?i[s.CARETLOOSE]:i[s.CARET];return e.replace(n,(function(r,n,i,o,s){var c;return t("caret",e,r,n,i,o,s),A(n)?c="":A(i)?c=">="+n+".0.0 <"+(+n+1)+".0.0":A(o)?c="0"===n?">="+n+"."+i+".0 <"+n+"."+(+i+1)+".0":">="+n+"."+i+".0 <"+(+n+1)+".0.0":s?(t("replaceCaret pr",s),c="0"===n?"0"===i?">="+n+"."+i+"."+o+"-"+s+" <"+n+"."+i+"."+(+o+1):">="+n+"."+i+"."+o+"-"+s+" <"+n+"."+(+i+1)+".0":">="+n+"."+i+"."+o+"-"+s+" <"+(+n+1)+".0.0"):(t("no pr"),c="0"===n?"0"===i?">="+n+"."+i+"."+o+" <"+n+"."+i+"."+(+o+1):">="+n+"."+i+"."+o+" <"+n+"."+(+i+1)+".0":">="+n+"."+i+"."+o+" <"+(+n+1)+".0.0"),t("caret return",c),c}))}(e,r)})).join(" ")}(e,r),t("caret",e),e=function(e,r){return e.trim().split(/\s+/).map((function(e){return function(e,r){var n=r.loose?i[s.TILDELOOSE]:i[s.TILDE];return e.replace(n,(function(r,n,i,o,s){var c;return t("tilde",e,r,n,i,o,s),A(n)?c="":A(i)?c=">="+n+".0.0 <"+(+n+1)+".0.0":A(o)?c=">="+n+"."+i+".0 <"+n+"."+(+i+1)+".0":s?(t("replaceTilde pr",s),c=">="+n+"."+i+"."+o+"-"+s+" <"+n+"."+(+i+1)+".0"):c=">="+n+"."+i+"."+o+" <"+n+"."+(+i+1)+".0",t("tilde return",c),c}))}(e,r)})).join(" ")}(e,r),t("tildes",e),e=function(e,r){return t("replaceXRanges",e,r),e.split(/\s+/).map((function(e){return function(e,r){e=e.trim();var n=r.loose?i[s.XRANGELOOSE]:i[s.XRANGE];return e.replace(n,(function(n,i,o,s,c,a){t("xRange",e,n,i,o,s,c,a);var u=A(o),l=u||A(s),f=l||A(c),p=f;return"="===i&&p&&(i=""),a=r.includePrerelease?"-0":"",u?n=">"===i||"<"===i?"<0.0.0-0":"*":i&&p?(l&&(s=0),c=0,">"===i?(i=">=",l?(o=+o+1,s=0,c=0):(s=+s+1,c=0)):"<="===i&&(i="<",l?o=+o+1:s=+s+1),n=i+o+"."+s+"."+c+a):l?n=">="+o+".0.0"+a+" <"+(+o+1)+".0.0"+a:f&&(n=">="+o+"."+s+".0"+a+" <"+o+"."+(+s+1)+".0"+a),t("xRange return",n),n}))}(e,r)})).join(" ")}(e,r),t("xrange",e),e=function(e,r){return t("replaceStars",e,r),e.trim().replace(i[s.STAR],"")}(e,r),t("stars",e),e}(e,this.options)}),this).join(" ").split(/\s+/);return this.options.loose&&(c=c.filter((function(e){return!!e.match(o)}))),c=c.map((function(e){return new O(e,this.options)}),this)},R.prototype.intersects=function(e,r){if(!(e instanceof R))throw new TypeError("a Range is required");return this.set.some((function(t){return I(t,r)&&e.set.some((function(e){return I(e,r)&&t.every((function(t){return e.every((function(e){return t.intersects(e,r)}))}))}))}))},r.toComparators=function(e,r){return new R(e,r).set.map((function(e){return e.map((function(e){return e.value})).join(" ").trim().split(" ")}))},R.prototype.test=function(e){if(!e)return!1;if("string"==typeof e)try{e=new f(e,this.options)}catch(e){return!1}for(var r=0;r<this.set.length;r++)if(x(this.set[r],e,this.options))return!0;return!1},r.satisfies=k,r.maxSatisfying=function(e,r,t){var n=null,i=null;try{var o=new R(r,t)}catch(e){return null}return e.forEach((function(e){o.test(e)&&(n&&-1!==i.compare(e)||(i=new f(n=e,t)))})),n},r.minSatisfying=function(e,r,t){var n=null,i=null;try{var o=new R(r,t)}catch(e){return null}return e.forEach((function(e){o.test(e)&&(n&&1!==i.compare(e)||(i=new f(n=e,t)))})),n},r.minVersion=function(e,r){e=new R(e,r);var t=new f("0.0.0");if(e.test(t))return t;if(t=new f("0.0.0-0"),e.test(t))return t;t=null;for(var n=0;n<e.set.length;++n){e.set[n].forEach((function(e){var r=new f(e.semver.version);switch(e.operator){case">":0===r.prerelease.length?r.patch++:r.prerelease.push(0),r.raw=r.format();case"":case">=":t&&!v(t,r)||(t=r);break;case"<":case"<=":break;default:throw new Error("Unexpected operation: "+e.operator)}}))}if(t&&e.test(t))return t;return null},r.validRange=function(e,r){try{return new R(e,r).range||"*"}catch(e){return null}},r.ltr=function(e,r,t){return L(e,r,"<",t)},r.gtr=function(e,r,t){return L(e,r,">",t)},r.outside=L,r.prerelease=function(e,r){var t=l(e,r);return t&&t.prerelease.length?t.prerelease:null},r.intersects=function(e,r,t){return e=new R(e,t),r=new R(r,t),e.intersects(r)},r.coerce=function(e,r){if(e instanceof f)return e;"number"==typeof e&&(e=String(e));if("string"!=typeof e)return null;var t=null;if((r=r||{}).rtl){for(var n;(n=i[s.COERCERTL].exec(e))&&(!t||t.index+t[0].length!==e.length);)t&&n.index+n[0].length===t.index+t[0].length||(t=n),i[s.COERCERTL].lastIndex=n.index+n[1].length+n[2].length;i[s.COERCERTL].lastIndex=-1}else t=e.match(i[s.COERCE]);if(null===t)return null;return l(t[2]+"."+(t[3]||"0")+"."+(t[4]||"0"),r)}}(0,$y.exports);var Uy=g($y.exports);const By=global,Xy=new Proxy(global.target,{set:function(e,r,t){return Reflect.set(...arguments)},get:function(e,r,t){return zy(`make ${r}`),Reflect.get(...arguments)}}),qy=["packages","codemods","eslint"],Wy=["eslint","scripts","benchmark",...qy,"*.{js,cjs,mjs,ts}","--format","codeframe","--ext",".js,.cjs,.mjs,.ts"],Vy=By.which("yarn").stdout,Hy=process.execPath;function zy(...e){console.log.apply(console,e)}function Yy(e,r,t,n=!0){zy(`${e.replaceAll(Vy,"yarn").replaceAll(Hy,"node")} ${r.join(" ")}`);try{return s.execFileSync(e,r,{stdio:n?"inherit":void 0,cwd:t&&l.default.resolve(t),env:process.env})}catch(t){throw n&&0!=t.status&&(console.error(new Error(`\ncommand: ${e} ${r.join(" ")}\ncode: ${t.status}`)),process.exit(t.status)),t}}function Ky(e,r,t){return Yy(Vy,e,r,t)}function Jy(e,r,t){return Yy(Hy,e,r,t)}function Qy(e,r){const t=process.env;process.env=Object.assign(Object.assign({},t),r),e(),process.env=t}function Zy(e,t,n){const i="./build/"+e.toLowerCase();By.rm("-rf",i),By.mkdir("-p","build"),Yy("git",["clone","--filter=blob:none","--sparse","--single-branch","--shallow-since='2 years ago'",t,i]),Yy("git",["sparse-checkout","set",...n],i),Yy("git",["checkout","-q",function(e){const t=r.readFileSync("./Makefile","utf8").match(new RegExp(`${e}_COMMIT = (\\w{40})`))[1];if(!t)throw new Error(`Could not find ${e}_COMMIT in Makefile`);return t}(e)],i)}By.config.verbose=!0,Xy["clean-all"]=function(){By.rm("-rf",["node_modules","package-lock.json",".changelog"]),qy.forEach((e=>{By.rm("-rf",`${e}/*/test/tmp`),By.rm("-rf",`${e}/*/package-lock.json`)})),Xy.clean(),Xy["clean-lib"]()},Xy.clean=function(){Xy["test-clean"](),By.rm("-rf",[".npmrc","coverage","packages/*/npm-debug*","node_modules/.cache"])},Xy["test-clean"]=function(){qy.forEach((e=>{By.rm("-rf",`${e}/*/test/tmp`),By.rm("-rf",`${e}/*/test-fixtures.json`)}))},Xy["clean-lib"]=function(){By.rm("-rf",qy.map((e=>`${e}/*/lib`)))},Xy["clean-runtime-helpers"]=function(){By.rm("-rf",["packages/babel-runtime/helpers/**/*.js","packages/babel-runtime-corejs2/helpers/**/*.js","packages/babel-runtime-corejs3/helpers/**/*.js","packages/babel-runtime/helpers/**/*.mjs","packages/babel-runtime-corejs2/helpers/**/*.mjs","packages/babel-runtime-corejs3/helpers/**/*.mjs","packages/babel-runtime-corejs2/core-js"])},Xy["use-cjs"]=function(){Jy(["scripts/set-module-type.js","script"]),Xy.bootstrap()},Xy["use-esm"]=function(){Jy(["scripts/set-module-type.js","module"]),Xy.bootstrap()},Xy["bootstrap-only"]=function(){Xy["clean-all"](),Ky(["install"])},Xy.bootstrap=function(){Xy["bootstrap-only"](),Xy["generate-tsconfig"](),Xy.build()},Xy.build=function(){Xy["build-no-bundle"](),"true"!=process.env.BABEL_COVERAGE&&Xy["build-standalone"]()},Xy["build-standalone"]=function(){Ky(["gulp","build-babel-standalone"])},Xy["build-bundle"]=function(){Xy.clean(),Xy["clean-lib"](),Jy(["scripts/set-module-type.js"]),Ky(["gulp","build"]),Xy["build-flow-typings"](),Xy["build-dist"]()},Xy["build-no-bundle"]=function(){Xy.clean(),Xy["clean-lib"](),Jy(["scripts/set-module-type.js"]),Qy((()=>{Ky(["gulp","build-dev"])}),{BABEL_ENV:"development"}),Xy["build-flow-typings"](),Xy["build-dist"]()},Xy["build-flow-typings"]=function(){r.writeFileSync("packages/babel-types/lib/index.js.flow",Jy(["packages/babel-types/scripts/generators/flow.js"],void 0,!1))},Xy["build-dist"]=function(){Xy["build-plugin-transform-runtime-dist"]()},Xy["build-plugin-transform-runtime-dist"]=function(){Jy(["scripts/build-dist.js"],"packages/babel-plugin-transform-runtime")},Xy.prepublish=function(){process.env.BABEL_8_BREAKING?Jy(["scripts/set-module-type.js","module"]):Jy(["scripts/set-module-type.js","script"]),Xy["bootstrap-only"](),Qy((()=>{Xy["prepublish-build"](),Xy.test()}),{IS_PUBLISH:"true"}),Jy(["scripts/set-module-type.js","clean"])},Xy["prepublish-build"]=function(){Xy["clean-lib"](),Xy["clean-runtime-helpers"](),Qy((()=>{Xy["build-bundle"]()}),{NODE_ENV:"production",BABEL_ENV:"production",STRIP_BABEL_8_FLAG:"true"}),Qy((()=>{Xy["prepublish-build-standalone"](),Xy["clone-license"](),Xy["prepublish-prepare-dts"]()}),{NODE_ENV:"production",STRIP_BABEL_8_FLAG:"true"})},Xy["prepublish-build-standalone"]=function(){Qy((()=>{Xy["build-standalone"]()}),{BABEL_ENV:"production",IS_PUBLISH:"true"})},Xy["prepublish-prepare-dts"]=function(){Xy.tscheck(),Ky(["gulp","bundle-dts"]),Xy["build-typescript-legacy-typings"]()},Xy.tscheck=function(){Xy["generate-tsconfig"](),By.rm("-rf","tsconfig.tsbuildinfo"),By.rm("-rf","dts"),Ky(["tsc","-b","."])},Xy["generate-tsconfig"]=function(){Jy(["scripts/generators/tsconfig.js"]),Jy(["scripts/generators/archived-libs-typings.js"])},Xy["generate-type-helpers"]=function(){Ky(["gulp","generate-type-helpers"])},Xy["build-typescript-legacy-typings"]=function(){r.writeFileSync("packages/babel-types/lib/index-legacy.d.ts",Jy(["packages/babel-types/scripts/generators/typescript-legacy.js"],void 0,!1))},Xy["clone-license"]=function(){Jy(["scripts/clone-license.js"])},Xy.lint=function(){Qy((()=>{Ky(Wy)}),{BABEL_ENV:"test"})},Xy.fix=function(){Xy["fix-json"](),Xy["fix-js"]()},Xy["fix-js"]=function(){Ky([...Wy,"--fix"])},Xy["fix-json"]=function(){Ky(["prettier",`{${qy.join(",")}}/*/test/fixtures/**/options.json`,"--write","--loglevel","warn"])},Xy.watch=function(){Xy["build-no-bundle"](),Qy((()=>{Ky(["gulp","watch"])}),{BABEL_ENV:"development",WATCH_SKIP_BUILD:"true"})},Xy.test=function(){Xy.lint(),Xy["test-only"]()},Xy["test-only"]=function(e=[]){Ky(["jest",...e])},Xy["test-cov"]=function(){Xy.build(),Qy((()=>{Ky(["c8","jest"])}),{BABEL_ENV:"test",BABEL_COVERAGE:"true"})},Xy["bootstrap-test262"]=function(){Zy("TEST262","https://github.com/tc39/test262.git",["test","harness"])},Xy["bootstrap-typescript"]=function(){Zy("TYPESCRIPT","https://github.com/microsoft/TypeScript.git",["tests"])},Xy["bootstrap-flow"]=function(){Zy("FLOW","https://github.com/facebook/flow.git",["src/parser/test/flow"])},Xy["new-version-checklist"]=function(){console.log("\n!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!\n!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!\n!!!!!!                                                   !!!!!!\n!!!!!!              Update the version in                !!!!!!\n!!!!!! packages/babel-plugin-syntax-import-attributes/src/index.ts\n!!!!!!                    and                            !!!!!!\n!!!!!! /Users/nic/Documents/dev/babel/babel/packages/babel-preset-env/src/available-plugins.ts#L159\n!!!!!!                                                   !!!!!!\n!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!\n!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!\n    ".trim()),process.exit(1)},Xy["new-version"]=function(){Xy["new-version-checklist"](),Yy("git",["pull","--rebase"]),Ky(["release-tool","version","-f","@babel/standalone"])},Xy["new-babel-8-version-prepare"]=function(){Yy("git",["pull","--rebase"]);const e=JSON.parse(r.readFileSync("./package.json","utf8")),t=Uy.inc(e.version_babel8,"prerelease");return e.version_babel8=t,r.writeFileSync("./package.json",JSON.stringify(e,null,2)+"\n"),Yy("git",["add","./package.json"]),Yy("git",["commit","-m","Bump Babel 8 version to "+t]),t},Xy["new-babel-8-version"]=function(){const e=Xy["new-babel-8-version-prepare"]();Yy("git",["checkout","-b","release/v"+e]),qy.forEach((t=>{r.readdirSync(t).forEach((n=>{const i=`${t}/${n}/package.json`;try{const t=JSON.parse(r.readFileSync(i,"utf8"));t.peerDependencies["@babel/core"]=`^${e}`,r.writeFileSync(i,JSON.stringify(t,null,2)+"\n")}catch(e){}}))})),Ky(["release-tool","version",e,"--all"]),console.log(`Run \`git push upstream main release/v${e} --follow-tags\` to push the changes on GitHub and release them.`)};
