+++ packages/babel-node/package.json
@@ -25,7 +25,7 @@
   "dependencies": {
     "@babel/register": "workspace:^",
     "commander": "^4.0.1",
-    "core-js": "^3.26.0",
+    "core-js": "^3.30.2",
     "node-environment-flags": "^1.0.5",
     "regenerator-runtime": "^0.13.11",
     "v8flags": "^3.1.1"
