+++ packages/babel-plugin-proposal-decorators/package.json
@@ -35,7 +35,7 @@
     "@babel/traverse": "workspace:^",
     "@types/charcodes": "^0.2.0",
     "array.prototype.concat": "^1.0.2",
-    "babel-plugin-polyfill-es-shims": "^0.7.1",
+    "babel-plugin-polyfill-es-shims": "^0.9.0",
     "charcodes": "^0.2.0",
     "object.getownpropertydescriptors": "^2.1.1"
   },
