+++ packages/babel-plugin-proposal-duplicate-named-capturing-groups-regex/package.json
@@ -30,7 +30,7 @@
   "devDependencies": {
     "@babel/core": "workspace:^",
     "@babel/helper-plugin-test-runner": "workspace:^",
-    "core-js": "^3.26.0"
+    "core-js": "^3.30.2"
   },
   "engines": {
     "node": ">=6.9.0"
