+++ packages/babel-plugin-transform-runtime/package.json
@@ -22,9 +22,9 @@
   "dependencies": {
     "@babel/helper-module-imports": "workspace:^",
     "@babel/helper-plugin-utils": "workspace:^",
-    "babel-plugin-polyfill-corejs2": "^0.3.3",
-    "babel-plugin-polyfill-corejs3": "^0.6.0",
-    "babel-plugin-polyfill-regenerator": "^0.4.1",
+    "babel-plugin-polyfill-corejs2": "^0.4.1",
+    "babel-plugin-polyfill-corejs3": "^0.8.1",
+    "babel-plugin-polyfill-regenerator": "^0.5.0",
     "semver": "condition:BABEL_8_BREAKING ? ^7.3.4 : ^6.3.0"
   },
   "peerDependencies": {
