+++ packages/babel-plugin-transform-runtime/test/fixtures/regression/with-env-useBuiltIns-babel-7/input.js
@@ -1,0 +1,1 @@
+class A {}
