+++ packages/babel-plugin-transform-runtime/test/fixtures/regression/with-env-useBuiltIns-babel-7/output.js
@@ -1,0 +1,7 @@
+var _createClass = require("@babel/runtime/helpers/createClass");
+var _classCallCheck = require("@babel/runtime/helpers/classCallCheck");
+var A = /*#__PURE__*/_createClass(function A() {
+  "use strict";
+
+  _classCallCheck(this, A);
+});
