+++ packages/babel-plugin-transform-runtime/test/fixtures/regression/with-env-useBuiltIns/input.js
@@ -1,0 +1,1 @@
+class A {}
