+++ packages/babel-preset-env/package.json
@@ -88,10 +88,10 @@
     "@babel/plugin-transform-unicode-regex": "workspace:^",
     "@babel/preset-modules": "^0.1.5",
     "@babel/types": "workspace:^",
-    "babel-plugin-polyfill-corejs2": "^0.3.3",
-    "babel-plugin-polyfill-corejs3": "^0.6.0",
-    "babel-plugin-polyfill-regenerator": "^0.4.1",
-    "core-js-compat": "^3.25.1",
+    "babel-plugin-polyfill-corejs2": "^0.4.1",
+    "babel-plugin-polyfill-corejs3": "^0.8.1",
+    "babel-plugin-polyfill-regenerator": "^0.5.0",
+    "core-js-compat": "^3.30.2",
     "semver": "condition:BABEL_8_BREAKING ? ^7.3.4 : ^6.3.0"
   },
   "peerDependencies": {
