+++ packages/babel-runtime-corejs3/package.json
@@ -13,7 +13,7 @@
   },
   "author": "The Babel Team (https://babel.dev/team)",
   "dependencies": {
-    "core-js-pure": "^3.25.1",
+    "core-js-pure": "^3.30.2",
     "regenerator-runtime": "^0.13.11"
   },
   "exports": {
